NewscGPS_SME

NewscGPS generates a scGPS object in the SingleCellExperiment class for use with the scGPS package.This object contains an expression matrix, associated metadata (cells, genes, clusters). The data are expected to be normalised counts.

NewscGPS_SME(ExpressionMatrix = NULL, GeneMetadata = NULL,
  CellMetadata = NULL)

Arguments

ExpressionMatrix
An expression dataset in matrix format. Rows should represent a transcript and its normalised counts, while columns should represent individual cells.

GeneMetadata
A data frame or vector containing gene identifiers used in the expression matrix. The first column should hold the cell identifiers you are using in the expression matrix. Other columns contain information about the genes, such as their corresponding ENSEMBL transcript identifiers.

CellMetadata
A data frame containing cell identifiers (usually barcodes) and clustering information (the first column of the data frame contains clustering information). The column containing clustering information needs to be named as 'Cluster'. If clustering information is not available, users can run CORE function and add the information to the scGPS before running scGPS prediction

Value

This function generates an scGPS object belonging to the SingleCellExperiment.

See also

Examples

day2 <- sample1
t <-NewscGPS_SME(ExpressionMatrix = day2$dat2_counts, GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
colData(t); show(t); colnames(t)

#> DataFrame with 590 rows and 1 column
#>                            X
#>                    <integer>
#> 2_CCCATGTGTCTCAT-1         2
#> 1_TACGGCCTAGAACA-1         1
#> 2_ACGATCGAGTACGT-1         3
#> 2_TACGGCCTAAAGTG-1         2
#> 2_AACTCTTGGGTATC-1         2
#> ...                      ...
#> 1_TTCAGACTTCGTGA-1         1
#> 2_CGCAAATGCGTTGA-1         1
#> 1_CAAGTCGATTCTAC-1         1
#> 2_AAGATTACGAGCAG-1         1
#> 2_TATCTTCTCGCCTT-1         1

#> class: SummarizedExperiment 
#> dim: 16990 590 
#> metadata(0):
#> assays(1): ''
#> rownames(16990): FO538757.1_ENSG00000279457 AP006222.2_ENSG00000228463
#>   ... AC004556.1_ENSG00000276345 AC240274.1_ENSG00000271254
#> rowData names(2): GeneSymbol EnsemblID
#> colnames(590): 2_CCCATGTGTCTCAT-1 1_TACGGCCTAGAACA-1 ...
#>   2_AAGATTACGAGCAG-1 2_TATCTTCTCGCCTT-1
#> colData names(1): X

#>   [1] "2_CCCATGTGTCTCAT-1" "1_TACGGCCTAGAACA-1" "2_ACGATCGAGTACGT-1"
#>   [4] "2_TACGGCCTAAAGTG-1" "2_AACTCTTGGGTATC-1" "2_GAAAGATGTTTGGG-1"
#>   [7] "2_TCATCATGAGTCTG-1" "2_GACACTGAGGAGGT-1" "2_GCGATATGTACTGG-1"
#>  [10] "1_GATATTGAGCCCTT-1" "2_GGACGCACTGACTG-1" "2_TAGGCATGTCTATC-1"
#>  [13] "2_GCAGATACTCCAGA-1" "1_TAAGATACATCTTC-1" "1_TACTTTCTACCATG-1"
#>  [16] "1_CAAGAAGATTTCTG-1" "2_ATCGAGTGTGCACA-1" "2_CATACTACGGTATC-1"
#>  [19] "2_CCTCATCTAAAACG-1" "2_GGACGCTGACACAC-1" "1_ATTACCACAAGGTA-1"
#>  [22] "2_ATTTCCGAACCTCC-1" "1_CCTCTACTCTACCC-1" "2_AGAAGATGCTCCCA-1"
#>  [25] "2_CTATAGCTCTCAGA-1" "2_AAGTTCCTGTAAGA-1" "2_GGCACGTGCGTACA-1"
#>  [28] "2_GCGAGAGACTAGCA-1" "2_TTCACAACTTCCAT-1" "2_ATCCTAACGCTGAT-1"
#>  [31] "1_ACGGAACTGCTACA-1" "2_TGCAATCTCTGCTC-1" "1_CGACTCACTGACAC-1"
#>  [34] "2_CCGAAAACGACACT-1" "2_CAGACCCTCTTCGC-1" "2_CCCAGACTCCAGTA-1"
#>  [37] "2_GCGACTCTGCATCA-1" "2_GAACACACGAAACA-1" "2_GTCCCATGTCCTGC-1"
#>  [40] "2_AACTCACTGGAGCA-1" "2_ATCTTGACGGTTAC-1" "2_AAAGACGAGTTAGC-1"
#>  [43] "2_CACATGGATTTGTC-1" "1_CCGACTACTCCTTA-1" "2_TTACACACCCCGTT-1"
#>  [46] "2_CCATCCGAAAGGCG-1" "2_TTCTACGATGAAGA-1" "2_GCATGATGAGTGCT-1"
#>  [49] "2_CGCTACACGGTCAT-1" "2_TTCAACACAGCAAA-1" "2_GTAGACTGCTTGAG-1"
#>  [52] "2_TTCATCGAGGGTGA-1" "1_CGAAGACTACTGTG-1" "2_AGAGCTACGGACGA-1"
#>  [55] "2_GCTATACTGTTGTG-1" "2_AGAGTGCTTGCTAG-1" "1_AGAATACTCTGATG-1"
#>  [58] "2_CTACGCACGGAAAT-1" "2_GATATCCTTCGCTC-1" "2_ATTAGATGCCACAA-1"
#>  [61] "1_CGAGTATGGAGACG-1" "1_GCGGACTGCTTGGA-1" "2_ACTCCCGACCACAA-1"
#>  [64] "2_GAGCAGGAAGAATG-1" "2_AGGGAGTGGTAAGA-1" "2_TAATGAACAAGGGC-1"
#>  [67] "2_GCACGGACCTCATT-1" "2_TCCGGACTAGAATG-1" "1_AAAGACGACACTGA-1"
#>  [70] "2_AGAAAGTGCATTTC-1" "1_CACATACTCGGTAT-1" "1_AACGGTTGCGCAAT-1"
#>  [73] "2_ATCCATACTGTGCA-1" "2_ATGAGAGATATGCG-1" "2_GTGTCAGATATGCG-1"
#>  [76] "2_AAGCCTGAGCTATG-1" "1_AACTCTTGTGACAC-1" "1_AACCCAGAGCATAC-1"
#>  [79] "2_GTTGACGATCTAGG-1" "1_TAGATTGAGTATCG-1" "1_CCCAGTTGCGTTGA-1"
#>  [82] "2_GTTAGGTGCCAAGT-1" "2_TTCCTAGACGAATC-1" "2_AAGCACTGGGTGAG-1"
#>  [85] "2_AATCCTTGGCGTAT-1" "2_TACCGCTGTCAGAC-1" "2_TGGAACTGGAGCTT-1"
#>  [88] "1_GAACCAACCTGAGT-1" "2_CGACTCTGCTGTCC-1" "1_CATCATACAGGTCT-1"
#>  [91] "2_ACCCTCGACCTATT-1" "2_TGGATTCTGTAAGA-1" "2_ACATGGTGCAGAGG-1"
#>  [94] "1_TCGATTTGTTGGTG-1" "2_CCCGGAGACAAGCT-1" "2_CTTATCGACCTATT-1"
#>  [97] "1_GTTCAGGACTTACT-1" "2_CGACCTACCCAATG-1" "1_CAAGGACTCTAGCA-1"
#> [100] "2_CGACCGGAGTACAC-1" "2_GGGCACACGCGTTA-1" "1_GTCCAAGATTTCGT-1"
#> [103] "2_CAAGCTGAATCGGT-1" "2_GGAACACTCTACTT-1" "1_CCAGTGCTACCTTT-1"
#> [106] "2_TCATTCGAGTCACA-1" "2_GATTCGGACGAATC-1" "1_ACGGAACTCTTCTA-1"
#> [109] "2_ACCCGTACGGTTCA-1" "2_CATAAAACGTTCAG-1" "2_GCCTACACTTTCAC-1"
#> [112] "2_TGGTAGTGGAGATA-1" "1_CTTACTGAGTCTAG-1" "1_AAGATTACCAGTTG-1"
#> [115] "2_ATATGAACGTATCG-1" "1_GCACAATGATCGGT-1" "2_GGGAACGATTCCGC-1"
#> [118] "1_GCGAGAGAATTCCT-1" "1_TACTACTGAGGTTC-1" "2_ATTGATGACCACAA-1"
#> [121] "1_CCAGTCTGCTCGAA-1" "2_TTTCTACTACCAAC-1" "2_AGAAACGAGAAACA-1"
#> [124] "2_ATTTCTCTTCTACT-1" "2_GTGTGATGCATGAC-1" "2_ACTGCCACTTCTTG-1"
#> [127] "2_TAGAGCACGCATCA-1" "2_TATCAGCTTTGTCT-1" "1_ATGTCGGAGTTAGC-1"
#> [130] "2_ACAATTGATGCTCC-1" "1_AGTAATACGTACCA-1" "1_ACTGAGACTTCTTG-1"
#> [133] "2_AAATCTGATACAGC-1" "2_GTTGAGTGTTAGGC-1" "2_GCACAAACAAGCCT-1"
#> [136] "2_TACGAGTGTGAAGA-1" "2_CAGGAACTACCCTC-1" "1_CTCGAGCTGACAAA-1"
#> [139] "2_TGCCCAACCTGAGT-1" "2_AGTACGTGAAGGGC-1" "2_CACTATACCTGACA-1"
#> [142] "2_ATCGCCACATTGGC-1" "2_TTCAGACTAAAGCA-1" "2_CACTGAGAACGGGA-1"
#> [145] "2_ACTACTACGATACC-1" "2_TGGAGGGACTACTT-1" "2_TGCAGATGATCGTG-1"
#> [148] "2_TTAGACCTTGGAAA-1" "1_AGTAAGGACCACCT-1" "1_GATTCGGAGGAAAT-1"
#> [151] "1_AGGTACACGTTTGG-1" "2_TAAGTAACGCGAGA-1" "2_TTCAAAGATCGACA-1"
#> [154] "2_AAGCACTGTCGTGA-1" "2_CCAAGAACACACTG-1" "2_CGTGCACTAGAGGC-1"
#> [157] "2_TATACGCTTCAGGT-1" "2_ATGCGATGGGTTTG-1" "2_GTTAGGTGAACGGG-1"
#> [160] "2_CGGTCACTACTCAG-1" "2_CAGCTCACTTCATC-1" "2_GTATCACTCATTGG-1"
#> [163] "2_TTGCTAACCTAGAC-1" "1_TGCTGAGATGCTCC-1" "1_GTTAGGTGGTTGAC-1"
#> [166] "2_TACTCAACGTAAAG-1" "2_GATCCCTGCTATTC-1" "2_ACAACCGACGGAGA-1"
#> [169] "2_TCAGACGAGTTTCT-1" "2_CGACGTCTCGTCTC-1" "2_GACCAAACTGCTTT-1"
#> [172] "2_CTGAAGACACAGCT-1" "2_TGGACCCTCCAATG-1" "1_TACTTTCTGAGAGC-1"
#> [175] "1_GTCGCACTCGAGTT-1" "2_CCCAACTGGGATTC-1" "2_AGGTGGGATTGGCA-1"
#> [178] "2_GACTACGAGAACTC-1" "2_CTAGTTACGTATGC-1" "1_GGCCGAACTTCTCA-1"
#> [181] "2_TAACGTCTGAGAGC-1" "2_TAGTCGGATTGAGC-1" "1_CACTATACTCGCCT-1"
#> [184] "1_GACCATGACCGCTT-1" "2_TGACTTACGAACTC-1" "1_ATCTACTGCTGTTT-1"
#> [187] "1_GATCGATGTACGCA-1" "2_TTCAACACATGTGC-1" "2_CATTGACTACGGAG-1"
#> [190] "2_TTACGACTCGAACT-1" "2_AATCGGTGGGATTC-1" "2_TTGACACTCTAGCA-1"
#> [193] "2_TGACCAGATTCTAC-1" "2_TAACCGGACTTCCG-1" "2_TCGCACACTGACTG-1"
#> [196] "1_GTTAACCTTGAGGG-1" "2_TGATTAGAGGTATC-1" "1_CAGCCTACATGGTC-1"
#> [199] "2_ATTGGTCTCTCCCA-1" "2_TACTAAGACACTTT-1" "2_TAAACAACCGCAAT-1"
#> [202] "2_TTTAGCTGTCTGGA-1" "1_TGTAGGTGACACAC-1" "2_TTCAGTACACCCTC-1"
#> [205] "2_CGACAAACAGCAAA-1" "2_GAACACACAGCCTA-1" "2_CGCTACTGCGCTAA-1"
#> [208] "2_CAAATTGAACTGGT-1" "2_AAGGCTACGCTAAC-1" "2_AGCAAAGACTTGAG-1"
#> [211] "2_AACCTACTCACTCC-1" "2_GAGTACACCATGGT-1" "2_CGGACTCTTTTGGG-1"
#> [214] "2_CTATACTGTGGAAA-1" "2_ACCACGCTAACAGA-1" "2_ATTCCATGAGGTTC-1"
#> [217] "2_CGGACTCTTCGTAG-1" "2_CGACCTTGACCAAC-1" "2_TTGCTATGCCTCGT-1"
#> [220] "2_GTATCTACATGGTC-1" "1_TCCCTACTCATTGG-1" "2_ACGTAGACAGCAAA-1"
#> [223] "2_GATTCTTGAGGGTG-1" "2_AAGAGATGTAACGC-1" "2_GCTCCATGTCCTCG-1"
#> [226] "1_GAGCTCCTGTAAGA-1" "2_AGAGCGGAAGCTAC-1" "2_CTACGGCTGCTTCC-1"
#> [229] "1_AACGCCCTGCTATG-1" "2_CTGACAGAACCTAG-1" "2_ATAAACACTGGATC-1"
#> [232] "2_ATACCTACTTTGCT-1" "2_GTCACAGATTTGTC-1" "1_AAGCACTGTGCAAC-1"
#> [235] "2_GAGGTGGAGAGAGC-1" "2_GGGCACACTGTGCA-1" "1_AGCATCGAATTCCT-1"
#> [238] "2_GGTAAAGAGTAGGG-1" "2_ATCTTTCTAGGTCT-1" "2_TCAGCAGATCTCGC-1"
#> [241] "1_CCTAAGGAAGGTCT-1" "2_TGATATGATAGCGT-1" "2_AGAGGTCTAGAGAT-1"
#> [244] "1_TTATGCACTAGAAG-1" "2_TACGCAGAGTATCG-1" "2_CCTCTACTACAGTC-1"
#> [247] "2_TACTAAGACACCAA-1" "2_TCAATCACCTGACA-1" "2_AATGAGGATTGGTG-1"
#> [250] "2_ACGAGGGAACCGAT-1" "2_ACTGTGGACATCAG-1" "2_CGACCACTCAGATC-1"
#> [253] "2_ACCTCCGAATTCGG-1" "1_AAGGTCTGTCACGA-1" "1_ATTGCACTCCAGTA-1"
#> [256] "2_CATAACCTTCAGTG-1" "2_GTTGAGTGTGAGAA-1" "2_CCAAGAACGCTAAC-1"
#> [259] "2_AGAGAAACTCAGGT-1" "2_AAGAAGACAAACGA-1" "1_CCAGGTCTGTATGC-1"
#> [262] "1_AGACACACTTTGTC-1" "2_AGTTAAACAACCTG-1" "2_TTCATGACTGCTAG-1"
#> [265] "2_GATCTTTGTCCGTC-1" "2_TAATGAACCCATGA-1" "1_TATCCAACGGTCAT-1"
#> [268] "2_GGAGAGACAGTAGA-1" "2_GCAACCCTCTATTC-1" "2_TCACGAGATGGTGT-1"
#> [271] "2_AGGTTGTGATCGAC-1" "2_AGGCAACTGCGTAT-1" "2_CAATTCACGGTATC-1"
#> [274] "2_AAATCTGACTATGG-1" "2_ACGATTCTTCCAGA-1" "2_CAAATATGAACCGT-1"
#> [277] "2_GAACAGCTTTTGGG-1" "2_AACGCATGGACAGG-1" "2_GCTAGAACGGTCAT-1"
#> [280] "1_GGAATCTGCACAAC-1" "2_AGCCTCTGGTCTTT-1" "1_TAGAGCACGTCAAC-1"
#> [283] "2_ACAGACACGACAGG-1" "2_GGCGCATGCTGATG-1" "2_CAGCTCACACCCTC-1"
#> [286] "2_ACACGAACAAGAAC-1" "2_AGCTCGCTGGCGAA-1" "2_CGGCACGAGGTGGA-1"
#> [289] "2_GCAGATACAACAGA-1" "2_CAATCGGAATGTGC-1" "2_GCCAACCTTGTGCA-1"
#> [292] "2_GAAGTCACCTTAGG-1" "2_CCCGGAGATGATGC-1" "1_TTTCACGAAAAGTG-1"
#> [295] "1_GTGAACACCGTAGT-1" "1_GGAGGCCTTACTGG-1" "2_CTCGCATGCTCGAA-1"
#> [298] "2_AGGAACCTTTCGGA-1" "2_CTCGACACCTCTTA-1" "2_ATAATCGACTTCTA-1"
#> [301] "1_TAGAGCACGGACTT-1" "1_TAGCTACTTGGTTG-1" "2_GGCGGACTACTGTG-1"
#> [304] "1_ACGCACCTGCTTCC-1" "2_TAAGTAACGCCTTC-1" "1_GTCCACTGACTCAG-1"
#> [307] "2_TCACTATGTTCCGC-1" "2_AAAGGCCTTTCTTG-1" "2_TACGCCACGGGATG-1"
#> [310] "2_TAATGCCTCCTTGC-1" "2_GAGTGTTGCATTGG-1" "2_CATCAACTCTAGAC-1"
#> [313] "2_TACTCTGAGTAGCT-1" "2_AAGGTGCTTTGCGA-1" "2_TTCTCAGAGGTCAT-1"
#> [316] "1_AACGTCGAGAGAGC-1" "2_GGTAGTACACCTGA-1" "1_AGGGTTTGGCCAAT-1"
#> [319] "2_ACCTCCGATTGCTT-1" "2_ATTCGACTCGACTA-1" "2_CATAGTCTTGCAGT-1"
#> [322] "1_TAACCGGACATGAC-1" "2_GTGATTCTACTTTC-1" "2_ATATACGAAAGTGA-1"
#> [325] "1_CCAAGTGATTGCGA-1" "2_TTTGCATGGCCTTC-1" "2_GATATATGAGTCAC-1"
#> [328] "1_ACGCACCTAGGAGC-1" "2_AATACTGAACTGTG-1" "2_AGTTTCACAAACAG-1"
#> [331] "1_GCACACCTCATTCT-1" "2_GAGTCTGATTCAGG-1" "2_CCGCTATGCTGTGA-1"
#> [334] "2_AGCAAGCTGTTGCA-1" "2_TACTGGGATAGAGA-1" "1_GGCCAGACTAGAAG-1"
#> [337] "2_TAAAGTTGGAACCT-1" "2_AAGTTCCTTGAGGG-1" "2_AACTCTTGTTTCAC-1"
#> [340] "2_ATCAAATGACCTCC-1" "2_GGAGCCACTTCGTT-1" "2_ACCCAAGAACGTTG-1"
#> [343] "2_GTAACGTGAGAACA-1" "2_AGGGACGATTCTGT-1" "2_GAGCAGGACCACAA-1"
#> [346] "2_ACAGGTACTTCGGA-1" "2_ACTGTTACGGTTCA-1" "1_CTCCACGACGCTAA-1"
#> [349] "2_CTAACGGAGCAGAG-1" "2_CAATAATGTCGCAA-1" "2_AGAAGATGGGTTAC-1"
#> [352] "1_CCAGATGATGCATG-1" "2_GGGAAGACAGGTTC-1" "2_TCGCCATGTCGTTT-1"
#> [355] "1_CAGACAACCTCCCA-1" "2_TCTCAAACTGCCCT-1" "1_CATACTTGACCCTC-1"
#> [358] "2_GGAAGGTGACTGTG-1" "1_ATGCACGAAGAGAT-1" "2_AGACTGACTGTGAC-1"
#> [361] "2_AGTTTAGAGTCGTA-1" "2_CGAGAACTCTGACA-1" "2_CCGGTACTGGTTCA-1"
#> [364] "2_GCTCGACTCCTCCA-1" "1_TCGGTAGAGTTCAG-1" "1_CACTCCGACTCTTA-1"
#> [367] "2_CCATCCGATCTAGG-1" "2_GAAGCGGAAGCGTT-1" "2_ATCTCAACGTACCA-1"
#> [370] "2_ACGACAACCCATGA-1" "2_GAAGTCTGTGCACA-1" "2_GCCACGGACCTATT-1"
#> [373] "2_CCATTAACAAAACG-1" "2_AAGTTATGATGTGC-1" "2_CCACGGGATCCTCG-1"
#> [376] "1_AGCATCGATGGTCA-1" "2_GCCTGACTACGGGA-1" "2_CTTTACGAATGGTC-1"
#> [379] "2_AGAACGCTAGTTCG-1" "2_TAGAATACTGTCAG-1" "2_TGGACCCTCACCAA-1"
#> [382] "2_ACCATTACAGAGTA-1" "1_CTCCATCTCTACGA-1" "2_AACCTTACGCTAAC-1"
#> [385] "1_GAGGTGGATGCTAG-1" "2_AGAGATGATCAGGT-1" "1_TGCCGACTGGTAAA-1"
#> [388] "2_CAGAAGCTCCTTTA-1" "1_ATTGCTACGTCTTT-1" "1_AACTCACTTGGGAG-1"
#> [391] "2_GCACCACTCTCTAT-1" "2_TTGCATTGTCCTTA-1" "2_GAAAGATGAGAGGC-1"
#> [394] "2_ACACAGACTCAGAC-1" "2_AGTTCTACATCGAC-1" "2_ATGAAACTCTCATT-1"
#> [397] "1_GCAACTGAGCTGAT-1" "2_ACCTCGTGTCGTGA-1" "2_GTGGAGGAAGGTTC-1"
#> [400] "1_GCCCAGGAAGCGTT-1" "1_ATCATGCTAGCAAA-1" "2_ACTTGTACTACTTC-1"
#> [403] "2_TCACCTCTTTGTGG-1" "2_ATATAGTGCAACCA-1" "2_TTAGAATGGTAGGG-1"
#> [406] "2_GGAGAGACATTGGC-1" "2_TTGCTAACCGTACA-1" "2_CACATACTGGTCTA-1"
#> [409] "2_AATACTGATCCTGC-1" "2_TAACTCACGTCATG-1" "2_AGTCTTACTCCTAT-1"
#> [412] "2_CTGATGGAGTACGT-1" "2_TTGGTACTCATTCT-1" "2_CAATGGACTCTGGA-1"
#> [415] "2_TCAGTACTATTCTC-1" "2_GCCACGGACTCCAC-1" "2_CCACTTCTTCTAGG-1"
#> [418] "1_TAGTTAGACCAGTA-1" "1_TAGAGAGACCTGAA-1" "2_CTCGACACTATCTC-1"
#> [421] "2_CTTACAACTTCTGT-1" "2_GGAATCTGGTTGCA-1" "1_CATAAATGAGTCAC-1"
#> [424] "2_ACTTGTACAAGCAA-1" "2_AGCTGCCTTTAGGC-1" "1_GATCCGCTCTGTAG-1"
#> [427] "2_ATTAAGACATTCCT-1" "1_AGCCGGTGGATACC-1" "1_GAAGTGCTAGGGTG-1"
#> [430] "2_CAGACTGACGACAT-1" "1_CAGACAACGGAGGT-1" "2_AGTAGGCTCATGAC-1"
#> [433] "1_ACTCAGGATCGATG-1" "1_CACTGCTGTCACGA-1" "2_CACCCATGCTTACT-1"
#> [436] "2_TTCGATTGTCGCCT-1" "2_CAAACTCTCCCAAA-1" "2_TTACCATGCCCTTG-1"
#> [439] "2_TCCGAAGAGTAGCT-1" "1_GGCGCATGATGCTG-1" "2_GCACGTCTTGTCAG-1"
#> [442] "1_ACTGTGGATCTCCG-1" "2_TAGACGTGCGTAAC-1" "1_CGGACCGAACGACT-1"
#> [445] "2_GTACTACTAGGTCT-1" "2_CATTTGTGGAAACA-1" "1_TAATGATGAGAGGC-1"
#> [448] "1_GTCAACGAACGGGA-1" "2_CAATCGGAAGAATG-1" "2_CATGTTTGGAGGAC-1"
#> [451] "2_GCCTACACGTCGAT-1" "2_CTTACTGAGGTTTG-1" "2_AGTACTCTTAGTCG-1"
#> [454] "2_GGGTTAACCTGCTC-1" "1_GGTGGAGATGACCA-1" "1_CCTGAGCTCTGCAA-1"
#> [457] "2_CGCTCATGTGCCTC-1" "2_CGAAGTACTCTGGA-1" "2_GATAATACAGATCC-1"
#> [460] "2_GATATTGAGCCTTC-1" "1_ACTCTCCTTGCATG-1" "1_CGAAGTACAGCACT-1"
#> [463] "2_TTATGCACGGAGGT-1" "2_TCAGAGACAGTGTC-1" "2_AGCATCGAGTTGAC-1"
#> [466] "2_CGTACAGACATTGG-1" "2_TCCCGATGAGTAGA-1" "2_CAGCCTACTCAGTG-1"
#> [469] "2_TAATGTGATTCCGC-1" "2_AATTGTGAACACTG-1" "2_AAGCGACTCAGTCA-1"
#> [472] "2_TAAGATTGTGATGC-1" "2_GACGTATGTGAACC-1" "1_GCTCAGCTTCTGGA-1"
#> [475] "2_CGCAAATGGGTCTA-1" "2_ATATGAACTTCTTG-1" "1_AAGCCTGAGAGGCA-1"
#> [478] "2_ACGCAATGTTTGTC-1" "2_GGACGCTGTCGACA-1" "1_TTATGGCTGGGAGT-1"
#> [481] "2_GTACCCTGATCGGT-1" "1_GATCTTACTTCTGT-1" "2_CCCTACGATAGCGT-1"
#> [484] "2_TTCATGTGTAAGGA-1" "2_GCAAGACTCCTTGC-1" "2_AGCGGGCTCTGTTT-1"
#> [487] "2_TGATATGATATTCC-1" "2_AAACGGCTGTTCGA-1" "2_AACTCGGATATCGG-1"
#> [490] "2_AAGTAACTTCAGAC-1" "2_AATAGGGACACTGA-1" "2_ACGCGGTGCTCATT-1"
#> [493] "1_CATTACACATGCTG-1" "1_AACAGAGAGCCCTT-1" "2_GCGAGAGAGAAACA-1"
#> [496] "2_TCTATGTGATGGTC-1" "2_CAAGCCCTGCTATG-1" "2_ATTCCATGCCTAAG-1"
#> [499] "1_TTCACAACTCTATC-1" "2_AATTCCTGATCGTG-1" "1_CCAAAGTGCATTCT-1"
#> [502] "2_ACCCAAGAGCCAAT-1" "2_TTGTCATGCCCGTT-1" "2_TTGAGGTGACGCAT-1"
#> [505] "2_GCTATACTCTGGTA-1" "1_CTATTGACGATAAG-1" "2_TCAGCGCTAGAGTA-1"
#> [508] "2_ATGAAGGAGGAAAT-1" "2_GACTACGACAATCG-1" "2_TTAGTCACGAACCT-1"
#> [511] "2_CGTTATACGTCATG-1" "2_ATGTCACTAGTACC-1" "2_TGACACGACACAAC-1"
#> [514] "2_CATAAAACCTCTAT-1" "1_ATGCGATGGTCTTT-1" "2_ACCGCGGAGCCATA-1"
#> [517] "2_TGAGCTGATGTGCA-1" "2_GATCGAACCCACAA-1" "1_TGGCAATGGTACAC-1"
#> [520] "2_CCGAAAACGGGACA-1" "1_CCCGAACTACCTAG-1" "2_GCGAAGGAGGTACT-1"
#> [523] "2_CCTTAATGCGTACA-1" "2_CTCAGCACGCTATG-1" "2_GACGCCGACTGAAC-1"
#> [526] "2_CCCAACTGAGAGGC-1" "2_CGTAACGATCGTGA-1" "2_GTACTACTAGTACC-1"
#> [529] "1_TTCAGTTGGGCAAG-1" "2_GAACGTTGGACGTT-1" "2_CCCTAGTGCTACCC-1"
#> [532] "2_ATACAATGTGGTAC-1" "2_GAGGACGAGAATCC-1" "2_CATAACCTGGTACT-1"
#> [535] "2_AAGCCATGTCATTC-1" "2_GCATGATGAAACGA-1" "2_CCATAGGATGCATG-1"
#> [538] "2_ACAATTGAGCATCA-1" "2_TCGTAGGAAAACAG-1" "2_ATCATCTGTCACCC-1"
#> [541] "2_AGAATGGATATGGC-1" "2_AATGGAGACTATGG-1" "2_GAAGCTACGAGACG-1"
#> [544] "2_AAGTCTCTTGTTCT-1" "2_ACATTCTGGGTAGG-1" "2_ACTGTGGACTGCAA-1"
#> [547] "2_GACCTCTGTGCAGT-1" "2_ACTCTATGTAACCG-1" "2_GGCACTCTACTCAG-1"
#> [550] "2_ACGTCAGATCGCTC-1" "2_TGGAAAGAGTACAC-1" "1_AGATCGTGGCTGTA-1"
#> [553] "1_GTACGTGACTCAGA-1" "2_GTACTACTGTCATG-1" "2_ATCTGTTGTTACCT-1"
#> [556] "1_TGTACTTGCGAGAG-1" "2_TATCGACTCTCAGA-1" "2_AGGCAGGAGACTAC-1"
#> [559] "2_ATCATGCTTTATCC-1" "1_AAGTTATGGTTGTG-1" "2_ATAACATGGTAGGG-1"
#> [562] "1_ACACGTGATTCTAC-1" "1_CCCGAACTGGCATT-1" "2_TAGCCCTGTACTCT-1"
#> [565] "2_GGAGTTACACGCTA-1" "2_CGCACTTGGAAACA-1" "2_ATGCACGAGTAAAG-1"
#> [568] "2_GTTAAATGGCAGAG-1" "2_CTTCACCTCGAATC-1" "2_CTGAGAACGGGACA-1"
#> [571] "2_AACGGTACGGATCT-1" "1_GCATTGGAGTGTAC-1" "2_CCCTCAGACTCTTA-1"
#> [574] "2_TCAGACGACTAAGC-1" "2_TCGACCTGGTCTGA-1" "2_CATTAGCTACCTGA-1"
#> [577] "1_GAGCGCTGCTCATT-1" "2_CTTGATTGCTTGTT-1" "2_ATGTTGCTCTGTCC-1"
#> [580] "2_TTCAAAGACTTGTT-1" "1_GCCTGACTATGCCA-1" "1_AGCCTCACAAGTGA-1"
#> [583] "2_ACAATCCTCGAATC-1" "2_TATGCGGAGCGTTA-1" "2_TGATTCTGGTTGAC-1"
#> [586] "1_TTCAGACTTCGTGA-1" "2_CGCAAATGCGTTGA-1" "1_CAAGTCGATTCTAC-1"
#> [589] "2_AAGATTACGAGCAG-1" "2_TATCTTCTCGCCTT-1"