PCA

Select top variable genes and perform prcomp

PCA_scGPS(expression.matrix = NULL, ngenes = 1500, scaling = TRUE,
  npcs = 50)

Arguments

expression.matrix
An expression matrix, with genes in rows

ngenes
number of genes used for clustering calculations.

scaling
a logical of whether we want to scale the matrix

npcs
an integer specifying the number of principal components to use.

Value

a list containing PCA results and variance explained

Examples

day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
t <-PCA_scGPS(expression.matrix=assay(mixedpop1))

#> [1] "Preparing PCA inputs using the top 1500 genes ..."
#> [1] "Computing PCA values..."