annotate_scGPS functionally annotates the identified clusters

often we need to label clusters with unique biological characters. One of the common approach to annotate a cluster is to perform functional enrichment analysis. The annotate_scGPS implements ReactomePA and clusterProfiler for this analysis type in R. The function require installation of several databases as described below.

annotate_scGPS(DEgeneList, pvalueCutoff = 0.05, gene_symbol = TRUE,
  output_filename = "PathwayEnrichment.xlsx", output_path = NULL)

Arguments

DEgeneList
is a vector of gene symbols, convertable to ENTREZID

Value

write enrichment test output to a file and an enrichment test object for plotting

Examples

genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap_scGPS(nboots = 2,mixedpop1 = mixedpop1, mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list())

#> Error in colData(mixedpop1): object 'mixedpop1' not found

enrichment_test <- annotate_scGPS(genes$Merged_unique, pvalueCutoff=0.05, gene_symbol=TRUE,output_filename = "PathwayEnrichment.xlsx", output_path = NULL )

#> Error in genes$Merged_unique: $ operator is invalid for atomic vectors

dotplot(enrichment_test, showCategory=15)

#> Error in dotplot(enrichment_test, showCategory = 15): object 'enrichment_test' not found