BootStrap runs for both scGPS training and prediction

ElasticNet and LDA prediction for each of all the subpopulations in the new mixed population after training the model for a subpopulation in the first mixed population. The number of bootstraps to be run can be specified.

bootstrap(nboots = 1, genes = genes, mixedpop1 = mixedpop1,
  mixedpop2 = mixedpop2, c_selectID = NULL, listData = list(),
  cluster_mixedpop1 = NULL, cluster_mixedpop2 = NULL,
  trainset_ratio = 0.5, LDA_run = TRUE)

Arguments

nboots
a number specifying how many bootstraps to be run

genes
a gene list to build the model

mixedpop1
a SingleCellExperiment object from a mixed population for training

mixedpop2
a SingleCellExperiment object from a target mixed population for prediction

c_selectID
the root cluster in mixedpop1 to becompared to clusters in mixedpop2

listData
a list object, which contains trained results for the first mixed population

cluster_mixedpop1
a vector of cluster assignment for mixedpop1

cluster_mixedpop2
a vector of cluster assignment for mixedpop2

trainset_ratio
a number specifying the proportion of cells to be part of the training subpopulation

LDA_run
logical, if the LDA prediction is added to compare to ElasticNet

Value

a list with prediction results written in to the index out_idx

See also

Examples

day2 <- sample1
mixedpop1 <-new_scGPS_object(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-new_scGPS_object(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
cluster_mixedpop1 <- colData(mixedpop1)[,1]
cluster_mixedpop2 <- colData(mixedpop2)[,1]
c_selectID <- 2
test <- bootstrap(nboots = 1, mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, listData =list(),
    cluster_mixedpop1 = cluster_mixedpop1,
    cluster_mixedpop2 = cluster_mixedpop2, c_selectID = c_selectID)

#> Total 203 cells as source subpop

#> Total 387 cells in remaining subpops

#> subsampling 102 cells for training source subpop

#> subsampling 102 cells in remaining subpops for training

#> use 6 genes for training model

#> use 6 genes 204 cells for testing model

#> rename remaining subpops to 1_3

#> there are 102 cells in class 1_3 and 102 cells in class 2

#> standardizing prediction/target dataset

#> performning elasticnet model training...

#> performning LDA model training...

#> Loading required package: lattice

#> Loading required package: ggplot2

#> 
#> Attaching package: 'ggplot2'

#> The following object is masked from 'package:scGPS':
#> 
#>     annotate

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> extracting deviance and best gene features...

#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df       %Dev   Lambda
#>  [1,]  0 -3.684e-15 0.112100
#>  [2,]  2  9.346e-03 0.102100
#>  [3,]  2  1.799e-02 0.093030
#>  [4,]  2  2.519e-02 0.084770
#>  [5,]  2  3.120e-02 0.077240
#>  [6,]  2  3.623e-02 0.070380
#>  [7,]  3  4.066e-02 0.064130
#>  [8,]  3  4.445e-02 0.058430
#>  [9,]  3  4.763e-02 0.053240
#> [10,]  3  5.030e-02 0.048510
#> [11,]  4  5.278e-02 0.044200
#> [12,]  4  5.499e-02 0.040270
#> [13,]  5  5.747e-02 0.036690
#> [14,]  5  5.966e-02 0.033430
#> [15,]  5  6.152e-02 0.030460
#> [16,]  5  6.308e-02 0.027760
#> [17,]  5  6.440e-02 0.025290
#> [18,]  5  6.552e-02 0.023050
#> [19,]  5  6.646e-02 0.021000
#> [20,]  5  6.725e-02 0.019130
#> [21,]  5  6.791e-02 0.017430
#> [22,]  5  6.847e-02 0.015880
#> [23,]  5  6.894e-02 0.014470
#> [24,]  5  6.933e-02 0.013190
#> [25,]  5  6.966e-02 0.012020
#> [26,]  5  6.994e-02 0.010950
#> [27,]  5  7.017e-02 0.009976
#> [28,]  5  7.037e-02 0.009090
#> [29,]  5  7.053e-02 0.008282
#> [30,]  5  7.067e-02 0.007546
#> [31,]  5  7.078e-02 0.006876
#> [32,]  5  7.088e-02 0.006265
#> [33,]  5  7.096e-02 0.005709
#> [34,]  5  7.102e-02 0.005201
#> [35,]  5  7.108e-02 0.004739
#> [36,]  5  7.112e-02 0.004318
#> [37,]  5  7.116e-02 0.003935
#> [38,]  5  7.120e-02 0.003585
#> [39,]  5  7.122e-02 0.003267
#> [40,]  5  7.124e-02 0.002976
#> [41,]  5  7.126e-02 0.002712
#> [42,]  5  7.128e-02 0.002471
#> [43,]  5  7.129e-02 0.002252
#> [44,]  5  7.130e-02 0.002052
#> [45,]  5  7.131e-02 0.001869

#> lambda min is at location 39

#> the leave-out cells in the source subpop is 101

#> use 102 target subpops cells for leave-out test set

#> standardizing the leave-out target and source subpops...

#> start ElasticNet prediction for estimating accuracy...

#> start LDA prediction for estimating accuracy for 195 cells and 6 genes...

#> evaluation accuracy ElasticNet 0.661538461538462

#> evaluation accuracy LDA 0.656410256410256

#> done training for bootstrap 1, moving to prediction...

#> standardizing target subpops before prediction...

#> predicting from source to target subpop 1...

#> number of cells in the target subpop 1 is 187

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running elasticNet classification...

#> ElasticNet for target subpop 1 has no solution

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 1 is 68.4491978609626

#> predicting from source to target subpop 2...

#> number of cells in the target subpop 2 is 140

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 2 is 8.57142857142857

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 2 is 4.28571428571429

#> predicting from source to target subpop 3...

#> number of cells in the target subpop 3 is 133

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 3 is 81.9548872180451

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 3 is 50.3759398496241

#> predicting from source to target subpop 4...

#> number of cells in the target subpop 4 is 40

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 4 is 70

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 4 is 35

names(test)

#> [1] "Accuracy"          "ElasticNetGenes"   "Deviance"         
#> [4] "ElasticNetFit"     "LDAFit"            "predictor_S1"     
#> [7] "ElasticNetPredict" "LDAPredict"       

test$ElasticNetPredict

#> [[1]]
#> [[1]][[1]]
#> NULL
#> 
#> [[1]][[2]]
#> NULL
#> 
#> [[1]][[3]]
#> NULL
#> 
#> [[1]][[4]]
#> NULL
#> 
#> [[1]][[5]]
#> [1] "ElasticNet for subpop1 in target mixedpop2"
#> 
#> [[1]][[6]]
#> [1] "NA"
#> 
#> [[1]][[7]]
#> [1] "ElasticNet for subpop2 in target mixedpop2"
#> 
#> [[1]][[8]]
#> [1] 8.571429
#> 
#> [[1]][[9]]
#> [1] "ElasticNet for subpop3 in target mixedpop2"
#> 
#> [[1]][[10]]
#> [1] 81.95489
#> 
#> [[1]][[11]]
#> [1] "ElasticNet for subpop4 in target mixedpop2"
#> 
#> [[1]][[12]]
#> [1] 70
#> 
#> 

test$LDAPredict

#> [[1]]
#> [[1]][[1]]
#> NULL
#> 
#> [[1]][[2]]
#> NULL
#> 
#> [[1]][[3]]
#> NULL
#> 
#> [[1]][[4]]
#> NULL
#> 
#> [[1]][[5]]
#> [1] "LDA for subpop 1 in target mixedpop2"
#> 
#> [[1]][[6]]
#> [1] 68.4492
#> 
#> [[1]][[7]]
#> [1] "LDA for subpop 2 in target mixedpop2"
#> 
#> [[1]][[8]]
#> [1] 4.285714
#> 
#> [[1]][[9]]
#> [1] "LDA for subpop 3 in target mixedpop2"
#> 
#> [[1]][[10]]
#> [1] 50.37594
#> 
#> [[1]][[11]]
#> [1] "LDA for subpop 4 in target mixedpop2"
#> 
#> [[1]][[12]]
#> [1] 35
#> 
#>