BootStrap runs for both scGPS training and prediction with parallel option

same as bootstrap_prediction, but with an multicore option

bootstrap_parallel(
  ncores = 4,
  nboots = 1,
  genes = genes,
  mixedpop1 = mixedpop1,
  mixedpop2 = mixedpop2,
  c_selectID,
  listData = list(),
  cluster_mixedpop1 = NULL,
  cluster_mixedpop2 = NULL
)

Arguments

ncores
a number specifying how many cpus to be used for running

nboots
a number specifying how many bootstraps to be run

genes
a gene list to build the model

mixedpop1
a SingleCellExperiment object from a mixed population for training

mixedpop2
a SingleCellExperiment object from a target mixed population for prediction

c_selectID
the root cluster in mixedpop1 to becompared to clusters in mixedpop2

listData
a list object, which contains trained results for the first mixed population

cluster_mixedpop1
a vector of cluster assignment for mixedpop1

cluster_mixedpop2
a vector of cluster assignment for mixedpop2

Value

a list with prediction results written in to the index out_idx

Author

Quan Nguyen, 2017-11-25

Examples

day2 <- day_2_cardio_cell_sample
mixedpop1 <-new_scGPS_object(ExpressionMatrix = day2$dat2_counts, 
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- day_5_cardio_cell_sample
mixedpop2 <-new_scGPS_object(ExpressionMatrix = day5$dat5_counts, 
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-training_gene_sample
genes <-genes$Merged_unique
#prl_boots <- bootstrap_parallel(ncores = 4, nboots = 1, genes=genes,
#    mixedpop1 = mixedpop2, mixedpop2 = mixedpop2,  c_selectID=1,
#    listData =list())
#prl_boots[[1]]$ElasticNetPredict
#prl_boots[[1]]$LDAPredict