BootStrap runs for both scGPS training and prediction

ElasticNet and LDA prediction for each of all the subpopulations in the new mixed population after training the model for a subpopulation in the first mixed population. The number of bootstraps to be run can be specified.

bootstrap_scGPS(nboots = 1, genes = genes, mixedpop1 = mixedpop1,
  mixedpop2 = mixedpop2, c_selectID = NULL, listData = list(),
  cluster_mixedpop1 = NULL, cluster_mixedpop2 = NULL,
  trainset_ratio = 0.5, LDA_run = TRUE)

Arguments

nboots
a number specifying how many bootstraps to be run

genes
a gene list to build the model

mixedpop1
a SingleCellExperiment object from a mixed population for training

mixedpop2
a SingleCellExperiment object from a target mixed population for prediction

c_selectID
the root cluster in mixedpop1 to becompared to clusters in mixedpop2

listData
a list object, which contains trained results for the first mixed population

cluster_mixedpop1
a vector of cluster assignment for mixedpop1

cluster_mixedpop2
a vector of cluster assignment for mixedpop2

trainset_ratio
a number specifying the proportion of cells to be part of the training subpopulation

LDA_run
logical, if the LDA prediction is added to compare to ElasticNet

Value

a list with prediction results written in to the index out_idx

See also

Examples

day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
cluster_mixedpop1 <- colData(mixedpop1)[,1]
cluster_mixedpop2 <- colData(mixedpop2)[,1]
c_selectID <- 2
test <- bootstrap_scGPS(nboots = 2, mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, listData =list(),
    cluster_mixedpop1 = cluster_mixedpop1,
    cluster_mixedpop2 = cluster_mixedpop2, c_selectID = c_selectID)

#> [1] "Total 203 cells as source subpop"
#> [1] "Total 387 cells in remaining subpops"
#> [1] "subsampling 102 cells for training source subpop"
#> [1] "subsampling 102 cells in remaining subpops for training"
#> [1] "use 107 genes for training model"
#> [1] "use 107 genes 204 cells for testing model"
#> [1] "rename remaining subpops to 1_3"
#> [1] "there are 102 cells in class 1_3 and 102 cells in class 2"
#> [1] "removing 10 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = y_cat, family = "binomial") 
#> 
#>        Df      %Dev    Lambda
#>   [1,]  0 1.281e-15 3.421e-01
#>   [2,]  1 5.742e-02 3.117e-01
#>   [3,]  1 1.056e-01 2.840e-01
#>   [4,]  2 1.513e-01 2.588e-01
#>   [5,]  2 1.926e-01 2.358e-01
#>   [6,]  2 2.279e-01 2.148e-01
#>   [7,]  2 2.584e-01 1.957e-01
#>   [8,]  2 2.848e-01 1.784e-01
#>   [9,]  2 3.077e-01 1.625e-01
#>  [10,]  2 3.277e-01 1.481e-01
#>  [11,]  2 3.450e-01 1.349e-01
#>  [12,]  3 3.616e-01 1.229e-01
#>  [13,]  4 3.798e-01 1.120e-01
#>  [14,]  4 3.975e-01 1.021e-01
#>  [15,]  6 4.162e-01 9.300e-02
#>  [16,]  7 4.363e-01 8.473e-02
#>  [17,]  7 4.546e-01 7.721e-02
#>  [18,]  8 4.712e-01 7.035e-02
#>  [19,]  9 4.864e-01 6.410e-02
#>  [20,]  9 5.003e-01 5.840e-02
#>  [21,] 12 5.138e-01 5.322e-02
#>  [22,] 12 5.275e-01 4.849e-02
#>  [23,] 16 5.417e-01 4.418e-02
#>  [24,] 18 5.604e-01 4.026e-02
#>  [25,] 20 5.780e-01 3.668e-02
#>  [26,] 21 5.946e-01 3.342e-02
#>  [27,] 21 6.096e-01 3.045e-02
#>  [28,] 23 6.242e-01 2.775e-02
#>  [29,] 24 6.378e-01 2.528e-02
#>  [30,] 26 6.511e-01 2.304e-02
#>  [31,] 29 6.652e-01 2.099e-02
#>  [32,] 31 6.791e-01 1.912e-02
#>  [33,] 32 6.932e-01 1.743e-02
#>  [34,] 33 7.064e-01 1.588e-02
#>  [35,] 34 7.187e-01 1.447e-02
#>  [36,] 36 7.308e-01 1.318e-02
#>  [37,] 39 7.435e-01 1.201e-02
#>  [38,] 40 7.557e-01 1.094e-02
#>  [39,] 41 7.677e-01 9.972e-03
#>  [40,] 43 7.790e-01 9.086e-03
#>  [41,] 44 7.901e-01 8.279e-03
#>  [42,] 45 8.008e-01 7.543e-03
#>  [43,] 45 8.110e-01 6.873e-03
#>  [44,] 46 8.218e-01 6.263e-03
#>  [45,] 49 8.328e-01 5.706e-03
#>  [46,] 50 8.432e-01 5.199e-03
#>  [47,] 50 8.528e-01 4.737e-03
#>  [48,] 51 8.619e-01 4.317e-03
#>  [49,] 55 8.706e-01 3.933e-03
#>  [50,] 56 8.791e-01 3.584e-03
#>  [51,] 58 8.872e-01 3.265e-03
#>  [52,] 60 8.953e-01 2.975e-03
#>  [53,] 61 9.029e-01 2.711e-03
#>  [54,] 62 9.103e-01 2.470e-03
#>  [55,] 62 9.172e-01 2.251e-03
#>  [56,] 63 9.240e-01 2.051e-03
#>  [57,] 63 9.305e-01 1.869e-03
#>  [58,] 63 9.365e-01 1.703e-03
#>  [59,] 63 9.421e-01 1.551e-03
#>  [60,] 62 9.472e-01 1.413e-03
#>  [61,] 62 9.519e-01 1.288e-03
#>  [62,] 62 9.562e-01 1.173e-03
#>  [63,] 63 9.601e-01 1.069e-03
#>  [64,] 64 9.637e-01 9.742e-04
#>  [65,] 62 9.670e-01 8.877e-04
#>  [66,] 62 9.700e-01 8.088e-04
#>  [67,] 62 9.727e-01 7.370e-04
#>  [68,] 62 9.751e-01 6.715e-04
#>  [69,] 63 9.773e-01 6.119e-04
#>  [70,] 63 9.793e-01 5.575e-04
#>  [71,] 63 9.812e-01 5.080e-04
#>  [72,] 63 9.829e-01 4.628e-04
#>  [73,] 63 9.844e-01 4.217e-04
#>  [74,] 63 9.858e-01 3.843e-04
#>  [75,] 64 9.870e-01 3.501e-04
#>  [76,] 66 9.882e-01 3.190e-04
#>  [77,] 66 9.892e-01 2.907e-04
#>  [78,] 66 9.902e-01 2.649e-04
#>  [79,] 66 9.910e-01 2.413e-04
#>  [80,] 66 9.919e-01 2.199e-04
#>  [81,] 67 9.926e-01 2.004e-04
#>  [82,] 66 9.932e-01 1.826e-04
#>  [83,] 66 9.938e-01 1.663e-04
#>  [84,] 66 9.944e-01 1.516e-04
#>  [85,] 66 9.949e-01 1.381e-04
#>  [86,] 66 9.953e-01 1.258e-04
#>  [87,] 68 9.957e-01 1.147e-04
#>  [88,] 68 9.961e-01 1.045e-04
#>  [89,] 68 9.964e-01 9.518e-05
#>  [90,] 68 9.968e-01 8.673e-05
#>  [91,] 68 9.970e-01 7.902e-05
#>  [92,] 69 9.973e-01 7.200e-05
#>  [93,] 69 9.975e-01 6.561e-05
#>  [94,] 69 9.977e-01 5.978e-05
#>  [95,] 69 9.979e-01 5.447e-05
#>  [96,] 69 9.981e-01 4.963e-05
#>  [97,] 70 9.983e-01 4.522e-05
#>  [98,] 70 9.984e-01 4.120e-05
#>  [99,] 70 9.986e-01 3.754e-05
#> [100,] 70 9.987e-01 3.421e-05
#> [1] "lambda min is at location 17"
#> [1] "the leave-out cells in the source subpop is 101"
#> [1] "use 102 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 203 cells and 97 genes..."
#> [1] "evaluation accuracy ElasticNet 0.901477832512315"
#> [1] "evaluation accuracy LDA 0.827586206896552"
#> [1] "done training for bootstrap 1, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 17305"
#> [1] "Number of genes in the model present in the target data is 97"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 97 genes and 187 cells. The trained model has 97 genes"
#> [1] "first 10 genes in model "
#>  [1] "NPPB"   "FABP3"  "LCK"    "DMBX1"  "ABCA4"  "TPM3"   "RGS4"   "RXRG"  
#>  [9] "NR5A2"  "LEFTY1"
#> [1] "first 10 genes in target "
#>  [1] "NPPB"   "FABP3"  "LCK"    "DMBX1"  "ABCA4"  "TPM3"   "RGS4"   "RXRG"  
#>  [9] "NR5A2"  "LEFTY1"
#> [1] "running elasticNet classification..."

#> Error in which(ElasticNet_result[, 1] == c_selectID): object 'c_selectID' not found

names(test)

#> Error in eval(expr, envir, enclos): object 'test' not found

test$ElasticNetPredict

#> Error in eval(expr, envir, enclos): object 'test' not found

test$LDAPredict

#> Error in eval(expr, envir, enclos): object 'test' not found