Main prediction function applying the optimal ElasticNet and LDA models

Predict a new mixed population after training the model for a subpopulation in the first mixed population. All subpopulations in the new target mixed population will be predicted, where each targeted subpopulation will have a transition score from the orginal subpopulation to the new subpopulation.

predicting_scGPS(listData = NULL, cluster_mixedpop2 = NULL,
  mixedpop2 = NULL, out_idx = NULL, standardize = TRUE,
  LDA_run = FALSE)

Arguments

listData
a list object containing trained results for the selected subpopulation in the first mixed population

cluster_mixedpop2
a vector of cluster assignment for mixedpop2

mixedpop2
a SingleCellExperiment object from the target mixed population of importance, e.g. differentially expressed genes that are most significant

out_idx
a number to specify index to write results into the list output. This is needed for running bootstrap.

standardize
a logical of whether to standardize the data

LDA_run
logical, if the LDA prediction is added to compare to ElasticNet, the LDA model needs to be trained from the training_scGPS before inputting to this prediction step

Value

a list with prediction results written in to the index out_idx

Examples

c_selectID<-1
out_idx<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
listData  <- training_scGPS(genes,
    cluster_mixedpop1 = colData(mixedpop1)[, 1], mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, c_selectID, listData =list(), out_idx=out_idx)

#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 107 genes for training model"
#> [1] "use 107 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "removing 11 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = y_cat, family = "binomial") 
#> 
#>        Df       %Dev    Lambda
#>   [1,]  0 -2.563e-15 3.453e-01
#>   [2,]  2  6.516e-02 3.146e-01
#>   [3,]  2  1.224e-01 2.866e-01
#>   [4,]  2  1.710e-01 2.612e-01
#>   [5,]  2  2.125e-01 2.380e-01
#>   [6,]  2  2.483e-01 2.168e-01
#>   [7,]  2  2.792e-01 1.976e-01
#>   [8,]  2  3.061e-01 1.800e-01
#>   [9,]  2  3.295e-01 1.640e-01
#>  [10,]  2  3.499e-01 1.495e-01
#>  [11,]  2  3.678e-01 1.362e-01
#>  [12,]  2  3.834e-01 1.241e-01
#>  [13,]  2  3.972e-01 1.131e-01
#>  [14,]  2  4.092e-01 1.030e-01
#>  [15,]  3  4.216e-01 9.386e-02
#>  [16,]  5  4.348e-01 8.553e-02
#>  [17,]  5  4.472e-01 7.793e-02
#>  [18,]  5  4.582e-01 7.101e-02
#>  [19,]  5  4.680e-01 6.470e-02
#>  [20,]  6  4.768e-01 5.895e-02
#>  [21,]  6  4.853e-01 5.371e-02
#>  [22,]  7  4.942e-01 4.894e-02
#>  [23,]  8  5.067e-01 4.459e-02
#>  [24,]  9  5.184e-01 4.063e-02
#>  [25,] 10  5.307e-01 3.702e-02
#>  [26,] 11  5.419e-01 3.373e-02
#>  [27,] 14  5.545e-01 3.074e-02
#>  [28,] 18  5.679e-01 2.801e-02
#>  [29,] 23  5.843e-01 2.552e-02
#>  [30,] 26  5.997e-01 2.325e-02
#>  [31,] 29  6.147e-01 2.119e-02
#>  [32,] 30  6.296e-01 1.930e-02
#>  [33,] 33  6.442e-01 1.759e-02
#>  [34,] 35  6.585e-01 1.603e-02
#>  [35,] 36  6.724e-01 1.460e-02
#>  [36,] 39  6.859e-01 1.331e-02
#>  [37,] 39  6.988e-01 1.212e-02
#>  [38,] 41  7.109e-01 1.105e-02
#>  [39,] 44  7.227e-01 1.006e-02
#>  [40,] 48  7.353e-01 9.171e-03
#>  [41,] 50  7.480e-01 8.356e-03
#>  [42,] 52  7.606e-01 7.614e-03
#>  [43,] 53  7.728e-01 6.937e-03
#>  [44,] 53  7.844e-01 6.321e-03
#>  [45,] 54  7.956e-01 5.759e-03
#>  [46,] 56  8.063e-01 5.248e-03
#>  [47,] 56  8.167e-01 4.782e-03
#>  [48,] 58  8.269e-01 4.357e-03
#>  [49,] 62  8.370e-01 3.970e-03
#>  [50,] 62  8.467e-01 3.617e-03
#>  [51,] 64  8.563e-01 3.296e-03
#>  [52,] 63  8.654e-01 3.003e-03
#>  [53,] 64  8.742e-01 2.736e-03
#>  [54,] 65  8.825e-01 2.493e-03
#>  [55,] 66  8.904e-01 2.272e-03
#>  [56,] 67  8.987e-01 2.070e-03
#>  [57,] 65  9.060e-01 1.886e-03
#>  [58,] 63  9.128e-01 1.718e-03
#>  [59,] 65  9.190e-01 1.566e-03
#>  [60,] 63  9.251e-01 1.427e-03
#>  [61,] 62  9.308e-01 1.300e-03
#>  [62,] 62  9.360e-01 1.184e-03
#>  [63,] 64  9.411e-01 1.079e-03
#>  [64,] 65  9.459e-01 9.833e-04
#>  [65,] 65  9.504e-01 8.960e-04
#>  [66,] 65  9.546e-01 8.164e-04
#>  [67,] 66  9.586e-01 7.439e-04
#>  [68,] 67  9.622e-01 6.778e-04
#>  [69,] 67  9.656e-01 6.176e-04
#>  [70,] 68  9.686e-01 5.627e-04
#>  [71,] 69  9.713e-01 5.127e-04
#>  [72,] 69  9.739e-01 4.672e-04
#>  [73,] 68  9.763e-01 4.257e-04
#>  [74,] 69  9.784e-01 3.879e-04
#>  [75,] 70  9.803e-01 3.534e-04
#>  [76,] 70  9.821e-01 3.220e-04
#>  [77,] 71  9.837e-01 2.934e-04
#>  [78,] 71  9.852e-01 2.673e-04
#>  [79,] 71  9.865e-01 2.436e-04
#>  [80,] 71  9.877e-01 2.219e-04
#>  [81,] 72  9.888e-01 2.022e-04
#>  [82,] 73  9.898e-01 1.843e-04
#>  [83,] 73  9.907e-01 1.679e-04
#>  [84,] 74  9.916e-01 1.530e-04
#>  [85,] 74  9.923e-01 1.394e-04
#>  [86,] 74  9.930e-01 1.270e-04
#>  [87,] 75  9.936e-01 1.157e-04
#>  [88,] 75  9.942e-01 1.054e-04
#>  [89,] 75  9.947e-01 9.607e-05
#>  [90,] 75  9.952e-01 8.754e-05
#>  [91,] 75  9.956e-01 7.976e-05
#>  [92,] 75  9.960e-01 7.268e-05
#>  [93,] 75  9.963e-01 6.622e-05
#>  [94,] 75  9.967e-01 6.034e-05
#>  [95,] 75  9.970e-01 5.498e-05
#>  [96,] 75  9.972e-01 5.009e-05
#>  [97,] 75  9.975e-01 4.564e-05
#>  [98,] 75  9.977e-01 4.159e-05
#>  [99,] 76  9.979e-01 3.789e-05
#> [100,] 76  9.981e-01 3.453e-05
#> [1] "lambda min is at location 14"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "evaluation accuracy ElasticNet 0.857142857142857"

listData  <- predicting_scGPS(listData =listData,  mixedpop2 = mixedpop2,
    out_idx=out_idx, cluster_mixedpop2 = colData(mixedpop2)[, 1])

#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 17306"
#> [1] "Number of genes in the model present in the target data is 96"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 96 genes and 187 cells. The trained model has 96 genes"
#> [1] "first 10 genes in model "
#>  [1] "FABP3"  "LCK"    "DMBX1"  "ABCA4"  "TPM3"   "RGS4"   "RXRG"   "NR5A2" 
#>  [9] "LEFTY1" "LEFTY2"
#> [1] "first 10 genes in target "
#>  [1] "FABP3"  "LCK"    "DMBX1"  "ABCA4"  "TPM3"   "RGS4"   "RXRG"   "NR5A2" 
#>  [9] "LEFTY1" "LEFTY2"
#> [1] "running elasticNet classification..."

#> Error in which(ElasticNet_result[, 1] == c_selectID): object 'c_selectID' not found