summarise bootstrap runs for Lasso model, from n bootstraps

the training and prediction results from bootstrap were written to the object LSOLDA_dat, the reformat_LASSO summarises prediction for n bootstrap runs

reformat_LASSO(c_selectID = NULL, mp_selectID = NULL,
  LSOLDA_dat = NULL, nPredSubpop = NULL, Nodes_group = "#7570b3",
  nboots = 2)

Arguments

c_selectID
is the original cluster to be projected

mp_selectID
is the target mixedpop to project to

LSOLDA_dat
is the results from the bootstrap

nPredSubpop
is the number of clusters in the target mixedpop row_cluster <-length(unique(target_cluster))

Nodes_group
string representation of hexidecimal color code for node

nboots
is an integer for how many bootstraps are run

Value

a dataframe containg information for the Lasso prediction results, each column contains prediction results for all subpopulations from each bootstrap run

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-new_scGPS_object(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-new_scGPS_object(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap(nboots = 2, mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> Total 224 cells as source subpop

#> Total 366 cells in remaining subpops

#> subsampling 112 cells for training source subpop

#> subsampling 112 cells in remaining subpops for training

#> use 6 genes for training model

#> use 6 genes 224 cells for testing model

#> rename remaining subpops to 2_3

#> there are 112 cells in class 2_3 and 112 cells in class 1

#> removing 1 genes with no variance

#> standardizing prediction/target dataset

#> performning elasticnet model training...

#> performning LDA model training...

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> extracting deviance and best gene features...

#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev   Lambda
#>  [1,]  0 2.243e-15 0.102100
#>  [2,]  2 6.663e-03 0.093010
#>  [3,]  2 1.439e-02 0.084750
#>  [4,]  2 2.089e-02 0.077220
#>  [5,]  2 2.637e-02 0.070360
#>  [6,]  2 3.103e-02 0.064110
#>  [7,]  2 3.498e-02 0.058410
#>  [8,]  2 3.835e-02 0.053220
#>  [9,]  2 4.123e-02 0.048500
#> [10,]  2 4.368e-02 0.044190
#> [11,]  2 4.578e-02 0.040260
#> [12,]  3 4.815e-02 0.036690
#> [13,]  3 5.019e-02 0.033430
#> [14,]  3 5.194e-02 0.030460
#> [15,]  3 5.345e-02 0.027750
#> [16,]  4 5.478e-02 0.025290
#> [17,]  4 5.620e-02 0.023040
#> [18,]  4 5.741e-02 0.020990
#> [19,]  4 5.845e-02 0.019130
#> [20,]  4 5.933e-02 0.017430
#> [21,]  4 6.008e-02 0.015880
#> [22,]  4 6.072e-02 0.014470
#> [23,]  4 6.126e-02 0.013180
#> [24,]  4 6.172e-02 0.012010
#> [25,]  4 6.211e-02 0.010950
#> [26,]  4 6.244e-02 0.009973
#> [27,]  4 6.272e-02 0.009087
#> [28,]  4 6.296e-02 0.008280
#> [29,]  4 6.316e-02 0.007544
#> [30,]  4 6.333e-02 0.006874
#> [31,]  4 6.347e-02 0.006263
#> [32,]  4 6.359e-02 0.005707
#> [33,]  4 6.369e-02 0.005200
#> [34,]  4 6.377e-02 0.004738
#> [35,]  4 6.384e-02 0.004317
#> [36,]  4 6.390e-02 0.003934
#> [37,]  4 6.395e-02 0.003584
#> [38,]  4 6.399e-02 0.003266
#> [39,]  4 6.403e-02 0.002976
#> [40,]  4 6.406e-02 0.002711
#> [41,]  4 6.408e-02 0.002470
#> [42,]  4 6.410e-02 0.002251
#> [43,]  4 6.412e-02 0.002051
#> [44,]  4 6.413e-02 0.001869
#> [45,]  4 6.414e-02 0.001703
#> [46,]  4 6.415e-02 0.001552

#> lambda min is at location 17

#> the leave-out cells in the source subpop is 112

#> use 112 target subpops cells for leave-out test set

#> standardizing the leave-out target and source subpops...

#> start ElasticNet prediction for estimating accuracy...

#> start LDA prediction for estimating accuracy for 215 cells and 5 genes...

#> evaluation accuracy ElasticNet 0.655813953488372

#> evaluation accuracy LDA 0.651162790697674

#> done training for bootstrap 1, moving to prediction...

#> standardizing target subpops before prediction...

#> predicting from source to target subpop 1...

#> number of cells in the target subpop 1 is 187

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 1 is 14.4385026737968

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 1 is 53.475935828877

#> predicting from source to target subpop 2...

#> number of cells in the target subpop 2 is 140

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 2 is 19.2857142857143

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 2 is 12.1428571428571

#> predicting from source to target subpop 3...

#> number of cells in the target subpop 3 is 133

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 3 is 83.4586466165414

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 3 is 60.1503759398496

#> predicting from source to target subpop 4...

#> number of cells in the target subpop 4 is 40

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 4 is 67.5

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 4 is 45

#> Total 224 cells as source subpop

#> Total 366 cells in remaining subpops

#> subsampling 112 cells for training source subpop

#> subsampling 112 cells in remaining subpops for training

#> use 6 genes for training model

#> use 6 genes 224 cells for testing model

#> rename remaining subpops to 2_3

#> there are 112 cells in class 2_3 and 112 cells in class 1

#> removing 1 genes with no variance

#> standardizing prediction/target dataset

#> performning elasticnet model training...

#> performning LDA model training...

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> extracting deviance and best gene features...

#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df       %Dev   Lambda
#>  [1,]  0 -1.922e-15 0.092070
#>  [2,]  1  4.155e-03 0.083890
#>  [3,]  2  8.071e-03 0.076440
#>  [4,]  2  1.412e-02 0.069650
#>  [5,]  2  1.919e-02 0.063460
#>  [6,]  2  2.346e-02 0.057820
#>  [7,]  3  2.818e-02 0.052690
#>  [8,]  3  3.218e-02 0.048010
#>  [9,]  3  3.559e-02 0.043740
#> [10,]  3  3.850e-02 0.039860
#> [11,]  3  4.099e-02 0.036320
#> [12,]  3  4.311e-02 0.033090
#> [13,]  3  4.493e-02 0.030150
#> [14,]  3  4.648e-02 0.027470
#> [15,]  3  4.781e-02 0.025030
#> [16,]  3  4.894e-02 0.022810
#> [17,]  4  4.992e-02 0.020780
#> [18,]  4  5.075e-02 0.018930
#> [19,]  4  5.146e-02 0.017250
#> [20,]  4  5.206e-02 0.015720
#> [21,]  4  5.258e-02 0.014320
#> [22,]  4  5.301e-02 0.013050
#> [23,]  4  5.338e-02 0.011890
#> [24,]  4  5.369e-02 0.010840
#> [25,]  4  5.395e-02 0.009873
#> [26,]  4  5.417e-02 0.008996
#> [27,]  4  5.436e-02 0.008196
#> [28,]  4  5.452e-02 0.007468
#> [29,]  4  5.465e-02 0.006805
#> [30,]  4  5.476e-02 0.006200
#> [31,]  4  5.486e-02 0.005649
#> [32,]  4  5.493e-02 0.005148
#> [33,]  4  5.500e-02 0.004690
#> [34,]  4  5.506e-02 0.004274
#> [35,]  4  5.510e-02 0.003894
#> [36,]  4  5.514e-02 0.003548
#> [37,]  4  5.517e-02 0.003233
#> [38,]  4  5.520e-02 0.002946
#> [39,]  4  5.522e-02 0.002684
#> [40,]  4  5.524e-02 0.002446
#> [41,]  4  5.526e-02 0.002228
#> [42,]  4  5.527e-02 0.002030
#> [43,]  4  5.528e-02 0.001850
#> [44,]  4  5.529e-02 0.001686

#> lambda min is at location 32

#> the leave-out cells in the source subpop is 112

#> use 112 target subpops cells for leave-out test set

#> standardizing the leave-out target and source subpops...

#> start ElasticNet prediction for estimating accuracy...

#> start LDA prediction for estimating accuracy for 220 cells and 5 genes...

#> evaluation accuracy ElasticNet 0.668181818181818

#> evaluation accuracy LDA 0.659090909090909

#> done training for bootstrap 2, moving to prediction...

#> standardizing target subpops before prediction...

#> predicting from source to target subpop 1...

#> number of cells in the target subpop 1 is 187

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 1 is 61.4973262032086

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 1 is 55.0802139037433

#> predicting from source to target subpop 2...

#> number of cells in the target subpop 2 is 140

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 2 is 95

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 2 is 50

#> predicting from source to target subpop 3...

#> number of cells in the target subpop 3 is 133

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 3 is 96.9924812030075

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 3 is 64.6616541353383

#> predicting from source to target subpop 4...

#> number of cells in the target subpop 4 is 40

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 4 is 90

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 4 is 52.5

reformat_LASSO(LSOLDA_dat=LSOLDA_dat,
    nPredSubpop=length(unique(colData(mixedpop2)[,1])), c_selectID = 1,
    mp_selectID =2, nboots = 2)

#>          Boostrap1        Boostrap2 Target Source   Node NodeGroup    Value
#> 1 14.4385026737968 61.4973262032086 C1_MP2 C1_MP2 C1_MP2   #7570b3 37.96791
#> 2 19.2857142857143               95 C2_MP2 C1_MP2 C1_MP2   #7570b3 57.14286
#> 3 83.4586466165414 96.9924812030075 C3_MP2 C1_MP2 C1_MP2   #7570b3 90.22556
#> 4             67.5               90 C4_MP2 C1_MP2 C1_MP2   #7570b3 78.75000