get percent accuracy for Lasso model, from n bootstraps

The training results from training were written to

summary_accuracy(object = NULL)

Arguments

object
is a list containing the training results from the summary_accuracy summarise n bootstraps

Value

a vector of percent accuracy for the selected subpopulation

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-new_scGPS_object(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-new_scGPS_object(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap(nboots = 1,mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> Total 224 cells as source subpop

#> Total 366 cells in remaining subpops

#> subsampling 112 cells for training source subpop

#> subsampling 112 cells in remaining subpops for training

#> use 6 genes for training model

#> use 6 genes 224 cells for testing model

#> rename remaining subpops to 2_3

#> there are 112 cells in class 2_3 and 112 cells in class 1

#> standardizing prediction/target dataset

#> performning elasticnet model training...

#> performning LDA model training...

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> extracting deviance and best gene features...

#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev    Lambda
#>  [1,]  0 1.762e-15 0.1654000
#>  [2,]  1 1.341e-02 0.1507000
#>  [3,]  1 2.457e-02 0.1373000
#>  [4,]  1 3.388e-02 0.1251000
#>  [5,]  1 4.166e-02 0.1140000
#>  [6,]  1 4.816e-02 0.1039000
#>  [7,]  1 5.361e-02 0.0946300
#>  [8,]  1 5.817e-02 0.0862200
#>  [9,]  2 6.496e-02 0.0785600
#> [10,]  2 7.103e-02 0.0715800
#> [11,]  2 7.625e-02 0.0652200
#> [12,]  3 8.236e-02 0.0594300
#> [13,]  3 8.788e-02 0.0541500
#> [14,]  3 9.267e-02 0.0493400
#> [15,]  3 9.687e-02 0.0449600
#> [16,]  4 1.006e-01 0.0409600
#> [17,]  4 1.041e-01 0.0373200
#> [18,]  5 1.072e-01 0.0340100
#> [19,]  5 1.100e-01 0.0309900
#> [20,]  5 1.126e-01 0.0282300
#> [21,]  5 1.148e-01 0.0257300
#> [22,]  5 1.168e-01 0.0234400
#> [23,]  5 1.185e-01 0.0213600
#> [24,]  5 1.200e-01 0.0194600
#> [25,]  5 1.214e-01 0.0177300
#> [26,]  5 1.226e-01 0.0161600
#> [27,]  5 1.236e-01 0.0147200
#> [28,]  5 1.245e-01 0.0134100
#> [29,]  5 1.253e-01 0.0122200
#> [30,]  5 1.261e-01 0.0111400
#> [31,]  5 1.267e-01 0.0101500
#> [32,]  5 1.272e-01 0.0092450
#> [33,]  5 1.277e-01 0.0084240
#> [34,]  5 1.281e-01 0.0076760
#> [35,]  5 1.285e-01 0.0069940
#> [36,]  5 1.288e-01 0.0063720
#> [37,]  5 1.290e-01 0.0058060
#> [38,]  5 1.293e-01 0.0052900
#> [39,]  5 1.295e-01 0.0048200
#> [40,]  5 1.296e-01 0.0043920
#> [41,]  5 1.298e-01 0.0040020
#> [42,]  5 1.299e-01 0.0036470
#> [43,]  5 1.300e-01 0.0033230
#> [44,]  5 1.301e-01 0.0030270
#> [45,]  5 1.302e-01 0.0027580
#> [46,]  5 1.302e-01 0.0025130
#> [47,]  5 1.303e-01 0.0022900
#> [48,]  5 1.303e-01 0.0020870
#> [49,]  5 1.304e-01 0.0019010
#> [50,]  5 1.304e-01 0.0017320
#> [51,]  5 1.304e-01 0.0015780
#> [52,]  5 1.305e-01 0.0014380
#> [53,]  5 1.305e-01 0.0013100
#> [54,]  5 1.305e-01 0.0011940
#> [55,]  5 1.305e-01 0.0010880
#> [56,]  5 1.305e-01 0.0009913
#> [57,]  5 1.305e-01 0.0009033

#> lambda min is at location 18

#> the leave-out cells in the source subpop is 112

#> use 112 target subpops cells for leave-out test set

#> standardizing the leave-out target and source subpops...

#> start ElasticNet prediction for estimating accuracy...

#> start LDA prediction for estimating accuracy for 216 cells and 6 genes...

#> evaluation accuracy ElasticNet 0.611111111111111

#> evaluation accuracy LDA 0.615740740740741

#> done training for bootstrap 1, moving to prediction...

#> standardizing target subpops before prediction...

#> predicting from source to target subpop 1...

#> number of cells in the target subpop 1 is 187

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running elasticNet classification...

#> ElasticNet for target subpop 1 has no solution

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 1 is 23.5294117647059

#> predicting from source to target subpop 2...

#> number of cells in the target subpop 2 is 140

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 2 is 39.2857142857143

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 2 is 51.4285714285714

#> predicting from source to target subpop 3...

#> number of cells in the target subpop 3 is 133

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 3 is 85.7142857142857

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 3 is 36.8421052631579

#> predicting from source to target subpop 4...

#> number of cells in the target subpop 4 is 40

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 4 is 72.5

#> Number of genes in the target data, but not in model genes is 4994

#> Number of genes in the model present in the target data is 6

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1MAP2TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 4 is 37.5

summary_accuracy(LSOLDA_dat)

#> [1] 61.57407

summary_deviance(LSOLDA_dat)

#> $allDeviance
#> [1] "0.1072"
#> 
#> $DeviMax
#>          dat_DE$Dfd          Deviance           DEgenes
#> 1                 0            0.1072    genes_cluster1
#> 2                 1            0.1072    genes_cluster1
#> 3                 2            0.1072    genes_cluster1
#> 4                 3            0.1072    genes_cluster1
#> 5                 4            0.1072    genes_cluster1
#> 6                 5            0.1072    genes_cluster1
#> 7 remaining DEgenes remaining DEgenes remaining DEgenes
#> 
#> $LassoGenesMax
#> NULL
#>