get percent deviance explained for Lasso model, from n bootstraps

the training results from training_scGPS were written to the object LSOLDA_dat, the summary_devidance summarises deviance explained for n bootstrap runs and also returns the best deviance matrix for plotting, as well as the best matrix with Lasso genes and coefficients

summary_deviance(object = NULL)

Arguments

object
is a list containing the training results from training_scGPS

Value

a list containing three elements, with a vector of percent maximum deviance explained, a dataframe containg information for the full deviance, and a dataframe containing gene names and coefficients of the best model

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo,
                    CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap_scGPS(nboots = 2,mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "removing 1 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev   Lambda
#>  [1,]  0 2.723e-15 0.185700
#>  [2,]  1 1.690e-02 0.169200
#>  [3,]  1 3.098e-02 0.154200
#>  [4,]  1 4.273e-02 0.140500
#>  [5,]  1 5.255e-02 0.128000
#>  [6,]  1 6.078e-02 0.116600
#>  [7,]  1 6.768e-02 0.106300
#>  [8,]  2 7.587e-02 0.096820
#>  [9,]  2 8.479e-02 0.088220
#> [10,]  2 9.249e-02 0.080380
#> [11,]  2 9.918e-02 0.073240
#> [12,]  2 1.050e-01 0.066740
#> [13,]  2 1.101e-01 0.060810
#> [14,]  2 1.146e-01 0.055410
#> [15,]  2 1.185e-01 0.050480
#> [16,]  2 1.219e-01 0.046000
#> [17,]  2 1.249e-01 0.041910
#> [18,]  2 1.275e-01 0.038190
#> [19,]  2 1.298e-01 0.034800
#> [20,]  2 1.317e-01 0.031710
#> [21,]  2 1.334e-01 0.028890
#> [22,]  2 1.349e-01 0.026320
#> [23,]  2 1.361e-01 0.023980
#> [24,]  2 1.372e-01 0.021850
#> [25,]  2 1.381e-01 0.019910
#> [26,]  2 1.389e-01 0.018140
#> [27,]  2 1.396e-01 0.016530
#> [28,]  3 1.402e-01 0.015060
#> [29,]  3 1.409e-01 0.013720
#> [30,]  3 1.414e-01 0.012510
#> [31,]  3 1.419e-01 0.011390
#> [32,]  4 1.423e-01 0.010380
#> [33,]  4 1.427e-01 0.009460
#> [34,]  4 1.431e-01 0.008619
#> [35,]  4 1.434e-01 0.007854
#> [36,]  4 1.436e-01 0.007156
#> [37,]  4 1.438e-01 0.006520
#> [38,]  4 1.440e-01 0.005941
#> [39,]  4 1.441e-01 0.005413
#> [40,]  4 1.443e-01 0.004932
#> [41,]  4 1.444e-01 0.004494
#> [42,]  4 1.444e-01 0.004095
#> [43,]  4 1.445e-01 0.003731
#> [44,]  4 1.446e-01 0.003400
#> [45,]  4 1.446e-01 0.003098
#> [46,]  4 1.447e-01 0.002822
#> [47,]  4 1.447e-01 0.002572
#> [48,]  4 1.447e-01 0.002343
#> [49,]  4 1.447e-01 0.002135
#> [50,]  4 1.448e-01 0.001945
#> [51,]  4 1.448e-01 0.001773
#> [52,]  4 1.448e-01 0.001615
#> [53,]  4 1.448e-01 0.001472
#> [54,]  4 1.448e-01 0.001341
#> [1] "lambda min is at location 32"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 217 cells and 5 genes..."
#> [1] "evaluation accuracy ElasticNet 0.617511520737327"
#> [1] "evaluation accuracy LDA 0.608294930875576"
#> [1] "done training for bootstrap 1, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 1 is 2.67379679144385"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 0"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 8.57142857142857"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 5"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 4.51127819548872"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 0"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 12.5"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 2.5"
#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "removing 1 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev   Lambda
#>  [1,]  0 1.762e-15 0.138800
#>  [2,]  1 9.445e-03 0.126500
#>  [3,]  1 1.730e-02 0.115300
#>  [4,]  1 2.384e-02 0.105000
#>  [5,]  1 2.930e-02 0.095690
#>  [6,]  1 3.385e-02 0.087190
#>  [7,]  1 3.765e-02 0.079440
#>  [8,]  1 4.082e-02 0.072380
#>  [9,]  1 4.347e-02 0.065950
#> [10,]  1 4.568e-02 0.060090
#> [11,]  2 4.835e-02 0.054760
#> [12,]  2 5.115e-02 0.049890
#> [13,]  2 5.350e-02 0.045460
#> [14,]  3 5.575e-02 0.041420
#> [15,]  3 5.828e-02 0.037740
#> [16,]  3 6.041e-02 0.034390
#> [17,]  3 6.222e-02 0.031330
#> [18,]  3 6.375e-02 0.028550
#> [19,]  3 6.505e-02 0.026010
#> [20,]  3 6.616e-02 0.023700
#> [21,]  3 6.709e-02 0.021600
#> [22,]  3 6.789e-02 0.019680
#> [23,]  3 6.856e-02 0.017930
#> [24,]  3 6.913e-02 0.016340
#> [25,]  4 6.962e-02 0.014890
#> [26,]  4 7.004e-02 0.013560
#> [27,]  4 7.039e-02 0.012360
#> [28,]  4 7.068e-02 0.011260
#> [29,]  4 7.093e-02 0.010260
#> [30,]  4 7.114e-02 0.009349
#> [31,]  4 7.132e-02 0.008518
#> [32,]  4 7.147e-02 0.007762
#> [33,]  4 7.159e-02 0.007072
#> [34,]  4 7.170e-02 0.006444
#> [35,]  4 7.179e-02 0.005871
#> [36,]  4 7.186e-02 0.005350
#> [37,]  4 7.192e-02 0.004874
#> [38,]  4 7.197e-02 0.004441
#> [39,]  4 7.202e-02 0.004047
#> [40,]  4 7.205e-02 0.003687
#> [41,]  4 7.208e-02 0.003360
#> [42,]  4 7.211e-02 0.003061
#> [43,]  4 7.213e-02 0.002789
#> [44,]  4 7.215e-02 0.002542
#> [45,]  4 7.216e-02 0.002316
#> [46,]  4 7.217e-02 0.002110
#> [47,]  4 7.218e-02 0.001923
#> [48,]  4 7.219e-02 0.001752
#> [1] "lambda min is at location 22"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 216 cells and 5 genes..."
#> [1] "evaluation accuracy ElasticNet 0.643518518518518"
#> [1] "evaluation accuracy LDA 0.648148148148148"
#> [1] "done training for bootstrap 2, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 1 is 11.2299465240642"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 7.48663101604278"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 37.8571428571429"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 13.5714285714286"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 88.7218045112782"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 15.7894736842105"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 70"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 7.5"

summary_deviance(LSOLDA_dat)

#> $allDeviance
#> [1] "0.1423"  "0.06789"
#> 
#> $DeviMax
#>          dat_DE$Dfd          Deviance           DEgenes
#> 1                 0            0.1423    genes_cluster1
#> 2                 1            0.1423    genes_cluster1
#> 3                 2            0.1423    genes_cluster1
#> 4                 3            0.1423    genes_cluster1
#> 5                 4            0.1423    genes_cluster1
#> 6 remaining DEgenes remaining DEgenes remaining DEgenes
#> 
#> $LassoGenesMax
#> NULL
#>