get percent deviance explained for Lasso model, from n bootstraps

the training results from training were written to the object LSOLDA_dat, the summary_prediction summarises prediction for n bootstrap runs

summary_prediction_lasso(LSOLDA_dat = NULL, nPredSubpop = NULL)

Arguments

LSOLDA_dat
is a list containing the training results from training

nPredSubpop
is the number of subpopulations in the target mixed population

Value

a dataframe containg information for the Lasso prediction results, each column contains prediction results for all subpopulations from each bootstrap run

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-new_scGPS_object(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-new_scGPS_object(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap(nboots = 1,mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> Total 224 cells as source subpop

#> Total 366 cells in remaining subpops

#> subsampling 112 cells for training source subpop

#> subsampling 112 cells in remaining subpops for training

#> use 6 genes for training model

#> use 6 genes 224 cells for testing model

#> rename remaining subpops to 2_3

#> there are 112 cells in class 2_3 and 112 cells in class 1

#> removing 1 genes with no variance

#> standardizing prediction/target dataset

#> performning elasticnet model training...

#> performning LDA model training...

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> extracting deviance and best gene features...

#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev   Lambda
#>  [1,]  0 2.243e-15 0.136200
#>  [2,]  1 9.093e-03 0.124100
#>  [3,]  1 1.665e-02 0.113100
#>  [4,]  1 2.295e-02 0.103000
#>  [5,]  1 2.819e-02 0.093870
#>  [6,]  1 3.257e-02 0.085530
#>  [7,]  2 3.838e-02 0.077930
#>  [8,]  2 4.379e-02 0.071010
#>  [9,]  2 4.839e-02 0.064700
#> [10,]  2 5.233e-02 0.058950
#> [11,]  2 5.570e-02 0.053720
#> [12,]  3 5.987e-02 0.048940
#> [13,]  3 6.373e-02 0.044600
#> [14,]  4 6.751e-02 0.040630
#> [15,]  4 7.102e-02 0.037020
#> [16,]  4 7.406e-02 0.033740
#> [17,]  4 7.669e-02 0.030740
#> [18,]  4 7.896e-02 0.028010
#> [19,]  4 8.092e-02 0.025520
#> [20,]  4 8.262e-02 0.023250
#> [21,]  4 8.409e-02 0.021190
#> [22,]  4 8.536e-02 0.019300
#> [23,]  4 8.645e-02 0.017590
#> [24,]  4 8.739e-02 0.016030
#> [25,]  4 8.819e-02 0.014600
#> [26,]  4 8.888e-02 0.013310
#> [27,]  4 8.947e-02 0.012120
#> [28,]  4 8.997e-02 0.011050
#> [29,]  4 9.040e-02 0.010070
#> [30,]  4 9.076e-02 0.009171
#> [31,]  4 9.107e-02 0.008356
#> [32,]  4 9.134e-02 0.007614
#> [33,]  4 9.156e-02 0.006938
#> [34,]  4 9.175e-02 0.006321
#> [35,]  4 9.190e-02 0.005760
#> [36,]  4 9.204e-02 0.005248
#> [37,]  4 9.215e-02 0.004782
#> [38,]  4 9.224e-02 0.004357
#> [39,]  4 9.232e-02 0.003970
#> [40,]  4 9.239e-02 0.003617
#> [41,]  4 9.245e-02 0.003296
#> [42,]  4 9.249e-02 0.003003
#> [43,]  4 9.253e-02 0.002736
#> [44,]  4 9.257e-02 0.002493
#> [45,]  4 9.259e-02 0.002272
#> [46,]  4 9.262e-02 0.002070
#> [47,]  4 9.263e-02 0.001886
#> [48,]  4 9.265e-02 0.001719
#> [49,]  4 9.266e-02 0.001566
#> [50,]  4 9.267e-02 0.001427
#> [51,]  4 9.268e-02 0.001300

#> lambda min is at location 8

#> the leave-out cells in the source subpop is 112

#> use 112 target subpops cells for leave-out test set

#> standardizing the leave-out target and source subpops...

#> start ElasticNet prediction for estimating accuracy...

#> start LDA prediction for estimating accuracy for 216 cells and 5 genes...

#> evaluation accuracy ElasticNet 0.648148148148148

#> evaluation accuracy LDA 0.564814814814815

#> done training for bootstrap 1, moving to prediction...

#> standardizing target subpops before prediction...

#> predicting from source to target subpop 1...

#> number of cells in the target subpop 1 is 187

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 1 is 8.02139037433155

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 1 is 15.5080213903743

#> predicting from source to target subpop 2...

#> number of cells in the target subpop 2 is 140

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 2 is 37.8571428571429

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 2 is 8.57142857142857

#> predicting from source to target subpop 3...

#> number of cells in the target subpop 3 is 133

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 3 is 10.5263157894737

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 3 is 24.0601503759398

#> predicting from source to target subpop 4...

#> number of cells in the target subpop 4 is 40

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 4 is 22.5

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 4 is 12.5

summary_prediction_lasso(LSOLDA_dat=LSOLDA_dat, nPredSubpop=4)

#>                 V1                                      names
#> 1 8.02139037433155 ElasticNet for subpop1 in target mixedpop2
#> 2 37.8571428571429 ElasticNet for subpop2 in target mixedpop2
#> 3 10.5263157894737 ElasticNet for subpop3 in target mixedpop2
#> 4             22.5 ElasticNet for subpop4 in target mixedpop2