Main model training function for finding the best model that characterises a subpopulation

Training a haft of all cells to find optimal ElasticNet and LDA models to predict a subpopulation

training(genes = NULL, cluster_mixedpop1 = NULL, mixedpop1 = NULL,
  mixedpop2 = NULL, c_selectID = NULL, listData = list(),
  out_idx = 1, standardize = TRUE, trainset_ratio = 0.5,
  LDA_run = FALSE)

Arguments

genes
a vector of gene names (for ElasticNet shrinkage); gene symbols must be in the same format with gene names in subpop2. Note that genes are listed by the order of importance, e.g. differentially expressed genes that are most significan, so that if the gene list contains too many genes, only the top 500 genes are used.

cluster_mixedpop1
a vector of cluster assignment in mixedpop1

mixedpop1
is a SingleCellExperiment object from the train mixed population

mixedpop2
is a SingleCellExperiment object from the target mixed population

c_selectID
a selected number to specify which subpopulation to be used for training

listData
list to store output in

out_idx
a number to specify index to write results into the list output. This is needed for running bootstrap.

standardize
a logical value specifying whether or not to standardize the train matrix

trainset_ratio
a number specifying the proportion of cells to be part of the training subpopulation

LDA_run
logical, if the LDA run is added to compare to ElasticNet

Value

a list with prediction results written in to the indexed out_idx

Examples

c_selectID<-1
out_idx<-1
day2 <- sample1
mixedpop1 <-new_scGPS_object(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-new_scGPS_object(ExpressionMatrix = day5$dat5_counts,
GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
listData  <- training(genes,
    cluster_mixedpop1 = colData(mixedpop1)[, 1],
    mixedpop1 = mixedpop1, mixedpop2 = mixedpop2, c_selectID,
    listData =list(), out_idx=out_idx, trainset_ratio = 0.5)

#> Total 224 cells as source subpop

#> Total 366 cells in remaining subpops

#> subsampling 112 cells for training source subpop

#> subsampling 112 cells in remaining subpops for training

#> use 6 genes for training model

#> use 6 genes 224 cells for testing model

#> rename remaining subpops to 2_3

#> there are 112 cells in class 2_3 and 112 cells in class 1

#> standardizing prediction/target dataset

#> performning elasticnet model training...

#> extracting deviance and best gene features...

#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev   Lambda
#>  [1,]  0 2.243e-15 0.112400
#>  [2,]  2 1.102e-02 0.102400
#>  [3,]  2 2.055e-02 0.093340
#>  [4,]  2 2.858e-02 0.085050
#>  [5,]  2 3.537e-02 0.077490
#>  [6,]  2 4.116e-02 0.070610
#>  [7,]  2 4.609e-02 0.064330
#>  [8,]  2 5.030e-02 0.058620
#>  [9,]  2 5.391e-02 0.053410
#> [10,]  2 5.700e-02 0.048670
#> [11,]  2 5.966e-02 0.044340
#> [12,]  4 6.203e-02 0.040400
#> [13,]  4 6.480e-02 0.036810
#> [14,]  4 6.719e-02 0.033540
#> [15,]  4 6.925e-02 0.030560
#> [16,]  4 7.102e-02 0.027850
#> [17,]  4 7.255e-02 0.025370
#> [18,]  4 7.386e-02 0.023120
#> [19,]  4 7.499e-02 0.021070
#> [20,]  5 7.602e-02 0.019200
#> [21,]  5 7.693e-02 0.017490
#> [22,]  5 7.770e-02 0.015940
#> [23,]  5 7.837e-02 0.014520
#> [24,]  5 7.895e-02 0.013230
#> [25,]  5 7.944e-02 0.012060
#> [26,]  5 7.986e-02 0.010980
#> [27,]  5 8.022e-02 0.010010
#> [28,]  5 8.053e-02 0.009119
#> [29,]  5 8.079e-02 0.008309
#> [30,]  5 8.102e-02 0.007571
#> [31,]  5 8.121e-02 0.006898
#> [32,]  5 8.137e-02 0.006286
#> [33,]  5 8.151e-02 0.005727
#> [34,]  5 8.162e-02 0.005218
#> [35,]  5 8.172e-02 0.004755
#> [36,]  5 8.181e-02 0.004332
#> [37,]  5 8.188e-02 0.003947
#> [38,]  5 8.193e-02 0.003597
#> [39,]  5 8.198e-02 0.003277
#> [40,]  5 8.203e-02 0.002986
#> [41,]  5 8.206e-02 0.002721
#> [42,]  5 8.209e-02 0.002479
#> [43,]  5 8.212e-02 0.002259
#> [44,]  5 8.214e-02 0.002058
#> [45,]  5 8.215e-02 0.001875
#> [46,]  5 8.217e-02 0.001709
#> [47,]  5 8.218e-02 0.001557
#> [48,]  5 8.219e-02 0.001419
#> [49,]  5 8.220e-02 0.001293

#> lambda min is at location 10

#> the leave-out cells in the source subpop is 112

#> use 112 target subpops cells for leave-out test set

#> standardizing the leave-out target and source subpops...

#> start ElasticNet prediction for estimating accuracy...

#> evaluation accuracy ElasticNet 0.624413145539906

names(listData)

#> [1] "Accuracy"        "ElasticNetGenes" "Deviance"        "ElasticNetFit"  
#> [5] "LDAFit"          "predictor_S1"   

listData$Accuracy

#> [[1]]
#> [[1]][[1]]
#> [[1]][[1]][[1]]
#> [1] 133
#> 
#> [[1]][[1]][[2]]
#> [1] 80
#> 
#> 
#>