BootStrap runs for both scGPS training and prediction

ElasticNet and LDA prediction for each of all the subpopulations in the new mixed population after training the model for a subpopulation in the first mixed population. The number of bootstraps to be run can be specified.

bootstrap_scGPS(nboots = 1, genes = genes, mixedpop1 = mixedpop1,
  mixedpop2 = mixedpop2, c_selectID = NULL, listData = list(),
  cluster_mixedpop1 = NULL, cluster_mixedpop2 = NULL,
  trainset_ratio = 0.5, LDA_run = TRUE)

Arguments

nboots
a number specifying how many bootstraps to be run

genes
a gene list to build the model

mixedpop1
a SingleCellExperiment object from a mixed population for training

mixedpop2
a SingleCellExperiment object from a target mixed population for prediction

c_selectID
the root cluster in mixedpop1 to becompared to clusters in mixedpop2

listData
a list object, which contains trained results for the first mixed population

cluster_mixedpop1
a vector of cluster assignment for mixedpop1

cluster_mixedpop2
a vector of cluster assignment for mixedpop2

trainset_ratio
a number specifying the proportion of cells to be part of the training subpopulation

LDA_run
logical, if the LDA prediction is added to compare to ElasticNet

Value

a list with prediction results written in to the index out_idx

See also

Examples

day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
cluster_mixedpop1 <- colData(mixedpop1)[,1]
cluster_mixedpop2 <- colData(mixedpop2)[,1]
c_selectID <- 2
test <- bootstrap_scGPS(nboots = 2, mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, listData =list(),
    cluster_mixedpop1 = cluster_mixedpop1,
    cluster_mixedpop2 = cluster_mixedpop2, c_selectID = c_selectID)

#> [1] "Total 203 cells as source subpop"
#> [1] "Total 387 cells in remaining subpops"
#> [1] "subsampling 102 cells for training source subpop"
#> [1] "subsampling 102 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 204 cells for testing model"
#> [1] "rename remaining subpops to 1_3"
#> [1] "there are 102 cells in class 1_3 and 102 cells in class 2"
#> [1] "removing 1 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev   Lambda
#>  [1,]  0 8.010e-16 0.153100
#>  [2,]  1 1.149e-02 0.139500
#>  [3,]  1 2.104e-02 0.127100
#>  [4,]  1 2.900e-02 0.115800
#>  [5,]  1 3.563e-02 0.105500
#>  [6,]  1 4.116e-02 0.096140
#>  [7,]  1 4.578e-02 0.087600
#>  [8,]  2 5.214e-02 0.079820
#>  [9,]  2 5.759e-02 0.072730
#> [10,]  2 6.223e-02 0.066270
#> [11,]  2 6.621e-02 0.060380
#> [12,]  2 6.961e-02 0.055010
#> [13,]  3 7.284e-02 0.050130
#> [14,]  3 7.563e-02 0.045670
#> [15,]  3 7.802e-02 0.041620
#> [16,]  3 8.007e-02 0.037920
#> [17,]  3 8.182e-02 0.034550
#> [18,]  3 8.332e-02 0.031480
#> [19,]  3 8.460e-02 0.028680
#> [20,]  3 8.570e-02 0.026140
#> [21,]  3 8.663e-02 0.023810
#> [22,]  3 8.742e-02 0.021700
#> [23,]  3 8.809e-02 0.019770
#> [24,]  3 8.866e-02 0.018010
#> [25,]  3 8.914e-02 0.016410
#> [26,]  3 8.955e-02 0.014960
#> [27,]  3 8.989e-02 0.013630
#> [28,]  3 9.018e-02 0.012420
#> [29,]  4 9.044e-02 0.011310
#> [30,]  4 9.073e-02 0.010310
#> [31,]  4 9.100e-02 0.009393
#> [32,]  4 9.122e-02 0.008559
#> [33,]  4 9.141e-02 0.007798
#> [34,]  4 9.157e-02 0.007105
#> [35,]  4 9.170e-02 0.006474
#> [36,]  4 9.181e-02 0.005899
#> [37,]  4 9.190e-02 0.005375
#> [38,]  4 9.198e-02 0.004898
#> [39,]  4 9.204e-02 0.004462
#> [40,]  4 9.210e-02 0.004066
#> [41,]  4 9.214e-02 0.003705
#> [42,]  4 9.218e-02 0.003376
#> [43,]  4 9.221e-02 0.003076
#> [44,]  4 9.223e-02 0.002803
#> [45,]  4 9.226e-02 0.002554
#> [46,]  4 9.227e-02 0.002327
#> [47,]  4 9.229e-02 0.002120
#> [48,]  4 9.230e-02 0.001932
#> [49,]  4 9.231e-02 0.001760
#> [50,]  4 9.232e-02 0.001604
#> [1] "lambda min is at location 14"
#> [1] "the leave-out cells in the source subpop is 101"
#> [1] "use 102 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 195 cells and 5 genes..."
#> [1] "evaluation accuracy ElasticNet 0.615384615384615"
#> [1] "evaluation accuracy LDA 0.625641025641026"
#> [1] "done training for bootstrap 1, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 1 is 14.9732620320856"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 63.1016042780749"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 15"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 8.57142857142857"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 82.7067669172932"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 51.1278195488722"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 77.5"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 40"
#> [1] "Total 203 cells as source subpop"
#> [1] "Total 387 cells in remaining subpops"
#> [1] "subsampling 102 cells for training source subpop"
#> [1] "subsampling 102 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 204 cells for testing model"
#> [1] "rename remaining subpops to 1_3"
#> [1] "there are 102 cells in class 1_3 and 102 cells in class 2"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev   Lambda
#>  [1,]  0 6.410e-16 0.122900
#>  [2,]  2 8.098e-03 0.112000
#>  [3,]  2 1.482e-02 0.102100
#>  [4,]  2 2.041e-02 0.092990
#>  [5,]  3 2.615e-02 0.084730
#>  [6,]  3 3.543e-02 0.077200
#>  [7,]  3 4.330e-02 0.070350
#>  [8,]  3 4.994e-02 0.064100
#>  [9,]  3 5.555e-02 0.058400
#> [10,]  4 6.070e-02 0.053210
#> [11,]  4 6.527e-02 0.048490
#> [12,]  4 6.919e-02 0.044180
#> [13,]  4 7.253e-02 0.040250
#> [14,]  3 7.541e-02 0.036680
#> [15,]  3 7.786e-02 0.033420
#> [16,]  3 7.996e-02 0.030450
#> [17,]  3 8.176e-02 0.027750
#> [18,]  3 8.331e-02 0.025280
#> [19,]  3 8.463e-02 0.023040
#> [20,]  3 8.576e-02 0.020990
#> [21,]  3 8.673e-02 0.019120
#> [22,]  3 8.756e-02 0.017430
#> [23,]  3 8.827e-02 0.015880
#> [24,]  3 8.887e-02 0.014470
#> [25,]  3 8.938e-02 0.013180
#> [26,]  3 8.981e-02 0.012010
#> [27,]  3 9.018e-02 0.010940
#> [28,]  5 9.057e-02 0.009971
#> [29,]  5 9.101e-02 0.009086
#> [30,]  5 9.138e-02 0.008278
#> [31,]  5 9.169e-02 0.007543
#> [32,]  5 9.196e-02 0.006873
#> [33,]  5 9.219e-02 0.006262
#> [34,]  5 9.238e-02 0.005706
#> [35,]  5 9.254e-02 0.005199
#> [36,]  5 9.267e-02 0.004737
#> [37,]  5 9.279e-02 0.004316
#> [38,]  5 9.288e-02 0.003933
#> [39,]  5 9.296e-02 0.003584
#> [40,]  5 9.303e-02 0.003265
#> [41,]  5 9.309e-02 0.002975
#> [42,]  5 9.314e-02 0.002711
#> [43,]  5 9.318e-02 0.002470
#> [44,]  5 9.321e-02 0.002251
#> [45,]  5 9.324e-02 0.002051
#> [46,]  5 9.326e-02 0.001868
#> [47,]  5 9.328e-02 0.001702
#> [48,]  5 9.330e-02 0.001551
#> [49,]  5 9.331e-02 0.001413
#> [50,]  5 9.332e-02 0.001288
#> [51,]  5 9.333e-02 0.001173
#> [1] "lambda min is at location 16"
#> [1] "the leave-out cells in the source subpop is 101"
#> [1] "use 102 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 195 cells and 6 genes..."
#> [1] "evaluation accuracy ElasticNet 0.687179487179487"
#> [1] "evaluation accuracy LDA 0.682051282051282"
#> [1] "done training for bootstrap 2, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 1 is 44.3850267379679"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 62.5668449197861"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 47.1428571428571"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 22.1428571428571"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 95.4887218045113"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 48.8721804511278"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 85"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 45"

names(test)

#> [1] "Accuracy"          "ElasticNetGenes"   "Deviance"         
#> [4] "ElasticNetFit"     "LDAFit"            "predictor_S1"     
#> [7] "ElasticNetPredict" "LDAPredict"       

test$ElasticNetPredict

#> [[1]]
#> [[1]][[1]]
#> [1] "ElasticNet for subpop1 in target mixedpop2"
#> 
#> [[1]][[2]]
#> [1] 14.97326
#> 
#> [[1]][[3]]
#> [1] "ElasticNet for subpop2 in target mixedpop2"
#> 
#> [[1]][[4]]
#> [1] 15
#> 
#> [[1]][[5]]
#> [1] "ElasticNet for subpop3 in target mixedpop2"
#> 
#> [[1]][[6]]
#> [1] 82.70677
#> 
#> [[1]][[7]]
#> [1] "ElasticNet for subpop4 in target mixedpop2"
#> 
#> [[1]][[8]]
#> [1] 77.5
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] "ElasticNet for subpop1 in target mixedpop2"
#> 
#> [[2]][[2]]
#> [1] 44.38503
#> 
#> [[2]][[3]]
#> [1] "ElasticNet for subpop2 in target mixedpop2"
#> 
#> [[2]][[4]]
#> [1] 47.14286
#> 
#> [[2]][[5]]
#> [1] "ElasticNet for subpop3 in target mixedpop2"
#> 
#> [[2]][[6]]
#> [1] 95.48872
#> 
#> [[2]][[7]]
#> [1] "ElasticNet for subpop4 in target mixedpop2"
#> 
#> [[2]][[8]]
#> [1] 85
#> 
#> 

test$LDAPredict

#> [[1]]
#> [[1]][[1]]
#> [1] "LDA for subpop 1 in target mixedpop2"
#> 
#> [[1]][[2]]
#> [1] 63.1016
#> 
#> [[1]][[3]]
#> [1] "LDA for subpop 2 in target mixedpop2"
#> 
#> [[1]][[4]]
#> [1] 8.571429
#> 
#> [[1]][[5]]
#> [1] "LDA for subpop 3 in target mixedpop2"
#> 
#> [[1]][[6]]
#> [1] 51.12782
#> 
#> [[1]][[7]]
#> [1] "LDA for subpop 4 in target mixedpop2"
#> 
#> [[1]][[8]]
#> [1] 40
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] "LDA for subpop 1 in target mixedpop2"
#> 
#> [[2]][[2]]
#> [1] 62.56684
#> 
#> [[2]][[3]]
#> [1] "LDA for subpop 2 in target mixedpop2"
#> 
#> [[2]][[4]]
#> [1] 22.14286
#> 
#> [[2]][[5]]
#> [1] "LDA for subpop 3 in target mixedpop2"
#> 
#> [[2]][[6]]
#> [1] 48.87218
#> 
#> [[2]][[7]]
#> [1] "LDA for subpop 4 in target mixedpop2"
#> 
#> [[2]][[8]]
#> [1] 45
#> 
#>