BootStrap runs for both scGPS training and prediction

ElasticNet and LDA prediction for each of all the subpopulations in the new mixed population after training the model for a subpopulation in the first mixed population. The number of bootstraps to be run can be specified.

bootstrap_scGPS(nboots = 1, genes = genes, mixedpop1 = mixedpop1,
  mixedpop2 = mixedpop2, c_selectID = NULL, listData = list(),
  cluster_mixedpop1 = NULL, cluster_mixedpop2 = NULL,
  trainset_ratio = 0.5, LDA_run = TRUE)

Arguments

nboots
a number specifying how many bootstraps to be run

genes
a gene list to build the model

mixedpop1
a SingleCellExperiment object from a mixed population for training

mixedpop2
a SingleCellExperiment object from a target mixed population for prediction

c_selectID
the root cluster in mixedpop1 to becompared to clusters in mixedpop2

listData
a list object, which contains trained results for the first mixed population

cluster_mixedpop1
a vector of cluster assignment for mixedpop1

cluster_mixedpop2
a vector of cluster assignment for mixedpop2

trainset_ratio
a number specifying the proportion of cells to be part of the training subpopulation

LDA_run
logical, if the LDA prediction is added to compare to ElasticNet

Value

a list with prediction results written in to the index out_idx

See also

Examples

day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
cluster_mixedpop1 <- colData(mixedpop1)[,1]
cluster_mixedpop2 <- colData(mixedpop2)[,1]
c_selectID <- 2
test <- bootstrap_scGPS(nboots = 2, mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, listData =list(),
    cluster_mixedpop1 = cluster_mixedpop1,
    cluster_mixedpop2 = cluster_mixedpop2, c_selectID = c_selectID)

#> [1] "Total 203 cells as source subpop"
#> [1] "Total 387 cells in remaining subpops"
#> [1] "subsampling 102 cells for training source subpop"
#> [1] "subsampling 102 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 204 cells for testing model"
#> [1] "rename remaining subpops to 1_3"
#> [1] "there are 102 cells in class 1_3 and 102 cells in class 2"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df       %Dev   Lambda
#>  [1,]  0 -8.009e-16 0.128500
#>  [2,]  1  8.096e-03 0.117100
#>  [3,]  1  1.483e-02 0.106700
#>  [4,]  2  2.052e-02 0.097230
#>  [5,]  2  2.567e-02 0.088590
#>  [6,]  2  2.996e-02 0.080720
#>  [7,]  2  3.354e-02 0.073550
#>  [8,]  2  3.653e-02 0.067020
#>  [9,]  2  3.901e-02 0.061060
#> [10,]  2  4.109e-02 0.055640
#> [11,]  2  4.282e-02 0.050690
#> [12,]  3  4.485e-02 0.046190
#> [13,]  3  4.708e-02 0.042090
#> [14,]  3  4.894e-02 0.038350
#> [15,]  3  5.050e-02 0.034940
#> [16,]  3  5.180e-02 0.031840
#> [17,]  3  5.289e-02 0.029010
#> [18,]  3  5.380e-02 0.026430
#> [19,]  3  5.456e-02 0.024080
#> [20,]  3  5.520e-02 0.021940
#> [21,]  3  5.573e-02 0.019990
#> [22,]  3  5.617e-02 0.018220
#> [23,]  3  5.654e-02 0.016600
#> [24,]  3  5.685e-02 0.015130
#> [25,]  4  5.712e-02 0.013780
#> [26,]  4  5.745e-02 0.012560
#> [27,]  4  5.774e-02 0.011440
#> [28,]  5  5.799e-02 0.010430
#> [29,]  5  5.820e-02 0.009499
#> [30,]  5  5.838e-02 0.008655
#> [31,]  5  5.853e-02 0.007886
#> [32,]  5  5.865e-02 0.007186
#> [33,]  5  5.876e-02 0.006547
#> [34,]  5  5.884e-02 0.005966
#> [35,]  5  5.891e-02 0.005436
#> [36,]  5  5.897e-02 0.004953
#> [37,]  5  5.902e-02 0.004513
#> [38,]  5  5.906e-02 0.004112
#> [39,]  5  5.910e-02 0.003747
#> [40,]  5  5.913e-02 0.003414
#> [41,]  5  5.915e-02 0.003111
#> [42,]  5  5.917e-02 0.002834
#> [43,]  5  5.918e-02 0.002582
#> [44,]  5  5.920e-02 0.002353
#> [45,]  5  5.921e-02 0.002144
#> [46,]  5  5.922e-02 0.001954
#> [1] "lambda min is at location 30"
#> [1] "the leave-out cells in the source subpop is 101"
#> [1] "use 102 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 196 cells and 6 genes..."
#> [1] "evaluation accuracy ElasticNet 0.668367346938776"
#> [1] "evaluation accuracy LDA 0.668367346938776"
#> [1] "done training for bootstrap 1, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "ElasticNet for target subpop 1 has no solution"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 100"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 7.85714285714286"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 12.1428571428571"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 96.2406015037594"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 99.2481203007519"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 70"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 75"
#> [1] "Total 203 cells as source subpop"
#> [1] "Total 387 cells in remaining subpops"
#> [1] "subsampling 102 cells for training source subpop"
#> [1] "subsampling 102 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 204 cells for testing model"
#> [1] "rename remaining subpops to 1_3"
#> [1] "there are 102 cells in class 1_3 and 102 cells in class 2"
#> [1] "removing 1 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev   Lambda
#>  [1,]  0 1.602e-15 0.126800
#>  [2,]  1 7.885e-03 0.115600
#>  [3,]  1 1.444e-02 0.105300
#>  [4,]  1 1.990e-02 0.095950
#>  [5,]  1 2.444e-02 0.087430
#>  [6,]  1 2.823e-02 0.079660
#>  [7,]  1 3.139e-02 0.072590
#>  [8,]  1 3.402e-02 0.066140
#>  [9,]  1 3.622e-02 0.060260
#> [10,]  2 3.980e-02 0.054910
#> [11,]  2 4.307e-02 0.050030
#> [12,]  3 4.726e-02 0.045590
#> [13,]  4 5.127e-02 0.041540
#> [14,]  4 5.467e-02 0.037850
#> [15,]  4 5.756e-02 0.034480
#> [16,]  4 6.002e-02 0.031420
#> [17,]  4 6.211e-02 0.028630
#> [18,]  4 6.390e-02 0.026090
#> [19,]  4 6.542e-02 0.023770
#> [20,]  4 6.671e-02 0.021660
#> [21,]  4 6.781e-02 0.019730
#> [22,]  4 6.875e-02 0.017980
#> [23,]  4 6.954e-02 0.016380
#> [24,]  4 7.022e-02 0.014930
#> [25,]  4 7.079e-02 0.013600
#> [26,]  4 7.127e-02 0.012390
#> [27,]  4 7.168e-02 0.011290
#> [28,]  4 7.203e-02 0.010290
#> [29,]  4 7.232e-02 0.009375
#> [30,]  4 7.256e-02 0.008542
#> [31,]  4 7.277e-02 0.007783
#> [32,]  4 7.294e-02 0.007092
#> [33,]  4 7.309e-02 0.006462
#> [34,]  4 7.321e-02 0.005888
#> [35,]  4 7.331e-02 0.005365
#> [36,]  4 7.340e-02 0.004888
#> [37,]  4 7.347e-02 0.004454
#> [38,]  4 7.353e-02 0.004058
#> [39,]  4 7.358e-02 0.003698
#> [40,]  4 7.362e-02 0.003369
#> [41,]  4 7.366e-02 0.003070
#> [42,]  4 7.369e-02 0.002797
#> [43,]  4 7.371e-02 0.002549
#> [44,]  4 7.373e-02 0.002322
#> [45,]  4 7.375e-02 0.002116
#> [46,]  4 7.376e-02 0.001928
#> [47,]  4 7.377e-02 0.001757
#> [48,]  4 7.378e-02 0.001601
#> [1] "lambda min is at location 31"
#> [1] "the leave-out cells in the source subpop is 101"
#> [1] "use 102 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 196 cells and 5 genes..."
#> [1] "evaluation accuracy ElasticNet 0.586734693877551"
#> [1] "evaluation accuracy LDA 0.581632653061224"
#> [1] "done training for bootstrap 2, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 1 is 33.6898395721925"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 29.9465240641711"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 5.71428571428571"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 11.4285714285714"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 9.77443609022556"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 18.0451127819549"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 12.5"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 25"

names(test)

#> [1] "Accuracy"          "ElasticNetGenes"   "Deviance"         
#> [4] "ElasticNetFit"     "LDAFit"            "predictor_S1"     
#> [7] "ElasticNetPredict" "LDAPredict"       

test$ElasticNetPredict

#> [[1]]
#> [[1]][[1]]
#> [1] "ElasticNet for subpop1 in target mixedpop2"
#> 
#> [[1]][[2]]
#> [1] "NA"
#> 
#> [[1]][[3]]
#> [1] "ElasticNet for subpop2 in target mixedpop2"
#> 
#> [[1]][[4]]
#> [1] 7.857143
#> 
#> [[1]][[5]]
#> [1] "ElasticNet for subpop3 in target mixedpop2"
#> 
#> [[1]][[6]]
#> [1] 96.2406
#> 
#> [[1]][[7]]
#> [1] "ElasticNet for subpop4 in target mixedpop2"
#> 
#> [[1]][[8]]
#> [1] 70
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] "ElasticNet for subpop1 in target mixedpop2"
#> 
#> [[2]][[2]]
#> [1] 33.68984
#> 
#> [[2]][[3]]
#> [1] "ElasticNet for subpop2 in target mixedpop2"
#> 
#> [[2]][[4]]
#> [1] 5.714286
#> 
#> [[2]][[5]]
#> [1] "ElasticNet for subpop3 in target mixedpop2"
#> 
#> [[2]][[6]]
#> [1] 9.774436
#> 
#> [[2]][[7]]
#> [1] "ElasticNet for subpop4 in target mixedpop2"
#> 
#> [[2]][[8]]
#> [1] 12.5
#> 
#> 

test$LDAPredict

#> [[1]]
#> [[1]][[1]]
#> [1] "LDA for subpop 1 in target mixedpop2"
#> 
#> [[1]][[2]]
#> [1] 100
#> 
#> [[1]][[3]]
#> [1] "LDA for subpop 2 in target mixedpop2"
#> 
#> [[1]][[4]]
#> [1] 12.14286
#> 
#> [[1]][[5]]
#> [1] "LDA for subpop 3 in target mixedpop2"
#> 
#> [[1]][[6]]
#> [1] 99.24812
#> 
#> [[1]][[7]]
#> [1] "LDA for subpop 4 in target mixedpop2"
#> 
#> [[1]][[8]]
#> [1] 75
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> [1] "LDA for subpop 1 in target mixedpop2"
#> 
#> [[2]][[2]]
#> [1] 29.94652
#> 
#> [[2]][[3]]
#> [1] "LDA for subpop 2 in target mixedpop2"
#> 
#> [[2]][[4]]
#> [1] 11.42857
#> 
#> [[2]][[5]]
#> [1] "LDA for subpop 3 in target mixedpop2"
#> 
#> [[2]][[6]]
#> [1] 18.04511
#> 
#> [[2]][[7]]
#> [1] "LDA for subpop 4 in target mixedpop2"
#> 
#> [[2]][[8]]
#> [1] 25
#> 
#>