Main prediction function applying the optimal ElasticNet and LDA models

Predict a new mixed population after training the model for a subpopulation in the first mixed population. All subpopulations in the new target mixed population will be predicted, where each targeted subpopulation will have a transition score from the orginal subpopulation to the new subpopulation.

predicting(listData = NULL, cluster_mixedpop2 = NULL,
  mixedpop2 = NULL, out_idx = NULL, standardize = TRUE,
  LDA_run = FALSE, c_selectID = NULL)

Arguments

listData
a list object containing trained results for the selected subpopulation in the first mixed population

cluster_mixedpop2
a vector of cluster assignment for mixedpop2

mixedpop2
a SingleCellExperiment object from the target mixed population of importance, e.g. differentially expressed genes that are most significant

out_idx
a number to specify index to write results into the list output. This is needed for running bootstrap.

standardize
a logical of whether to standardize the data

LDA_run
logical, if the LDA prediction is added to compare to ElasticNet, the LDA model needs to be trained from the training before inputting to this prediction step

c_selectID
a number to specify the trained cluster used for prediction

Value

a list with prediction results written in to the index out_idx

Examples

c_selectID<-1
out_idx<-1
day2 <- sample1
mixedpop1 <-new_scGPS_object(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-new_scGPS_object(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
listData  <- training(genes,
    cluster_mixedpop1 = colData(mixedpop1)[, 1], mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, c_selectID, listData =list(), out_idx=out_idx)

#> Total 224 cells as source subpop

#> Total 366 cells in remaining subpops

#> subsampling 112 cells for training source subpop

#> subsampling 112 cells in remaining subpops for training

#> use 6 genes for training model

#> use 6 genes 224 cells for testing model

#> rename remaining subpops to 2_3

#> there are 112 cells in class 2_3 and 112 cells in class 1

#> removing 1 genes with no variance

#> standardizing prediction/target dataset

#> performning elasticnet model training...

#> extracting deviance and best gene features...

#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev   Lambda
#>  [1,]  0 1.923e-15 0.144500
#>  [2,]  1 1.024e-02 0.131700
#>  [3,]  1 1.875e-02 0.120000
#>  [4,]  1 2.584e-02 0.109300
#>  [5,]  1 3.175e-02 0.099590
#>  [6,]  2 3.734e-02 0.090740
#>  [7,]  2 4.505e-02 0.082680
#>  [8,]  2 5.160e-02 0.075340
#>  [9,]  2 5.722e-02 0.068640
#> [10,]  2 6.207e-02 0.062540
#> [11,]  2 6.626e-02 0.056990
#> [12,]  2 6.990e-02 0.051930
#> [13,]  2 7.306e-02 0.047310
#> [14,]  3 7.650e-02 0.043110
#> [15,]  3 7.959e-02 0.039280
#> [16,]  3 8.227e-02 0.035790
#> [17,]  3 8.457e-02 0.032610
#> [18,]  3 8.655e-02 0.029710
#> [19,]  3 8.826e-02 0.027070
#> [20,]  3 8.973e-02 0.024670
#> [21,]  3 9.100e-02 0.022480
#> [22,]  3 9.208e-02 0.020480
#> [23,]  3 9.300e-02 0.018660
#> [24,]  3 9.378e-02 0.017000
#> [25,]  3 9.445e-02 0.015490
#> [26,]  3 9.502e-02 0.014120
#> [27,]  3 9.550e-02 0.012860
#> [28,]  3 9.591e-02 0.011720
#> [29,]  3 9.625e-02 0.010680
#> [30,]  3 9.654e-02 0.009730
#> [31,]  3 9.679e-02 0.008866
#> [32,]  3 9.699e-02 0.008078
#> [33,]  3 9.717e-02 0.007360
#> [34,]  3 9.731e-02 0.006706
#> [35,]  4 9.746e-02 0.006111
#> [36,]  4 9.759e-02 0.005568
#> [37,]  4 9.770e-02 0.005073
#> [38,]  4 9.780e-02 0.004623
#> [39,]  4 9.787e-02 0.004212
#> [40,]  4 9.794e-02 0.003838
#> [41,]  4 9.799e-02 0.003497
#> [42,]  4 9.804e-02 0.003186
#> [43,]  4 9.808e-02 0.002903
#> [44,]  4 9.811e-02 0.002645
#> [45,]  4 9.814e-02 0.002410
#> [46,]  4 9.816e-02 0.002196
#> [47,]  4 9.818e-02 0.002001
#> [48,]  4 9.819e-02 0.001823
#> [49,]  4 9.821e-02 0.001661
#> [50,]  4 9.822e-02 0.001514
#> [51,]  4 9.823e-02 0.001379

#> lambda min is at location 30

#> the leave-out cells in the source subpop is 112

#> use 112 target subpops cells for leave-out test set

#> standardizing the leave-out target and source subpops...

#> start ElasticNet prediction for estimating accuracy...

#> evaluation accuracy ElasticNet 0.652777777777778

listData  <- predicting(listData =listData,  mixedpop2 = mixedpop2,
    out_idx=out_idx, cluster_mixedpop2 = colData(mixedpop2)[, 1],
    c_selectID = c_selectID)

#> standardizing target subpops before prediction...

#> predicting from source to target subpop 1...

#> number of cells in the target subpop 1 is 187

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 1 is 1.60427807486631

#> predicting from source to target subpop 2...

#> number of cells in the target subpop 2 is 140

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 2 is 36.4285714285714

#> predicting from source to target subpop 3...

#> number of cells in the target subpop 3 is 133

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 3 is 6.76691729323308

#> predicting from source to target subpop 4...

#> number of cells in the target subpop 4 is 40

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 4 is 20