Main prediction function applying the optimal ElasticNet and LDA models

Predict a new mixed population after training the model for a subpopulation in the first mixed population. All subpopulations in the new target mixed population will be predicted, where each targeted subpopulation will have a transition score from the orginal subpopulation to the new subpopulation.

predicting_scGPS(listData = NULL, cluster_mixedpop2 = NULL,
  mixedpop2 = NULL, out_idx = NULL, standardize = TRUE,
  LDA_run = FALSE, c_selectID = NULL)

Arguments

listData
a list object containing trained results for the selected subpopulation in the first mixed population

cluster_mixedpop2
a vector of cluster assignment for mixedpop2

mixedpop2
a SingleCellExperiment object from the target mixed population of importance, e.g. differentially expressed genes that are most significant

out_idx
a number to specify index to write results into the list output. This is needed for running bootstrap.

standardize
a logical of whether to standardize the data

LDA_run
logical, if the LDA prediction is added to compare to ElasticNet, the LDA model needs to be trained from the training_scGPS before inputting to this prediction step

c_selectID
a number to specify the trained cluster used for prediction

Value

a list with prediction results written in to the index out_idx

Examples

c_selectID<-1
out_idx<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
listData  <- training_scGPS(genes,
    cluster_mixedpop1 = colData(mixedpop1)[, 1], mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, c_selectID, listData =list(), out_idx=out_idx)

#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "removing 1 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev   Lambda
#>  [1,]  0 4.487e-15 0.133400
#>  [2,]  1 8.726e-03 0.121500
#>  [3,]  1 1.598e-02 0.110700
#>  [4,]  1 2.202e-02 0.100900
#>  [5,]  2 2.995e-02 0.091940
#>  [6,]  2 3.778e-02 0.083780
#>  [7,]  2 4.441e-02 0.076330
#>  [8,]  2 5.005e-02 0.069550
#>  [9,]  2 5.487e-02 0.063370
#> [10,]  2 5.901e-02 0.057740
#> [11,]  3 6.308e-02 0.052610
#> [12,]  3 6.690e-02 0.047940
#> [13,]  3 7.019e-02 0.043680
#> [14,]  3 7.303e-02 0.039800
#> [15,]  3 7.548e-02 0.036260
#> [16,]  4 7.786e-02 0.033040
#> [17,]  4 8.006e-02 0.030110
#> [18,]  4 8.196e-02 0.027430
#> [19,]  4 8.362e-02 0.025000
#> [20,]  4 8.505e-02 0.022780
#> [21,]  4 8.630e-02 0.020750
#> [22,]  4 8.737e-02 0.018910
#> [23,]  4 8.830e-02 0.017230
#> [24,]  4 8.911e-02 0.015700
#> [25,]  4 8.979e-02 0.014300
#> [26,]  4 9.038e-02 0.013030
#> [27,]  4 9.089e-02 0.011880
#> [28,]  4 9.132e-02 0.010820
#> [29,]  4 9.169e-02 0.009859
#> [30,]  4 9.201e-02 0.008983
#> [31,]  4 9.227e-02 0.008185
#> [32,]  4 9.250e-02 0.007458
#> [33,]  4 9.269e-02 0.006795
#> [34,]  4 9.285e-02 0.006192
#> [35,]  4 9.299e-02 0.005642
#> [36,]  4 9.311e-02 0.005140
#> [37,]  4 9.320e-02 0.004684
#> [38,]  4 9.328e-02 0.004268
#> [39,]  4 9.335e-02 0.003889
#> [40,]  4 9.341e-02 0.003543
#> [41,]  4 9.346e-02 0.003228
#> [42,]  4 9.350e-02 0.002942
#> [43,]  4 9.353e-02 0.002680
#> [44,]  4 9.356e-02 0.002442
#> [45,]  4 9.359e-02 0.002225
#> [46,]  4 9.361e-02 0.002027
#> [47,]  4 9.362e-02 0.001847
#> [48,]  4 9.364e-02 0.001683
#> [49,]  4 9.365e-02 0.001534
#> [50,]  4 9.366e-02 0.001397
#> [1] "lambda min is at location 12"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "evaluation accuracy ElasticNet 0.648148148148148"

listData  <- predicting_scGPS(listData =listData,  mixedpop2 = mixedpop2,
    out_idx=out_idx, cluster_mixedpop2 = colData(mixedpop2)[, 1],
    c_selectID = c_selectID)

#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 1 is 3.20855614973262"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 25.7142857142857"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 7.5187969924812"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 17.5"