Main prediction function applying the optimal ElasticNet and LDA models

Predict a new mixed population after training the model for a subpopulation in the first mixed population. All subpopulations in the new target mixed population will be predicted, where each targeted subpopulation will have a transition score from the orginal subpopulation to the new subpopulation.

predicting_scGPS(listData = NULL, cluster_mixedpop2 = NULL,
  mixedpop2 = NULL, out_idx = NULL, standardize = TRUE,
  LDA_run = FALSE, c_selectID = NULL)

Arguments

listData
a list object containing trained results for the selected subpopulation in the first mixed population

cluster_mixedpop2
a vector of cluster assignment for mixedpop2

mixedpop2
a SingleCellExperiment object from the target mixed population of importance, e.g. differentially expressed genes that are most significant

out_idx
a number to specify index to write results into the list output. This is needed for running bootstrap.

standardize
a logical of whether to standardize the data

LDA_run
logical, if the LDA prediction is added to compare to ElasticNet, the LDA model needs to be trained from the training_scGPS before inputting to this prediction step

c_selectID
a number to specify the trained cluster used for prediction

Value

a list with prediction results written in to the index out_idx

Examples

c_selectID<-1
out_idx<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
listData  <- training_scGPS(genes,
    cluster_mixedpop1 = colData(mixedpop1)[, 1], mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, c_selectID, listData =list(), out_idx=out_idx)

#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df       %Dev   Lambda
#>  [1,]  0 -1.922e-15 0.126100
#>  [2,]  1  7.789e-03 0.114900
#>  [3,]  1  1.426e-02 0.104700
#>  [4,]  1  1.965e-02 0.095360
#>  [5,]  2  2.748e-02 0.086890
#>  [6,]  2  3.446e-02 0.079170
#>  [7,]  2  4.036e-02 0.072140
#>  [8,]  2  4.539e-02 0.065730
#>  [9,]  2  4.967e-02 0.059890
#> [10,]  2  5.334e-02 0.054570
#> [11,]  3  5.687e-02 0.049720
#> [12,]  4  6.074e-02 0.045300
#> [13,]  4  6.432e-02 0.041280
#> [14,]  4  6.742e-02 0.037610
#> [15,]  4  7.012e-02 0.034270
#> [16,]  4  7.248e-02 0.031230
#> [17,]  4  7.453e-02 0.028450
#> [18,]  4  7.632e-02 0.025920
#> [19,]  4  7.789e-02 0.023620
#> [20,]  4  7.925e-02 0.021520
#> [21,]  5  8.045e-02 0.019610
#> [22,]  5  8.152e-02 0.017870
#> [23,]  5  8.245e-02 0.016280
#> [24,]  5  8.325e-02 0.014840
#> [25,]  5  8.395e-02 0.013520
#> [26,]  5  8.456e-02 0.012320
#> [27,]  5  8.508e-02 0.011220
#> [28,]  5  8.554e-02 0.010230
#> [29,]  5  8.593e-02 0.009317
#> [30,]  5  8.627e-02 0.008489
#> [31,]  5  8.656e-02 0.007735
#> [32,]  5  8.681e-02 0.007048
#> [33,]  5  8.702e-02 0.006422
#> [34,]  5  8.721e-02 0.005851
#> [35,]  5  8.736e-02 0.005331
#> [36,]  5  8.749e-02 0.004858
#> [37,]  5  8.761e-02 0.004426
#> [38,]  5  8.770e-02 0.004033
#> [39,]  5  8.779e-02 0.003675
#> [40,]  5  8.785e-02 0.003348
#> [41,]  5  8.791e-02 0.003051
#> [42,]  5  8.796e-02 0.002780
#> [43,]  5  8.800e-02 0.002533
#> [44,]  5  8.804e-02 0.002308
#> [45,]  5  8.807e-02 0.002103
#> [46,]  5  8.809e-02 0.001916
#> [47,]  5  8.811e-02 0.001746
#> [48,]  5  8.813e-02 0.001591
#> [49,]  5  8.815e-02 0.001449
#> [50,]  5  8.816e-02 0.001321
#> [51,]  5  8.817e-02 0.001203
#> [52,]  5  8.818e-02 0.001096
#> [1] "lambda min is at location 26"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "evaluation accuracy ElasticNet 0.643518518518518"

listData  <- predicting_scGPS(listData =listData,  mixedpop2 = mixedpop2,
    out_idx=out_idx, cluster_mixedpop2 = colData(mixedpop2)[, 1], c_selectID = c_selectID)

#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "ElasticNet for target subpop 1 has no solution"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "ElasticNet for target subpop 2 has no solution"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "ElasticNet for target subpop 3 has no solution"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "ElasticNet for target subpop 4 has no solution"