summarise bootstrap runs for Lasso model, from n bootstraps

the training and prediction results from bootstrap_scGPS were written to the object LSOLDA_dat, the reformat_LASSO summarises prediction for n bootstrap runs

reformat_LASSO(c_selectID = NULL, mp_selectID = NULL,
  LSOLDA_dat = NULL, nPredSubpop = NULL, Nodes_group = "#7570b3",
  nboots = 2)

Arguments

c_selectID
is the original cluster to be projected

mp_selectID
is the target mixedpop to project to

LSOLDA_dat
is the results from the bootstrap_scGPS

nPredSubpop
is the number of clusters in the target mixedpop row_cluster <-length(unique(target_cluster))

Nodes_group
string representation of hexidecimal color code for node

nboots
is an integer for how many bootstraps are run

Value

a dataframe containg information for the Lasso prediction results, each column contains prediction results for all subpopulations from each bootstrap run

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap_scGPS(nboots = 2,mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 107 genes for training model"
#> [1] "use 107 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "removing 10 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = y_cat, family = "binomial") 
#> 
#>        Df       %Dev    Lambda
#>   [1,]  0 -2.563e-15 3.550e-01
#>   [2,]  1  6.190e-02 3.235e-01
#>   [3,]  2  1.209e-01 2.947e-01
#>   [4,]  2  1.732e-01 2.686e-01
#>   [5,]  2  2.179e-01 2.447e-01
#>   [6,]  2  2.566e-01 2.230e-01
#>   [7,]  2  2.900e-01 2.032e-01
#>   [8,]  2  3.192e-01 1.851e-01
#>   [9,]  2  3.447e-01 1.687e-01
#>  [10,]  2  3.670e-01 1.537e-01
#>  [11,]  2  3.865e-01 1.400e-01
#>  [12,]  2  4.037e-01 1.276e-01
#>  [13,]  2  4.189e-01 1.163e-01
#>  [14,]  2  4.322e-01 1.059e-01
#>  [15,]  2  4.440e-01 9.652e-02
#>  [16,]  4  4.557e-01 8.794e-02
#>  [17,]  4  4.681e-01 8.013e-02
#>  [18,]  5  4.795e-01 7.301e-02
#>  [19,]  6  4.944e-01 6.653e-02
#>  [20,]  6  5.080e-01 6.062e-02
#>  [21,]  6  5.202e-01 5.523e-02
#>  [22,]  7  5.314e-01 5.032e-02
#>  [23,]  8  5.426e-01 4.585e-02
#>  [24,] 10  5.547e-01 4.178e-02
#>  [25,] 11  5.671e-01 3.807e-02
#>  [26,] 15  5.796e-01 3.469e-02
#>  [27,] 20  5.947e-01 3.160e-02
#>  [28,] 22  6.101e-01 2.880e-02
#>  [29,] 23  6.247e-01 2.624e-02
#>  [30,] 24  6.386e-01 2.391e-02
#>  [31,] 24  6.517e-01 2.178e-02
#>  [32,] 24  6.636e-01 1.985e-02
#>  [33,] 26  6.750e-01 1.809e-02
#>  [34,] 28  6.866e-01 1.648e-02
#>  [35,] 29  6.977e-01 1.501e-02
#>  [36,] 30  7.086e-01 1.368e-02
#>  [37,] 32  7.191e-01 1.247e-02
#>  [38,] 34  7.291e-01 1.136e-02
#>  [39,] 33  7.386e-01 1.035e-02
#>  [40,] 35  7.474e-01 9.430e-03
#>  [41,] 38  7.573e-01 8.592e-03
#>  [42,] 40  7.673e-01 7.829e-03
#>  [43,] 41  7.775e-01 7.133e-03
#>  [44,] 44  7.872e-01 6.500e-03
#>  [45,] 47  7.965e-01 5.922e-03
#>  [46,] 49  8.057e-01 5.396e-03
#>  [47,] 50  8.153e-01 4.917e-03
#>  [48,] 53  8.243e-01 4.480e-03
#>  [49,] 53  8.330e-01 4.082e-03
#>  [50,] 55  8.413e-01 3.719e-03
#>  [51,] 57  8.498e-01 3.389e-03
#>  [52,] 57  8.577e-01 3.088e-03
#>  [53,] 58  8.651e-01 2.814e-03
#>  [54,] 59  8.723e-01 2.564e-03
#>  [55,] 58  8.791e-01 2.336e-03
#>  [56,] 59  8.860e-01 2.128e-03
#>  [57,] 59  8.928e-01 1.939e-03
#>  [58,] 59  8.992e-01 1.767e-03
#>  [59,] 59  9.052e-01 1.610e-03
#>  [60,] 59  9.112e-01 1.467e-03
#>  [61,] 60  9.172e-01 1.337e-03
#>  [62,] 60  9.228e-01 1.218e-03
#>  [63,] 61  9.282e-01 1.110e-03
#>  [64,] 61  9.338e-01 1.011e-03
#>  [65,] 64  9.391e-01 9.213e-04
#>  [66,] 64  9.441e-01 8.395e-04
#>  [67,] 64  9.487e-01 7.649e-04
#>  [68,] 65  9.530e-01 6.969e-04
#>  [69,] 65  9.570e-01 6.350e-04
#>  [70,] 65  9.607e-01 5.786e-04
#>  [71,] 65  9.642e-01 5.272e-04
#>  [72,] 65  9.674e-01 4.804e-04
#>  [73,] 66  9.703e-01 4.377e-04
#>  [74,] 66  9.729e-01 3.988e-04
#>  [75,] 67  9.753e-01 3.634e-04
#>  [76,] 68  9.776e-01 3.311e-04
#>  [77,] 68  9.796e-01 3.017e-04
#>  [78,] 68  9.815e-01 2.749e-04
#>  [79,] 68  9.832e-01 2.505e-04
#>  [80,] 68  9.847e-01 2.282e-04
#>  [81,] 68  9.860e-01 2.079e-04
#>  [82,] 68  9.872e-01 1.895e-04
#>  [83,] 68  9.884e-01 1.726e-04
#>  [84,] 68  9.894e-01 1.573e-04
#>  [85,] 68  9.904e-01 1.433e-04
#>  [86,] 69  9.912e-01 1.306e-04
#>  [87,] 70  9.920e-01 1.190e-04
#>  [88,] 69  9.927e-01 1.084e-04
#>  [89,] 69  9.934e-01 9.879e-05
#>  [90,] 70  9.939e-01 9.001e-05
#>  [91,] 70  9.945e-01 8.202e-05
#>  [92,] 70  9.950e-01 7.473e-05
#>  [93,] 70  9.954e-01 6.809e-05
#>  [94,] 70  9.958e-01 6.204e-05
#>  [95,] 70  9.962e-01 5.653e-05
#>  [96,] 71  9.965e-01 5.151e-05
#>  [97,] 71  9.968e-01 4.693e-05
#>  [98,] 70  9.971e-01 4.276e-05
#>  [99,] 70  9.973e-01 3.896e-05
#> [100,] 70  9.976e-01 3.550e-05
#> [1] "lambda min is at location 38"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 224 cells and 97 genes..."
#> [1] "evaluation accuracy ElasticNet 0.799107142857143"
#> [1] "evaluation accuracy LDA 0.78125"
#> [1] "done training for bootstrap 1, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 17305"
#> [1] "Number of genes in the model present in the target data is 97"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 97 genes and 187 cells. The trained model has 97 genes"
#> [1] "first 10 genes in model "
#>  [1] "NPPB"   "FABP3"  "LCK"    "DMBX1"  "ABCA4"  "TPM3"   "RGS4"   "RXRG"  
#>  [9] "NR5A2"  "LEFTY1"
#> [1] "first 10 genes in target "
#>  [1] "NPPB"   "FABP3"  "LCK"    "DMBX1"  "ABCA4"  "TPM3"   "RGS4"   "RXRG"  
#>  [9] "NR5A2"  "LEFTY1"
#> [1] "running elasticNet classification..."
#> [1] "ElasticNet for target subpop 1 has no solution"
#> [1] "Number of genes in the target data, but not in model genes is 17305"
#> [1] "Number of genes in the model present in the target data is 97"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 97 genes and 187 cells. The trained model has 97 genes"
#> [1] "first 10 genes in model "
#>  [1] "NPPB"   "FABP3"  "LCK"    "DMBX1"  "ABCA4"  "TPM3"   "RGS4"   "RXRG"  
#>  [9] "NR5A2"  "LEFTY1"
#> [1] "first 10 genes in target "
#>  [1] "NPPB"   "FABP3"  "LCK"    "DMBX1"  "ABCA4"  "TPM3"   "RGS4"   "RXRG"  
#>  [9] "NR5A2"  "LEFTY1"
#> [1] "running LDA classification..."

#> Error in which(LDA_result[, 1] == c_selectID): object 'c_selectID' not found

reformat_LASSO(LSOLDA_dat=LSOLDA_dat,
    nPredSubpop=length(unique(colData(mixedpop2)[,1])), c_selectID = 1,
    mp_selectID =2)

#> Error in summary_prediction_lasso(LSOLDA_dat = LSOLDA_dat, nPredSubpop = nPredSubpop): object 'LSOLDA_dat' not found