summarise bootstrap runs for Lasso model, from n bootstraps

the training and prediction results from bootstrap_scGPS were written to the object LSOLDA_dat, the reformat_LASSO summarises prediction for n bootstrap runs

reformat_LASSO(c_selectID = NULL, mp_selectID = NULL,
  LSOLDA_dat = NULL, nPredSubpop = NULL, Nodes_group = "#7570b3",
  nboots = 2)

Arguments

c_selectID
is the original cluster to be projected

mp_selectID
is the target mixedpop to project to

LSOLDA_dat
is the results from the bootstrap_scGPS

nPredSubpop
is the number of clusters in the target mixedpop row_cluster <-length(unique(target_cluster))

Nodes_group
string representation of hexidecimal color code for node

nboots
is an integer for how many bootstraps are run

Value

a dataframe containg information for the Lasso prediction results, each column contains prediction results for all subpopulations from each bootstrap run

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap_scGPS(nboots = 2,mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "removing 1 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev   Lambda
#>  [1,]  0 8.009e-16 0.181500
#>  [2,]  1 1.615e-02 0.165400
#>  [3,]  1 2.960e-02 0.150700
#>  [4,]  1 4.083e-02 0.137300
#>  [5,]  1 5.021e-02 0.125100
#>  [6,]  1 5.807e-02 0.114000
#>  [7,]  1 6.465e-02 0.103900
#>  [8,]  1 7.017e-02 0.094650
#>  [9,]  3 7.654e-02 0.086240
#> [10,]  3 8.608e-02 0.078580
#> [11,]  3 9.424e-02 0.071600
#> [12,]  3 1.013e-01 0.065240
#> [13,]  3 1.074e-01 0.059440
#> [14,]  3 1.127e-01 0.054160
#> [15,]  3 1.173e-01 0.049350
#> [16,]  3 1.213e-01 0.044970
#> [17,]  3 1.248e-01 0.040970
#> [18,]  3 1.279e-01 0.037330
#> [19,]  3 1.306e-01 0.034020
#> [20,]  3 1.330e-01 0.030990
#> [21,]  3 1.350e-01 0.028240
#> [22,]  3 1.368e-01 0.025730
#> [23,]  3 1.384e-01 0.023450
#> [24,]  3 1.397e-01 0.021360
#> [25,]  4 1.410e-01 0.019470
#> [26,]  4 1.420e-01 0.017740
#> [27,]  4 1.429e-01 0.016160
#> [28,]  4 1.437e-01 0.014720
#> [29,]  4 1.444e-01 0.013420
#> [30,]  4 1.450e-01 0.012220
#> [31,]  4 1.455e-01 0.011140
#> [32,]  4 1.459e-01 0.010150
#> [33,]  4 1.463e-01 0.009248
#> [34,]  4 1.466e-01 0.008426
#> [35,]  4 1.469e-01 0.007677
#> [36,]  4 1.471e-01 0.006995
#> [37,]  4 1.473e-01 0.006374
#> [38,]  4 1.475e-01 0.005808
#> [39,]  4 1.476e-01 0.005292
#> [40,]  4 1.477e-01 0.004822
#> [41,]  4 1.478e-01 0.004393
#> [42,]  4 1.479e-01 0.004003
#> [43,]  4 1.480e-01 0.003647
#> [44,]  4 1.480e-01 0.003323
#> [45,]  4 1.481e-01 0.003028
#> [46,]  4 1.481e-01 0.002759
#> [47,]  4 1.481e-01 0.002514
#> [48,]  4 1.482e-01 0.002291
#> [49,]  4 1.482e-01 0.002087
#> [50,]  4 1.482e-01 0.001902
#> [51,]  4 1.482e-01 0.001733
#> [52,]  4 1.482e-01 0.001579
#> [53,]  4 1.483e-01 0.001439
#> [54,]  4 1.483e-01 0.001311
#> [1] "lambda min is at location 36"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 214 cells and 5 genes..."
#> [1] "evaluation accuracy ElasticNet 0.616822429906542"
#> [1] "evaluation accuracy LDA 0.61214953271028"
#> [1] "done training for bootstrap 1, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 1 is 47.5935828877005"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 31.0160427807487"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 62.8571428571429"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 47.8571428571429"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 90.2255639097744"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 41.3533834586466"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 75"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 42.5"
#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df       %Dev   Lambda
#>  [1,]  0 -1.442e-15 0.116900
#>  [2,]  1  6.708e-03 0.106600
#>  [3,]  1  1.228e-02 0.097090
#>  [4,]  1  1.692e-02 0.088460
#>  [5,]  1  2.078e-02 0.080600
#>  [6,]  1  2.400e-02 0.073440
#>  [7,]  3  2.882e-02 0.066920
#>  [8,]  3  3.447e-02 0.060970
#>  [9,]  3  3.924e-02 0.055560
#> [10,]  3  4.328e-02 0.050620
#> [11,]  3  4.672e-02 0.046120
#> [12,]  3  4.965e-02 0.042030
#> [13,]  3  5.216e-02 0.038290
#> [14,]  3  5.431e-02 0.034890
#> [15,]  4  5.618e-02 0.031790
#> [16,]  4  5.788e-02 0.028970
#> [17,]  4  5.933e-02 0.026390
#> [18,]  4  6.058e-02 0.024050
#> [19,]  4  6.164e-02 0.021910
#> [20,]  4  6.254e-02 0.019970
#> [21,]  4  6.332e-02 0.018190
#> [22,]  4  6.397e-02 0.016580
#> [23,]  4  6.454e-02 0.015100
#> [24,]  4  6.502e-02 0.013760
#> [25,]  4  6.542e-02 0.012540
#> [26,]  4  6.577e-02 0.011430
#> [27,]  4  6.606e-02 0.010410
#> [28,]  5  6.632e-02 0.009485
#> [29,]  5  6.661e-02 0.008643
#> [30,]  5  6.686e-02 0.007875
#> [31,]  5  6.708e-02 0.007175
#> [32,]  5  6.727e-02 0.006538
#> [33,]  5  6.742e-02 0.005957
#> [34,]  5  6.755e-02 0.005428
#> [35,]  5  6.766e-02 0.004946
#> [36,]  5  6.775e-02 0.004506
#> [37,]  5  6.783e-02 0.004106
#> [38,]  5  6.789e-02 0.003741
#> [39,]  5  6.794e-02 0.003409
#> [40,]  5  6.799e-02 0.003106
#> [41,]  4  6.803e-02 0.002830
#> [42,]  4  6.805e-02 0.002579
#> [43,]  4  6.808e-02 0.002350
#> [44,]  5  6.810e-02 0.002141
#> [45,]  5  6.812e-02 0.001951
#> [46,]  5  6.814e-02 0.001777
#> [47,]  5  6.815e-02 0.001620
#> [48,]  5  6.816e-02 0.001476
#> [49,]  5  6.817e-02 0.001345
#> [1] "lambda min is at location 17"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 215 cells and 6 genes..."
#> [1] "evaluation accuracy ElasticNet 0.66046511627907"
#> [1] "evaluation accuracy LDA 0.674418604651163"
#> [1] "done training for bootstrap 2, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "ElasticNet for target subpop 1 has no solution"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 0"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 90.7142857142857"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 84.2857142857143"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 4.51127819548872"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 0"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 35"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 22.5"

reformat_LASSO(LSOLDA_dat=LSOLDA_dat,
    nPredSubpop=length(unique(colData(mixedpop2)[,1])), c_selectID = 1,
    mp_selectID =2)

#> Warning: NAs introduced by coercion

#>          Boostrap1        Boostrap2 Target Source   Node NodeGroup    Value
#> 1 47.5935828877005               NA C1_MP2 C1_MP2 C1_MP2   #7570b3       NA
#> 2 62.8571428571429 90.7142857142857 C2_MP2 C1_MP2 C1_MP2   #7570b3 76.78571
#> 3 90.2255639097744 4.51127819548872 C3_MP2 C1_MP2 C1_MP2   #7570b3 47.36842
#> 4               75               35 C4_MP2 C1_MP2 C1_MP2   #7570b3 55.00000