get percent accuracy for Lasso model, from n bootstraps

The training results from training_scGPS were written to

summary_accuracy(object = NULL)

Arguments

object
is a list containing the training results from the summary_accuracy summarise n bootstraps

Value

a vector of percent accuracy for the selected subpopulation

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap_scGPS(nboots = 2,mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 107 genes for training model"
#> [1] "use 107 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "removing 12 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = y_cat, family = "binomial") 
#> 
#>        Df       %Dev    Lambda
#>   [1,]  0 -2.563e-15 3.653e-01
#>   [2,]  1  6.552e-02 3.328e-01
#>   [3,]  1  1.207e-01 3.032e-01
#>   [4,]  2  1.691e-01 2.763e-01
#>   [5,]  2  2.161e-01 2.518e-01
#>   [6,]  2  2.565e-01 2.294e-01
#>   [7,]  2  2.916e-01 2.090e-01
#>   [8,]  2  3.221e-01 1.904e-01
#>   [9,]  2  3.488e-01 1.735e-01
#>  [10,]  2  3.722e-01 1.581e-01
#>  [11,]  2  3.927e-01 1.441e-01
#>  [12,]  3  4.108e-01 1.313e-01
#>  [13,]  3  4.303e-01 1.196e-01
#>  [14,]  3  4.476e-01 1.090e-01
#>  [15,]  3  4.630e-01 9.930e-02
#>  [16,]  4  4.771e-01 9.048e-02
#>  [17,]  4  4.905e-01 8.244e-02
#>  [18,]  4  5.025e-01 7.512e-02
#>  [19,]  4  5.132e-01 6.844e-02
#>  [20,]  6  5.239e-01 6.236e-02
#>  [21,]  6  5.343e-01 5.682e-02
#>  [22,]  8  5.447e-01 5.177e-02
#>  [23,] 11  5.571e-01 4.717e-02
#>  [24,] 13  5.701e-01 4.298e-02
#>  [25,] 14  5.850e-01 3.917e-02
#>  [26,] 15  5.991e-01 3.569e-02
#>  [27,] 16  6.123e-01 3.252e-02
#>  [28,] 20  6.263e-01 2.963e-02
#>  [29,] 20  6.406e-01 2.700e-02
#>  [30,] 23  6.541e-01 2.460e-02
#>  [31,] 26  6.680e-01 2.241e-02
#>  [32,] 27  6.816e-01 2.042e-02
#>  [33,] 27  6.949e-01 1.861e-02
#>  [34,] 28  7.073e-01 1.695e-02
#>  [35,] 31  7.189e-01 1.545e-02
#>  [36,] 36  7.306e-01 1.408e-02
#>  [37,] 37  7.426e-01 1.282e-02
#>  [38,] 38  7.538e-01 1.169e-02
#>  [39,] 43  7.648e-01 1.065e-02
#>  [40,] 46  7.758e-01 9.702e-03
#>  [41,] 48  7.871e-01 8.840e-03
#>  [42,] 50  7.986e-01 8.054e-03
#>  [43,] 50  8.101e-01 7.339e-03
#>  [44,] 51  8.210e-01 6.687e-03
#>  [45,] 51  8.316e-01 6.093e-03
#>  [46,] 52  8.417e-01 5.552e-03
#>  [47,] 52  8.512e-01 5.058e-03
#>  [48,] 54  8.603e-01 4.609e-03
#>  [49,] 56  8.694e-01 4.200e-03
#>  [50,] 56  8.782e-01 3.826e-03
#>  [51,] 57  8.864e-01 3.487e-03
#>  [52,] 58  8.943e-01 3.177e-03
#>  [53,] 59  9.019e-01 2.895e-03
#>  [54,] 58  9.091e-01 2.637e-03
#>  [55,] 59  9.158e-01 2.403e-03
#>  [56,] 59  9.222e-01 2.190e-03
#>  [57,] 59  9.281e-01 1.995e-03
#>  [58,] 59  9.337e-01 1.818e-03
#>  [59,] 60  9.392e-01 1.656e-03
#>  [60,] 61  9.446e-01 1.509e-03
#>  [61,] 62  9.496e-01 1.375e-03
#>  [62,] 62  9.542e-01 1.253e-03
#>  [63,] 62  9.585e-01 1.142e-03
#>  [64,] 62  9.622e-01 1.040e-03
#>  [65,] 62  9.656e-01 9.478e-04
#>  [66,] 62  9.687e-01 8.636e-04
#>  [67,] 61  9.715e-01 7.869e-04
#>  [68,] 62  9.741e-01 7.170e-04
#>  [69,] 62  9.764e-01 6.533e-04
#>  [70,] 62  9.785e-01 5.953e-04
#>  [71,] 63  9.804e-01 5.424e-04
#>  [72,] 63  9.822e-01 4.942e-04
#>  [73,] 63  9.838e-01 4.503e-04
#>  [74,] 62  9.851e-01 4.103e-04
#>  [75,] 62  9.864e-01 3.739e-04
#>  [76,] 62  9.876e-01 3.406e-04
#>  [77,] 62  9.888e-01 3.104e-04
#>  [78,] 62  9.897e-01 2.828e-04
#>  [79,] 62  9.906e-01 2.577e-04
#>  [80,] 62  9.914e-01 2.348e-04
#>  [81,] 63  9.922e-01 2.139e-04
#>  [82,] 63  9.929e-01 1.949e-04
#>  [83,] 63  9.935e-01 1.776e-04
#>  [84,] 63  9.941e-01 1.618e-04
#>  [85,] 63  9.946e-01 1.475e-04
#>  [86,] 63  9.951e-01 1.344e-04
#>  [87,] 63  9.955e-01 1.224e-04
#>  [88,] 63  9.959e-01 1.115e-04
#>  [89,] 62  9.962e-01 1.016e-04
#>  [90,] 62  9.966e-01 9.261e-05
#>  [91,] 62  9.969e-01 8.438e-05
#>  [92,] 63  9.971e-01 7.688e-05
#>  [93,] 63  9.974e-01 7.005e-05
#>  [94,] 62  9.976e-01 6.383e-05
#>  [95,] 62  9.978e-01 5.816e-05
#>  [96,] 63  9.980e-01 5.299e-05
#>  [97,] 63  9.982e-01 4.828e-05
#>  [98,] 63  9.983e-01 4.400e-05
#>  [99,] 63  9.985e-01 4.009e-05
#> [100,] 64  9.986e-01 3.653e-05
#> [1] "lambda min is at location 15"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 224 cells and 95 genes..."
#> [1] "evaluation accuracy ElasticNet 0.84375"
#> [1] "evaluation accuracy LDA 0.794642857142857"
#> [1] "done training for bootstrap 1, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 17307"
#> [1] "Number of genes in the model present in the target data is 95"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 95 genes and 187 cells. The trained model has 95 genes"
#> [1] "first 10 genes in model "
#>  [1] "FABP3"  "LCK"    "DMBX1"  "TPM3"   "RXRG"   "NR5A2"  "LEFTY1" "LEFTY2"
#>  [9] "ACTN2"  "RYR2"  
#> [1] "first 10 genes in target "
#>  [1] "FABP3"  "LCK"    "DMBX1"  "TPM3"   "RXRG"   "NR5A2"  "LEFTY1" "LEFTY2"
#>  [9] "ACTN2"  "RYR2"  
#> [1] "running elasticNet classification..."

#> Error in which(ElasticNet_result[, 1] == c_selectID): object 'c_selectID' not found

summary_accuracy(LSOLDA_dat)

#> Error in summary_accuracy(LSOLDA_dat): object 'LSOLDA_dat' not found

summary_deviance(LSOLDA_dat)

#> Error in summary_deviance(LSOLDA_dat): object 'LSOLDA_dat' not found