get percent accuracy for Lasso model, from n bootstraps

The training results from training were written to

summary_accuracy(object = NULL)

Arguments

object
is a list containing the training results from the summary_accuracy summarise n bootstraps

Value

a vector of percent accuracy for the selected subpopulation

Author

Quan Nguyen, 2017-11-25

Examples

c_selectID<-1
day2 <- day_2_cardio_cell_sample
mixedpop1 <-new_scGPS_object(ExpressionMatrix = day2$dat2_counts, 
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- day_5_cardio_cell_sample
mixedpop2 <-new_scGPS_object(ExpressionMatrix = day5$dat5_counts, 
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-training_gene_sample
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap_prediction(nboots = 1,mixedpop1 = mixedpop1, 
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> done training for bootstrap 1, moving to prediction...

summary_accuracy(LSOLDA_dat)

#> [1] 68.2243

summary_deviance(LSOLDA_dat)

#> $allDeviance
#> [1] "4.44"
#> 
#> $DeviMax
#>          dat_DE$Dfd          Deviance           DEgenes
#> 1                 0              4.44    genes_cluster1
#> 2                 1              4.44    genes_cluster1
#> 3                 2              4.44    genes_cluster1
#> 4                 3              4.44    genes_cluster1
#> 5                 4              4.44    genes_cluster1
#> 6 remaining DEgenes remaining DEgenes remaining DEgenes
#> 
#> $LassoGenesMax
#> NULL
#>