get percent deviance explained for Lasso model, from n bootstraps

the training results from training were written to the object LSOLDA_dat, the summary_devidance summarises deviance explained for n bootstrap runs and also returns the best deviance matrix for plotting, as well as the best matrix with Lasso genes and coefficients

summary_deviance(object = NULL)

Arguments

object
is a list containing the training results from training

Value

a list containing three elements, with a vector of percent maximum deviance explained, a dataframe containg information for the full deviance, and a dataframe containing gene names and coefficients of the best model

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-new_scGPS_object(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-new_scGPS_object(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo,
                    CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap(nboots = 2,mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> Total 224 cells as source subpop

#> Total 366 cells in remaining subpops

#> subsampling 112 cells for training source subpop

#> subsampling 112 cells in remaining subpops for training

#> use 6 genes for training model

#> use 6 genes 224 cells for testing model

#> rename remaining subpops to 2_3

#> there are 112 cells in class 2_3 and 112 cells in class 1

#> removing 1 genes with no variance

#> standardizing prediction/target dataset

#> performning elasticnet model training...

#> performning LDA model training...

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> extracting deviance and best gene features...

#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev   Lambda
#>  [1,]  0 2.723e-15 0.101700
#>  [2,]  1 5.070e-03 0.092690
#>  [3,]  2 1.045e-02 0.084450
#>  [4,]  2 1.656e-02 0.076950
#>  [5,]  2 2.169e-02 0.070110
#>  [6,]  2 2.601e-02 0.063890
#>  [7,]  2 2.966e-02 0.058210
#>  [8,]  2 3.276e-02 0.053040
#>  [9,]  3 3.600e-02 0.048330
#> [10,]  3 3.934e-02 0.044030
#> [11,]  4 4.227e-02 0.040120
#> [12,]  4 4.485e-02 0.036560
#> [13,]  4 4.706e-02 0.033310
#> [14,]  4 4.895e-02 0.030350
#> [15,]  4 5.059e-02 0.027650
#> [16,]  4 5.200e-02 0.025200
#> [17,]  4 5.321e-02 0.022960
#> [18,]  4 5.425e-02 0.020920
#> [19,]  4 5.514e-02 0.019060
#> [20,]  4 5.591e-02 0.017370
#> [21,]  4 5.657e-02 0.015830
#> [22,]  4 5.714e-02 0.014420
#> [23,]  4 5.763e-02 0.013140
#> [24,]  4 5.804e-02 0.011970
#> [25,]  4 5.840e-02 0.010910
#> [26,]  4 5.870e-02 0.009939
#> [27,]  4 5.896e-02 0.009056
#> [28,]  4 5.918e-02 0.008251
#> [29,]  4 5.937e-02 0.007518
#> [30,]  4 5.953e-02 0.006850
#> [31,]  4 5.967e-02 0.006242
#> [32,]  4 5.978e-02 0.005687
#> [33,]  4 5.988e-02 0.005182
#> [34,]  4 5.996e-02 0.004722
#> [35,]  4 6.003e-02 0.004302
#> [36,]  4 6.009e-02 0.003920
#> [37,]  4 6.013e-02 0.003572
#> [38,]  4 6.018e-02 0.003254
#> [39,]  4 6.021e-02 0.002965
#> [40,]  4 6.024e-02 0.002702
#> [41,]  4 6.026e-02 0.002462
#> [42,]  4 6.028e-02 0.002243
#> [43,]  4 6.030e-02 0.002044
#> [44,]  4 6.032e-02 0.001862
#> [45,]  4 6.033e-02 0.001697
#> [46,]  4 6.034e-02 0.001546

#> lambda min is at location 20

#> the leave-out cells in the source subpop is 112

#> use 112 target subpops cells for leave-out test set

#> standardizing the leave-out target and source subpops...

#> start ElasticNet prediction for estimating accuracy...

#> start LDA prediction for estimating accuracy for 213 cells and 5 genes...

#> evaluation accuracy ElasticNet 0.676056338028169

#> evaluation accuracy LDA 0.676056338028169

#> done training for bootstrap 1, moving to prediction...

#> standardizing target subpops before prediction...

#> predicting from source to target subpop 1...

#> number of cells in the target subpop 1 is 187

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 1 is 32.0855614973262

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 1 is 38.5026737967914

#> predicting from source to target subpop 2...

#> number of cells in the target subpop 2 is 140

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 2 is 26.4285714285714

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 2 is 32.8571428571429

#> predicting from source to target subpop 3...

#> number of cells in the target subpop 3 is 133

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 3 is 66.9172932330827

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 3 is 50.3759398496241

#> predicting from source to target subpop 4...

#> number of cells in the target subpop 4 is 40

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 4 is 47.5

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 4 is 42.5

#> Total 224 cells as source subpop

#> Total 366 cells in remaining subpops

#> subsampling 112 cells for training source subpop

#> subsampling 112 cells in remaining subpops for training

#> use 6 genes for training model

#> use 6 genes 224 cells for testing model

#> rename remaining subpops to 2_3

#> there are 112 cells in class 2_3 and 112 cells in class 1

#> removing 1 genes with no variance

#> standardizing prediction/target dataset

#> performning elasticnet model training...

#> performning LDA model training...

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> extracting deviance and best gene features...

#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev   Lambda
#>  [1,]  0 2.723e-15 0.153700
#>  [2,]  1 1.158e-02 0.140100
#>  [3,]  1 2.122e-02 0.127600
#>  [4,]  1 2.925e-02 0.116300
#>  [5,]  1 3.595e-02 0.106000
#>  [6,]  1 4.155e-02 0.096550
#>  [7,]  2 4.652e-02 0.087970
#>  [8,]  2 5.334e-02 0.080160
#>  [9,]  2 5.917e-02 0.073030
#> [10,]  2 6.418e-02 0.066550
#> [11,]  2 6.849e-02 0.060630
#> [12,]  2 7.223e-02 0.055250
#> [13,]  2 7.547e-02 0.050340
#> [14,]  2 7.828e-02 0.045870
#> [15,]  2 8.073e-02 0.041790
#> [16,]  2 8.285e-02 0.038080
#> [17,]  2 8.470e-02 0.034700
#> [18,]  2 8.631e-02 0.031610
#> [19,]  2 8.770e-02 0.028810
#> [20,]  2 8.891e-02 0.026250
#> [21,]  2 8.995e-02 0.023920
#> [22,]  3 9.085e-02 0.021790
#> [23,]  3 9.204e-02 0.019860
#> [24,]  3 9.308e-02 0.018090
#> [25,]  3 9.397e-02 0.016480
#> [26,]  3 9.474e-02 0.015020
#> [27,]  4 9.547e-02 0.013690
#> [28,]  4 9.616e-02 0.012470
#> [29,]  4 9.676e-02 0.011360
#> [30,]  4 9.727e-02 0.010350
#> [31,]  4 9.770e-02 0.009433
#> [32,]  4 9.807e-02 0.008595
#> [33,]  4 9.839e-02 0.007831
#> [34,]  4 9.865e-02 0.007136
#> [35,]  4 9.888e-02 0.006502
#> [36,]  4 9.907e-02 0.005924
#> [37,]  4 9.923e-02 0.005398
#> [38,]  4 9.937e-02 0.004918
#> [39,]  4 9.948e-02 0.004481
#> [40,]  4 9.958e-02 0.004083
#> [41,]  4 9.966e-02 0.003720
#> [42,]  4 9.973e-02 0.003390
#> [43,]  4 9.979e-02 0.003089
#> [44,]  4 9.983e-02 0.002814
#> [45,]  4 9.987e-02 0.002564
#> [46,]  4 9.991e-02 0.002337
#> [47,]  4 9.993e-02 0.002129
#> [48,]  4 9.996e-02 0.001940
#> [49,]  4 9.998e-02 0.001768
#> [50,]  4 9.999e-02 0.001611
#> [51,]  4 1.000e-01 0.001467
#> [52,]  4 1.000e-01 0.001337
#> [53,]  4 1.000e-01 0.001218

#> lambda min is at location 16

#> the leave-out cells in the source subpop is 112

#> use 112 target subpops cells for leave-out test set

#> standardizing the leave-out target and source subpops...

#> start ElasticNet prediction for estimating accuracy...

#> start LDA prediction for estimating accuracy for 212 cells and 5 genes...

#> evaluation accuracy ElasticNet 0.674528301886792

#> evaluation accuracy LDA 0.679245283018868

#> done training for bootstrap 2, moving to prediction...

#> standardizing target subpops before prediction...

#> predicting from source to target subpop 1...

#> number of cells in the target subpop 1 is 187

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 1 is 4.27807486631016

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 1 is 0

#> predicting from source to target subpop 2...

#> number of cells in the target subpop 2 is 140

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 2 is 20.7142857142857

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 2 is 6.42857142857143

#> predicting from source to target subpop 3...

#> number of cells in the target subpop 3 is 133

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 3 is 7.5187969924812

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 3 is 0.75187969924812

#> predicting from source to target subpop 4...

#> number of cells in the target subpop 4 is 40

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 4 is 17.5

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 4 is 2.5

summary_deviance(LSOLDA_dat)

#> $allDeviance
#> [1] "0.05591" "0.08285"
#> 
#> $DeviMax
#>          dat_DE$Dfd          Deviance           DEgenes
#> 1                 0           0.08285    genes_cluster1
#> 2                 1           0.08285    genes_cluster1
#> 3                 2           0.08285    genes_cluster1
#> 4 remaining DEgenes remaining DEgenes remaining DEgenes
#> 
#> $LassoGenesMax
#> NULL
#>