get percent deviance explained for Lasso model, from n bootstraps

the training results from training_scGPS were written to the object LSOLDA_dat, the summary_devidance summarises deviance explained for n bootstrap runs and also returns the best deviance matrix for plotting, as well as the best matrix with Lasso genes and coefficients

summary_deviance(object = NULL)

Arguments

object
is a list containing the training results from training_scGPS

Value

a list containing three elements, with a vector of percent maximum deviance explained, a dataframe containg information for the full deviance, and a dataframe containing gene names and coefficients of the best model

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo,
                    CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap_scGPS(nboots = 2,mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "removing 1 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df     %Dev   Lambda
#>  [1,]  0 0.000000 0.109900
#>  [2,]  1 0.005932 0.100200
#>  [3,]  2 0.013350 0.091270
#>  [4,]  2 0.020760 0.083160
#>  [5,]  2 0.027060 0.075770
#>  [6,]  2 0.032430 0.069040
#>  [7,]  2 0.037030 0.062910
#>  [8,]  2 0.040970 0.057320
#>  [9,]  2 0.044370 0.052230
#> [10,]  3 0.047300 0.047590
#> [11,]  3 0.050910 0.043360
#> [12,]  3 0.054030 0.039510
#> [13,]  3 0.056720 0.036000
#> [14,]  3 0.059050 0.032800
#> [15,]  4 0.061160 0.029890
#> [16,]  4 0.063020 0.027230
#> [17,]  4 0.064640 0.024810
#> [18,]  4 0.066050 0.022610
#> [19,]  4 0.067260 0.020600
#> [20,]  4 0.068310 0.018770
#> [21,]  4 0.069210 0.017100
#> [22,]  4 0.069990 0.015580
#> [23,]  4 0.070660 0.014200
#> [24,]  4 0.071240 0.012940
#> [25,]  4 0.071730 0.011790
#> [26,]  4 0.072160 0.010740
#> [27,]  4 0.072520 0.009787
#> [28,]  4 0.072830 0.008917
#> [29,]  4 0.073100 0.008125
#> [30,]  4 0.073320 0.007403
#> [31,]  4 0.073510 0.006746
#> [32,]  4 0.073680 0.006146
#> [33,]  4 0.073810 0.005600
#> [34,]  4 0.073930 0.005103
#> [35,]  4 0.074030 0.004649
#> [36,]  4 0.074110 0.004236
#> [37,]  4 0.074180 0.003860
#> [38,]  4 0.074240 0.003517
#> [39,]  4 0.074290 0.003205
#> [40,]  4 0.074340 0.002920
#> [41,]  4 0.074370 0.002661
#> [42,]  4 0.074400 0.002424
#> [43,]  4 0.074430 0.002209
#> [44,]  4 0.074450 0.002013
#> [45,]  4 0.074460 0.001834
#> [46,]  4 0.074480 0.001671
#> [47,]  4 0.074490 0.001522
#> [48,]  4 0.074500 0.001387
#> [49,]  4 0.074510 0.001264
#> [1] "lambda min is at location 27"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 217 cells and 5 genes..."
#> [1] "evaluation accuracy ElasticNet 0.695852534562212"
#> [1] "evaluation accuracy LDA 0.700460829493088"
#> [1] "done training for bootstrap 1, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 1 is 37.9679144385027"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 52.4064171122995"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 17.8571428571429"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 40.7142857142857"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 63.9097744360902"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 62.406015037594"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 47.5"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 50"
#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df       %Dev   Lambda
#>  [1,]  0 -4.326e-15 0.129300
#>  [2,]  1  8.198e-03 0.117800
#>  [3,]  1  1.501e-02 0.107400
#>  [4,]  1  2.068e-02 0.097820
#>  [5,]  1  2.541e-02 0.089130
#>  [6,]  1  2.935e-02 0.081210
#>  [7,]  1  3.263e-02 0.074000
#>  [8,]  1  3.537e-02 0.067420
#>  [9,]  2  3.901e-02 0.061430
#> [10,]  2  4.257e-02 0.055980
#> [11,]  2  4.559e-02 0.051000
#> [12,]  2  4.816e-02 0.046470
#> [13,]  2  5.034e-02 0.042340
#> [14,]  2  5.221e-02 0.038580
#> [15,]  3  5.446e-02 0.035150
#> [16,]  5  5.732e-02 0.032030
#> [17,]  5  5.982e-02 0.029190
#> [18,]  5  6.197e-02 0.026590
#> [19,]  5  6.382e-02 0.024230
#> [20,]  5  6.542e-02 0.022080
#> [21,]  5  6.680e-02 0.020120
#> [22,]  5  6.799e-02 0.018330
#> [23,]  5  6.902e-02 0.016700
#> [24,]  5  6.991e-02 0.015220
#> [25,]  5  7.068e-02 0.013870
#> [26,]  5  7.134e-02 0.012630
#> [27,]  5  7.192e-02 0.011510
#> [28,]  5  7.241e-02 0.010490
#> [29,]  5  7.284e-02 0.009557
#> [30,]  5  7.321e-02 0.008708
#> [31,]  5  7.352e-02 0.007934
#> [32,]  5  7.379e-02 0.007229
#> [33,]  5  7.402e-02 0.006587
#> [34,]  5  7.422e-02 0.006002
#> [35,]  5  7.439e-02 0.005469
#> [36,]  5  7.453e-02 0.004983
#> [37,]  5  7.466e-02 0.004540
#> [38,]  5  7.476e-02 0.004137
#> [39,]  5  7.485e-02 0.003769
#> [40,]  5  7.492e-02 0.003435
#> [41,]  5  7.499e-02 0.003129
#> [42,]  5  7.504e-02 0.002851
#> [43,]  5  7.509e-02 0.002598
#> [44,]  5  7.512e-02 0.002367
#> [45,]  5  7.516e-02 0.002157
#> [46,]  5  7.518e-02 0.001965
#> [47,]  5  7.521e-02 0.001791
#> [48,]  5  7.522e-02 0.001632
#> [49,]  5  7.524e-02 0.001487
#> [50,]  5  7.525e-02 0.001355
#> [51,]  5  7.526e-02 0.001234
#> [52,]  5  7.527e-02 0.001125
#> [1] "lambda min is at location 10"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 218 cells and 6 genes..."
#> [1] "evaluation accuracy ElasticNet 0.610091743119266"
#> [1] "evaluation accuracy LDA 0.619266055045872"
#> [1] "done training for bootstrap 2, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 1 is 5.3475935828877"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 4.27807486631016"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 31.4285714285714"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 40"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 9.02255639097744"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 6.01503759398496"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 22.5"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 17.5"

summary_deviance(LSOLDA_dat)

#> $allDeviance
#> [1] "0.07252" "0.04257"
#> 
#> $DeviMax
#>          dat_DE$Dfd          Deviance           DEgenes
#> 1                 0           0.07252    genes_cluster1
#> 2                 1           0.07252    genes_cluster1
#> 3                 2           0.07252    genes_cluster1
#> 4                 3           0.07252    genes_cluster1
#> 5                 4           0.07252    genes_cluster1
#> 6 remaining DEgenes remaining DEgenes remaining DEgenes
#> 
#> $LassoGenesMax
#> NULL
#>