get percent deviance explained for Lasso model, from n bootstraps

the training results from training_scGPS were written to the object LSOLDA_dat, the summary_devidance summarises deviance explained for n bootstrap runs and also returns the best deviance matrix for plotting, as well as the best matrix with Lasso genes and coefficients

summary_deviance(object = NULL)

Arguments

object
is a list containing the training results from training_scGPS

Value

a list containing three elements, with a vector of percent maximum deviance explained, a dataframe containg information for the full deviance, and a dataframe containing gene names and coefficients of the best model

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo,
                    CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap_scGPS(nboots = 2,mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 107 genes for training model"
#> [1] "use 107 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "removing 13 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = y_cat, family = "binomial") 
#> 
#>        Df       %Dev    Lambda
#>   [1,]  0 -2.563e-15 3.395e-01
#>   [2,]  1  5.661e-02 3.094e-01
#>   [3,]  1  1.042e-01 2.819e-01
#>   [4,]  1  1.447e-01 2.568e-01
#>   [5,]  2  1.823e-01 2.340e-01
#>   [6,]  2  2.164e-01 2.132e-01
#>   [7,]  2  2.457e-01 1.943e-01
#>   [8,]  2  2.712e-01 1.770e-01
#>   [9,]  2  2.932e-01 1.613e-01
#>  [10,]  2  3.124e-01 1.470e-01
#>  [11,]  2  3.292e-01 1.339e-01
#>  [12,]  3  3.463e-01 1.220e-01
#>  [13,]  3  3.617e-01 1.112e-01
#>  [14,]  4  3.754e-01 1.013e-01
#>  [15,]  4  3.893e-01 9.230e-02
#>  [16,]  4  4.017e-01 8.410e-02
#>  [17,]  5  4.127e-01 7.663e-02
#>  [18,]  5  4.233e-01 6.982e-02
#>  [19,]  8  4.356e-01 6.362e-02
#>  [20,]  9  4.485e-01 5.797e-02
#>  [21,]  9  4.604e-01 5.282e-02
#>  [22,] 12  4.742e-01 4.813e-02
#>  [23,] 13  4.885e-01 4.385e-02
#>  [24,] 15  5.022e-01 3.996e-02
#>  [25,] 17  5.165e-01 3.641e-02
#>  [26,] 17  5.299e-01 3.317e-02
#>  [27,] 18  5.425e-01 3.023e-02
#>  [28,] 19  5.548e-01 2.754e-02
#>  [29,] 21  5.664e-01 2.509e-02
#>  [30,] 24  5.789e-01 2.286e-02
#>  [31,] 24  5.934e-01 2.083e-02
#>  [32,] 28  6.072e-01 1.898e-02
#>  [33,] 29  6.216e-01 1.730e-02
#>  [34,] 31  6.355e-01 1.576e-02
#>  [35,] 36  6.492e-01 1.436e-02
#>  [36,] 41  6.633e-01 1.308e-02
#>  [37,] 42  6.770e-01 1.192e-02
#>  [38,] 47  6.909e-01 1.086e-02
#>  [39,] 48  7.066e-01 9.897e-03
#>  [40,] 49  7.220e-01 9.018e-03
#>  [41,] 51  7.362e-01 8.217e-03
#>  [42,] 52  7.499e-01 7.487e-03
#>  [43,] 54  7.632e-01 6.822e-03
#>  [44,] 53  7.754e-01 6.216e-03
#>  [45,] 54  7.867e-01 5.664e-03
#>  [46,] 56  7.976e-01 5.161e-03
#>  [47,] 56  8.080e-01 4.702e-03
#>  [48,] 57  8.173e-01 4.284e-03
#>  [49,] 58  8.260e-01 3.904e-03
#>  [50,] 60  8.341e-01 3.557e-03
#>  [51,] 61  8.420e-01 3.241e-03
#>  [52,] 63  8.493e-01 2.953e-03
#>  [53,] 62  8.561e-01 2.691e-03
#>  [54,] 64  8.627e-01 2.452e-03
#>  [55,] 64  8.690e-01 2.234e-03
#>  [56,] 65  8.750e-01 2.035e-03
#>  [57,] 65  8.808e-01 1.855e-03
#>  [58,] 66  8.867e-01 1.690e-03
#>  [59,] 64  8.927e-01 1.540e-03
#>  [60,] 65  8.994e-01 1.403e-03
#>  [61,] 65  9.067e-01 1.278e-03
#>  [62,] 66  9.136e-01 1.165e-03
#>  [63,] 66  9.202e-01 1.061e-03
#>  [64,] 68  9.264e-01 9.670e-04
#>  [65,] 68  9.322e-01 8.811e-04
#>  [66,] 70  9.378e-01 8.028e-04
#>  [67,] 69  9.433e-01 7.315e-04
#>  [68,] 71  9.482e-01 6.665e-04
#>  [69,] 71  9.529e-01 6.073e-04
#>  [70,] 69  9.572e-01 5.533e-04
#>  [71,] 67  9.611e-01 5.042e-04
#>  [72,] 67  9.646e-01 4.594e-04
#>  [73,] 66  9.679e-01 4.186e-04
#>  [74,] 67  9.707e-01 3.814e-04
#>  [75,] 67  9.734e-01 3.475e-04
#>  [76,] 66  9.758e-01 3.166e-04
#>  [77,] 66  9.780e-01 2.885e-04
#>  [78,] 65  9.800e-01 2.629e-04
#>  [79,] 65  9.818e-01 2.395e-04
#>  [80,] 65  9.834e-01 2.183e-04
#>  [81,] 63  9.849e-01 1.989e-04
#>  [82,] 63  9.862e-01 1.812e-04
#>  [83,] 63  9.874e-01 1.651e-04
#>  [84,] 63  9.885e-01 1.504e-04
#>  [85,] 63  9.895e-01 1.371e-04
#>  [86,] 64  9.905e-01 1.249e-04
#>  [87,] 64  9.913e-01 1.138e-04
#>  [88,] 65  9.921e-01 1.037e-04
#>  [89,] 68  9.928e-01 9.448e-05
#>  [90,] 68  9.934e-01 8.608e-05
#>  [91,] 68  9.940e-01 7.844e-05
#>  [92,] 68  9.945e-01 7.147e-05
#>  [93,] 67  9.950e-01 6.512e-05
#>  [94,] 67  9.954e-01 5.933e-05
#>  [95,] 68  9.958e-01 5.406e-05
#>  [96,] 68  9.962e-01 4.926e-05
#>  [97,] 67  9.965e-01 4.488e-05
#>  [98,] 68  9.968e-01 4.090e-05
#>  [99,] 68  9.971e-01 3.726e-05
#> [100,] 68  9.974e-01 3.395e-05
#> [1] "lambda min is at location 13"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 224 cells and 94 genes..."
#> [1] "evaluation accuracy ElasticNet 0.825892857142857"
#> [1] "evaluation accuracy LDA 0.758928571428571"
#> [1] "done training for bootstrap 1, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 17308"
#> [1] "Number of genes in the model present in the target data is 94"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 94 genes and 187 cells. The trained model has 94 genes"
#> [1] "first 10 genes in model "
#>  [1] "NPPB"   "FABP3"  "LCK"    "DMBX1"  "ABCA4"  "TPM3"   "RGS4"   "RXRG"  
#>  [9] "NR5A2"  "LEFTY1"
#> [1] "first 10 genes in target "
#>  [1] "NPPB"   "FABP3"  "LCK"    "DMBX1"  "ABCA4"  "TPM3"   "RGS4"   "RXRG"  
#>  [9] "NR5A2"  "LEFTY1"
#> [1] "running elasticNet classification..."

#> Error in which(ElasticNet_result[, 1] == c_selectID): object 'c_selectID' not found

summary_deviance(LSOLDA_dat)

#> Error in summary_deviance(LSOLDA_dat): object 'LSOLDA_dat' not found