get percent deviance explained for Lasso model, from n bootstraps

the training results from training_scGPS were written to the object LSOLDA_dat, the summary_prediction summarises prediction for n bootstrap runs

summary_prediction_lasso(LSOLDA_dat = NULL, nPredSubpop = NULL)

Arguments

LSOLDA_dat
is a list containing the training results from training_scGPS

nPredSubpop
is the number of subpopulations in the target mixed population

Value

a dataframe containg information for the Lasso prediction results, each column contains prediction results for all subpopulations from each bootstrap run

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap_scGPS(nboots = 2,mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "removing 1 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df       %Dev   Lambda
#>  [1,]  0 -2.563e-15 0.132200
#>  [2,]  2  1.149e-02 0.120500
#>  [3,]  2  2.458e-02 0.109800
#>  [4,]  2  3.567e-02 0.100000
#>  [5,]  2  4.517e-02 0.091130
#>  [6,]  2  5.336e-02 0.083040
#>  [7,]  2  6.047e-02 0.075660
#>  [8,]  2  6.668e-02 0.068940
#>  [9,]  2  7.211e-02 0.062810
#> [10,]  2  7.687e-02 0.057230
#> [11,]  2  8.107e-02 0.052150
#> [12,]  2  8.476e-02 0.047520
#> [13,]  2  8.801e-02 0.043290
#> [14,]  2  9.088e-02 0.039450
#> [15,]  2  9.341e-02 0.035940
#> [16,]  2  9.563e-02 0.032750
#> [17,]  2  9.758e-02 0.029840
#> [18,]  2  9.929e-02 0.027190
#> [19,]  2  1.008e-01 0.024770
#> [20,]  2  1.021e-01 0.022570
#> [21,]  3  1.034e-01 0.020570
#> [22,]  3  1.046e-01 0.018740
#> [23,]  3  1.057e-01 0.017080
#> [24,]  3  1.065e-01 0.015560
#> [25,]  3  1.073e-01 0.014180
#> [26,]  3  1.080e-01 0.012920
#> [27,]  3  1.085e-01 0.011770
#> [28,]  3  1.090e-01 0.010720
#> [29,]  3  1.094e-01 0.009772
#> [30,]  3  1.098e-01 0.008904
#> [31,]  3  1.101e-01 0.008113
#> [32,]  3  1.103e-01 0.007392
#> [33,]  4  1.106e-01 0.006735
#> [34,]  4  1.108e-01 0.006137
#> [35,]  4  1.109e-01 0.005592
#> [36,]  4  1.111e-01 0.005095
#> [37,]  4  1.112e-01 0.004642
#> [38,]  4  1.113e-01 0.004230
#> [39,]  4  1.114e-01 0.003854
#> [40,]  4  1.115e-01 0.003512
#> [41,]  4  1.116e-01 0.003200
#> [42,]  4  1.116e-01 0.002916
#> [43,]  4  1.117e-01 0.002657
#> [44,]  4  1.117e-01 0.002421
#> [45,]  4  1.117e-01 0.002205
#> [46,]  4  1.117e-01 0.002010
#> [47,]  4  1.118e-01 0.001831
#> [48,]  4  1.118e-01 0.001668
#> [49,]  4  1.118e-01 0.001520
#> [50,]  4  1.118e-01 0.001385
#> [51,]  4  1.118e-01 0.001262
#> [52,]  4  1.118e-01 0.001150
#> [1] "lambda min is at location 41"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 219 cells and 5 genes..."
#> [1] "evaluation accuracy ElasticNet 0.671232876712329"
#> [1] "evaluation accuracy LDA 0.671232876712329"
#> [1] "done training for bootstrap 1, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 1 is 1.60427807486631"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 0"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 0.714285714285714"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 6.42857142857143"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 0.75187969924812"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 0"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 2.5"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 2.5"
#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df       %Dev   Lambda
#>  [1,]  0 -1.922e-15 0.143400
#>  [2,]  1  1.008e-02 0.130700
#>  [3,]  1  1.846e-02 0.119100
#>  [4,]  1  2.544e-02 0.108500
#>  [5,]  1  3.126e-02 0.098840
#>  [6,]  2  3.633e-02 0.090060
#>  [7,]  3  4.163e-02 0.082060
#>  [8,]  3  4.912e-02 0.074770
#>  [9,]  3  5.549e-02 0.068130
#> [10,]  3  6.091e-02 0.062080
#> [11,]  3  6.550e-02 0.056560
#> [12,]  4  6.945e-02 0.051540
#> [13,]  4  7.355e-02 0.046960
#> [14,]  4  7.708e-02 0.042790
#> [15,]  4  8.012e-02 0.038990
#> [16,]  4  8.275e-02 0.035520
#> [17,]  4  8.503e-02 0.032370
#> [18,]  4  8.700e-02 0.029490
#> [19,]  4  8.871e-02 0.026870
#> [20,]  4  9.019e-02 0.024480
#> [21,]  4  9.148e-02 0.022310
#> [22,]  4  9.259e-02 0.020330
#> [23,]  4  9.356e-02 0.018520
#> [24,]  4  9.440e-02 0.016880
#> [25,]  4  9.513e-02 0.015380
#> [26,]  4  9.576e-02 0.014010
#> [27,]  4  9.630e-02 0.012770
#> [28,]  4  9.677e-02 0.011630
#> [29,]  4  9.718e-02 0.010600
#> [30,]  4  9.752e-02 0.009657
#> [31,]  4  9.782e-02 0.008799
#> [32,]  4  9.808e-02 0.008017
#> [33,]  4  9.830e-02 0.007305
#> [34,]  4  9.849e-02 0.006656
#> [35,]  4  9.865e-02 0.006065
#> [36,]  4  9.878e-02 0.005526
#> [37,]  4  9.890e-02 0.005035
#> [38,]  5  9.900e-02 0.004588
#> [39,]  5  9.908e-02 0.004180
#> [40,]  5  9.915e-02 0.003809
#> [41,]  5  9.922e-02 0.003471
#> [42,]  5  9.927e-02 0.003162
#> [43,]  5  9.931e-02 0.002881
#> [44,]  5  9.935e-02 0.002625
#> [45,]  5  9.938e-02 0.002392
#> [46,]  5  9.940e-02 0.002180
#> [47,]  5  9.942e-02 0.001986
#> [48,]  5  9.944e-02 0.001810
#> [49,]  5  9.946e-02 0.001649
#> [50,]  5  9.947e-02 0.001502
#> [51,]  5  9.948e-02 0.001369
#> [52,]  5  9.949e-02 0.001247
#> [1] "lambda min is at location 24"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 218 cells and 6 genes..."
#> [1] "evaluation accuracy ElasticNet 0.68348623853211"
#> [1] "evaluation accuracy LDA 0.660550458715596"
#> [1] "done training for bootstrap 2, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "ElasticNet for target subpop 1 has no solution"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 1.06951871657754"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 66.4285714285714"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 85"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 0"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 6.76691729323308"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 15"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 25"

summary_prediction_lasso(LSOLDA_dat=LSOLDA_dat, nPredSubpop=4)

#>                  V1               V2                                      names
#> 1  1.60427807486631               NA ElasticNet for subpop1 in target mixedpop2
#> 2 0.714285714285714 66.4285714285714 ElasticNet for subpop2 in target mixedpop2
#> 3  0.75187969924812                0 ElasticNet for subpop3 in target mixedpop2
#> 4               2.5               15 ElasticNet for subpop4 in target mixedpop2