get percent deviance explained for Lasso model, from n bootstraps

the training results from training_scGPS were written to the object LSOLDA_dat, the summary_prediction summarises prediction for n bootstrap runs

summary_prediction_lasso(LSOLDA_dat = NULL, nPredSubpop = NULL)

Arguments

LSOLDA_dat
is a list containing the training results from training_scGPS

nPredSubpop
is the number of subpopulations in the target mixed population

Value

a dataframe containg information for the Lasso prediction results, each column contains prediction results for all subpopulations from each bootstrap run

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap_scGPS(nboots = 2,mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df      %Dev    Lambda
#>  [1,]  0 3.684e-15 0.1972000
#>  [2,]  1 1.907e-02 0.1797000
#>  [3,]  1 3.495e-02 0.1637000
#>  [4,]  1 4.822e-02 0.1492000
#>  [5,]  1 5.934e-02 0.1359000
#>  [6,]  1 6.866e-02 0.1238000
#>  [7,]  1 7.649e-02 0.1128000
#>  [8,]  1 8.307e-02 0.1028000
#>  [9,]  1 8.860e-02 0.0936900
#> [10,]  2 9.678e-02 0.0853600
#> [11,]  3 1.040e-01 0.0777800
#> [12,]  3 1.122e-01 0.0708700
#> [13,]  3 1.192e-01 0.0645700
#> [14,]  4 1.254e-01 0.0588400
#> [15,]  4 1.310e-01 0.0536100
#> [16,]  4 1.359e-01 0.0488500
#> [17,]  4 1.403e-01 0.0445100
#> [18,]  4 1.440e-01 0.0405500
#> [19,]  4 1.474e-01 0.0369500
#> [20,]  4 1.503e-01 0.0336700
#> [21,]  4 1.529e-01 0.0306800
#> [22,]  4 1.552e-01 0.0279500
#> [23,]  4 1.572e-01 0.0254700
#> [24,]  4 1.590e-01 0.0232100
#> [25,]  4 1.605e-01 0.0211500
#> [26,]  4 1.619e-01 0.0192700
#> [27,]  4 1.631e-01 0.0175600
#> [28,]  4 1.642e-01 0.0160000
#> [29,]  4 1.651e-01 0.0145700
#> [30,]  4 1.659e-01 0.0132800
#> [31,]  4 1.666e-01 0.0121000
#> [32,]  4 1.673e-01 0.0110300
#> [33,]  4 1.678e-01 0.0100500
#> [34,]  4 1.683e-01 0.0091530
#> [35,]  4 1.687e-01 0.0083400
#> [36,]  4 1.690e-01 0.0075990
#> [37,]  4 1.693e-01 0.0069240
#> [38,]  4 1.696e-01 0.0063090
#> [39,]  4 1.698e-01 0.0057490
#> [40,]  4 1.700e-01 0.0052380
#> [41,]  4 1.702e-01 0.0047730
#> [42,]  4 1.703e-01 0.0043490
#> [43,]  4 1.704e-01 0.0039620
#> [44,]  4 1.705e-01 0.0036100
#> [45,]  4 1.706e-01 0.0032900
#> [46,]  4 1.707e-01 0.0029970
#> [47,]  4 1.707e-01 0.0027310
#> [48,]  4 1.708e-01 0.0024880
#> [49,]  4 1.708e-01 0.0022670
#> [50,]  4 1.709e-01 0.0020660
#> [51,]  4 1.709e-01 0.0018820
#> [52,]  4 1.709e-01 0.0017150
#> [53,]  4 1.710e-01 0.0015630
#> [54,]  4 1.710e-01 0.0014240
#> [55,]  5 1.710e-01 0.0012970
#> [56,]  5 1.710e-01 0.0011820
#> [57,]  5 1.710e-01 0.0010770
#> [58,]  5 1.711e-01 0.0009815
#> [59,]  5 1.711e-01 0.0008943
#> [60,]  5 1.711e-01 0.0008148
#> [1] "lambda min is at location 17"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 216 cells and 6 genes..."
#> [1] "evaluation accuracy ElasticNet 0.615740740740741"
#> [1] "evaluation accuracy LDA 0.62037037037037"
#> [1] "done training for bootstrap 1, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 1 is 13.3689839572193"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 187 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 23.5294117647059"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 38.5714285714286"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 140 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 24.2857142857143"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 78.1954887218045"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 133 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 35.3383458646617"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 60"
#> [1] "Number of genes in the target data, but not in model genes is 4994"
#> [1] "Number of genes in the model present in the target data is 6"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 6 genes and 40 cells. The trained model has 6 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "MAP2"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 27.5"
#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 6 genes for training model"
#> [1] "use 6 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "removing 1 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df       %Dev   Lambda
#>  [1,]  0 -1.922e-15 0.143900
#>  [2,]  1  1.015e-02 0.131100
#>  [3,]  1  1.859e-02 0.119500
#>  [4,]  1  2.563e-02 0.108900
#>  [5,]  1  3.149e-02 0.099190
#>  [6,]  1  3.639e-02 0.090380
#>  [7,]  2  4.226e-02 0.082350
#>  [8,]  3  4.892e-02 0.075030
#>  [9,]  3  5.575e-02 0.068370
#> [10,]  3  6.160e-02 0.062290
#> [11,]  3  6.665e-02 0.056760
#> [12,]  3  7.104e-02 0.051720
#> [13,]  3  7.486e-02 0.047120
#> [14,]  3  7.819e-02 0.042940
#> [15,]  3  8.110e-02 0.039120
#> [16,]  3  8.365e-02 0.035650
#> [17,]  3  8.589e-02 0.032480
#> [18,]  3  8.784e-02 0.029590
#> [19,]  3  8.956e-02 0.026970
#> [20,]  3  9.106e-02 0.024570
#> [21,]  3  9.237e-02 0.022390
#> [22,]  3  9.352e-02 0.020400
#> [23,]  3  9.452e-02 0.018590
#> [24,]  3  9.539e-02 0.016940
#> [25,]  3  9.615e-02 0.015430
#> [26,]  3  9.681e-02 0.014060
#> [27,]  3  9.738e-02 0.012810
#> [28,]  4  9.790e-02 0.011670
#> [29,]  4  9.841e-02 0.010640
#> [30,]  4  9.885e-02 0.009691
#> [31,]  4  9.922e-02 0.008830
#> [32,]  4  9.954e-02 0.008046
#> [33,]  4  9.982e-02 0.007331
#> [34,]  4  1.001e-01 0.006680
#> [35,]  4  1.003e-01 0.006086
#> [36,]  4  1.004e-01 0.005546
#> [37,]  4  1.006e-01 0.005053
#> [38,]  4  1.007e-01 0.004604
#> [39,]  4  1.008e-01 0.004195
#> [40,]  4  1.009e-01 0.003822
#> [41,]  4  1.009e-01 0.003483
#> [42,]  4  1.010e-01 0.003173
#> [43,]  4  1.011e-01 0.002891
#> [44,]  4  1.011e-01 0.002635
#> [45,]  4  1.011e-01 0.002401
#> [46,]  4  1.012e-01 0.002187
#> [47,]  4  1.012e-01 0.001993
#> [48,]  4  1.012e-01 0.001816
#> [49,]  4  1.012e-01 0.001655
#> [50,]  4  1.012e-01 0.001508
#> [51,]  4  1.013e-01 0.001374
#> [52,]  4  1.013e-01 0.001252
#> [53,]  4  1.013e-01 0.001140
#> [1] "lambda min is at location 39"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 215 cells and 5 genes..."
#> [1] "evaluation accuracy ElasticNet 0.618604651162791"
#> [1] "evaluation accuracy LDA 0.627906976744186"
#> [1] "done training for bootstrap 2, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 1 is 54.5454545454545"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 1 is 52.4064171122995"
#> [1] "predicting from source to target subpop 2..."
#> [1] "number of cells in the target subpop 2 is 140"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 2 is 65.7142857142857"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 2 is 50"
#> [1] "predicting from source to target subpop 3..."
#> [1] "number of cells in the target subpop 3 is 133"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 3 is 96.9924812030075"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 3 is 61.6541353383459"
#> [1] "predicting from source to target subpop 4..."
#> [1] "number of cells in the target subpop 4 is 40"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running elasticNet classification..."
#> [1] "class probability prediction ElasticNet for target subpop 4 is 77.5"
#> [1] "Number of genes in the target data, but not in model genes is 4995"
#> [1] "Number of genes in the model present in the target data is 5"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes"
#> [1] "first 10 genes in model "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "first 10 genes in target "
#> [1] "GATA4"  "VIM"    "SNAI2"  "GJA1"   "TMEM88"
#> [1] "running LDA classification..."
#> [1] "class probability prediction LDA for target subpop 4 is 52.5"

summary_prediction_lasso(LSOLDA_dat=LSOLDA_dat, nPredSubpop=4)

#>                 V1               V2                                      names
#> 1 13.3689839572193 54.5454545454545 ElasticNet for subpop1 in target mixedpop2
#> 2 38.5714285714286 65.7142857142857 ElasticNet for subpop2 in target mixedpop2
#> 3 78.1954887218045 96.9924812030075 ElasticNet for subpop3 in target mixedpop2
#> 4               60             77.5 ElasticNet for subpop4 in target mixedpop2