get percent deviance explained for Lasso model, from n bootstraps

the training results from training_scGPS were written to the object LSOLDA_dat, the summary_prediction summarises prediction for n bootstrap runs

summary_prediction_lasso(LSOLDA_dat = NULL, nPredSubpop = NULL)

Arguments

LSOLDA_dat
is a list containing the training results from training_scGPS

nPredSubpop
is the number of subpopulations in the target mixed population

Value

a dataframe containg information for the Lasso prediction results, each column contains prediction results for all subpopulations from each bootstrap run

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
    GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap_scGPS(nboots = 2,mixedpop1 = mixedpop1,
    mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 107 genes for training model"
#> [1] "use 107 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "removing 7 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = y_cat, family = "binomial") 
#> 
#>        Df       %Dev    Lambda
#>   [1,]  0 -2.563e-15 3.026e-01
#>   [2,]  1  4.495e-02 2.757e-01
#>   [3,]  2  8.518e-02 2.512e-01
#>   [4,]  2  1.230e-01 2.289e-01
#>   [5,]  2  1.553e-01 2.086e-01
#>   [6,]  2  1.829e-01 1.900e-01
#>   [7,]  3  2.080e-01 1.731e-01
#>   [8,]  3  2.353e-01 1.578e-01
#>   [9,]  3  2.588e-01 1.437e-01
#>  [10,]  3  2.791e-01 1.310e-01
#>  [11,]  4  2.989e-01 1.193e-01
#>  [12,]  4  3.168e-01 1.087e-01
#>  [13,]  6  3.339e-01 9.908e-02
#>  [14,]  7  3.534e-01 9.028e-02
#>  [15,]  9  3.731e-01 8.226e-02
#>  [16,] 11  3.927e-01 7.495e-02
#>  [17,] 11  4.113e-01 6.829e-02
#>  [18,] 12  4.280e-01 6.222e-02
#>  [19,] 13  4.446e-01 5.670e-02
#>  [20,] 13  4.595e-01 5.166e-02
#>  [21,] 16  4.746e-01 4.707e-02
#>  [22,] 17  4.892e-01 4.289e-02
#>  [23,] 18  5.034e-01 3.908e-02
#>  [24,] 19  5.179e-01 3.561e-02
#>  [25,] 21  5.321e-01 3.244e-02
#>  [26,] 22  5.462e-01 2.956e-02
#>  [27,] 24  5.602e-01 2.694e-02
#>  [28,] 28  5.742e-01 2.454e-02
#>  [29,] 30  5.879e-01 2.236e-02
#>  [30,] 30  6.004e-01 2.038e-02
#>  [31,] 31  6.118e-01 1.857e-02
#>  [32,] 33  6.229e-01 1.692e-02
#>  [33,] 36  6.334e-01 1.541e-02
#>  [34,] 36  6.435e-01 1.404e-02
#>  [35,] 37  6.538e-01 1.280e-02
#>  [36,] 39  6.645e-01 1.166e-02
#>  [37,] 44  6.761e-01 1.062e-02
#>  [38,] 47  6.924e-01 9.680e-03
#>  [39,] 48  7.068e-01 8.820e-03
#>  [40,] 47  7.196e-01 8.037e-03
#>  [41,] 48  7.311e-01 7.323e-03
#>  [42,] 49  7.417e-01 6.672e-03
#>  [43,] 52  7.520e-01 6.079e-03
#>  [44,] 52  7.621e-01 5.539e-03
#>  [45,] 53  7.715e-01 5.047e-03
#>  [46,] 57  7.803e-01 4.599e-03
#>  [47,] 55  7.891e-01 4.190e-03
#>  [48,] 56  7.974e-01 3.818e-03
#>  [49,] 58  8.055e-01 3.479e-03
#>  [50,] 58  8.132e-01 3.170e-03
#>  [51,] 60  8.208e-01 2.888e-03
#>  [52,] 63  8.287e-01 2.632e-03
#>  [53,] 63  8.368e-01 2.398e-03
#>  [54,] 65  8.447e-01 2.185e-03
#>  [55,] 69  8.524e-01 1.991e-03
#>  [56,] 72  8.610e-01 1.814e-03
#>  [57,] 72  8.697e-01 1.653e-03
#>  [58,] 71  8.785e-01 1.506e-03
#>  [59,] 72  8.869e-01 1.372e-03
#>  [60,] 72  8.948e-01 1.250e-03
#>  [61,] 74  9.024e-01 1.139e-03
#>  [62,] 74  9.100e-01 1.038e-03
#>  [63,] 73  9.175e-01 9.458e-04
#>  [64,] 75  9.247e-01 8.617e-04
#>  [65,] 75  9.315e-01 7.852e-04
#>  [66,] 75  9.378e-01 7.154e-04
#>  [67,] 75  9.435e-01 6.519e-04
#>  [68,] 75  9.487e-01 5.940e-04
#>  [69,] 75  9.534e-01 5.412e-04
#>  [70,] 75  9.576e-01 4.931e-04
#>  [71,] 74  9.615e-01 4.493e-04
#>  [72,] 73  9.650e-01 4.094e-04
#>  [73,] 72  9.682e-01 3.730e-04
#>  [74,] 71  9.711e-01 3.399e-04
#>  [75,] 72  9.737e-01 3.097e-04
#>  [76,] 72  9.761e-01 2.822e-04
#>  [77,] 72  9.782e-01 2.571e-04
#>  [78,] 73  9.802e-01 2.343e-04
#>  [79,] 73  9.819e-01 2.135e-04
#>  [80,] 72  9.835e-01 1.945e-04
#>  [81,] 72  9.850e-01 1.772e-04
#>  [82,] 72  9.863e-01 1.615e-04
#>  [83,] 72  9.875e-01 1.471e-04
#>  [84,] 72  9.886e-01 1.341e-04
#>  [85,] 73  9.896e-01 1.221e-04
#>  [86,] 73  9.905e-01 1.113e-04
#>  [87,] 73  9.914e-01 1.014e-04
#>  [88,] 73  9.921e-01 9.240e-05
#>  [89,] 74  9.928e-01 8.419e-05
#>  [90,] 74  9.934e-01 7.671e-05
#>  [91,] 74  9.940e-01 6.990e-05
#>  [92,] 75  9.945e-01 6.369e-05
#>  [93,] 75  9.950e-01 5.803e-05
#>  [94,] 75  9.954e-01 5.288e-05
#>  [95,] 75  9.959e-01 4.818e-05
#>  [96,] 75  9.962e-01 4.390e-05
#>  [97,] 75  9.966e-01 4.000e-05
#>  [98,] 75  9.968e-01 3.645e-05
#>  [99,] 75  9.971e-01 3.321e-05
#> [100,] 75  9.974e-01 3.026e-05
#> [1] "lambda min is at location 19"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 224 cells and 100 genes..."
#> [1] "evaluation accuracy ElasticNet 0.866071428571429"
#> [1] "evaluation accuracy LDA 0.816964285714286"
#> [1] "done training for bootstrap 1, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 17302"
#> [1] "Number of genes in the model present in the target data is 100"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 100 genes and 187 cells. The trained model has 100 genes"
#> [1] "first 10 genes in model "
#>  [1] "NPPB"   "FABP3"  "LCK"    "DMBX1"  "ABCA4"  "TPM3"   "RGS4"   "RXRG"  
#>  [9] "NR5A2"  "LEFTY1"
#> [1] "first 10 genes in target "
#>  [1] "NPPB"   "FABP3"  "LCK"    "DMBX1"  "ABCA4"  "TPM3"   "RGS4"   "RXRG"  
#>  [9] "NR5A2"  "LEFTY1"
#> [1] "running elasticNet classification..."

#> Error in which(ElasticNet_result[, 1] == c_selectID): object 'c_selectID' not found

summary_prediction_lasso(LSOLDA_dat=LSOLDA_dat, nPredSubpop=4)

#> Error in summary_prediction_lasso(LSOLDA_dat = LSOLDA_dat, nPredSubpop = 4): object 'LSOLDA_dat' not found