get percent deviance explained for LDA model, from n bootstraps

the training results from training were written to the object LSOLDA_dat, the summary_prediction summarises prediction explained for n bootstrap runs and also returns the best deviance matrix for plotting, as well as the best matrix with Lasso genes and coefficients

summary_prediction_lda(LSOLDA_dat = NULL, nPredSubpop = NULL)

Arguments

LSOLDA_dat
is a list containing the training results from training

nPredSubpop
is the number of subpopulations in the target mixed population

Value

a dataframe containg information for the LDA prediction results, each column contains prediction results for all subpopulations from each bootstrap run

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-new_scGPS_object(ExpressionMatrix = day2$dat2_counts,
GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-new_scGPS_object(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap(nboots = 1,mixedpop1 = mixedpop1,
mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> Total 224 cells as source subpop

#> Total 366 cells in remaining subpops

#> subsampling 112 cells for training source subpop

#> subsampling 112 cells in remaining subpops for training

#> use 6 genes for training model

#> use 6 genes 224 cells for testing model

#> rename remaining subpops to 2_3

#> there are 112 cells in class 2_3 and 112 cells in class 1

#> removing 1 genes with no variance

#> standardizing prediction/target dataset

#> performning elasticnet model training...

#> performning LDA model training...

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> extracting deviance and best gene features...

#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = as.vector(dataset$Cluster_class), family = "binomial") 
#> 
#>       Df       %Dev   Lambda
#>  [1,]  0 -3.366e-15 0.150400
#>  [2,]  1  1.111e-02 0.137000
#>  [3,]  1  2.035e-02 0.124900
#>  [4,]  2  2.824e-02 0.113800
#>  [5,]  2  3.532e-02 0.103700
#>  [6,]  2  4.122e-02 0.094460
#>  [7,]  3  4.891e-02 0.086070
#>  [8,]  3  5.614e-02 0.078420
#>  [9,]  3  6.227e-02 0.071460
#> [10,]  4  6.920e-02 0.065110
#> [11,]  4  7.562e-02 0.059320
#> [12,]  4  8.120e-02 0.054050
#> [13,]  4  8.605e-02 0.049250
#> [14,]  4  9.029e-02 0.044880
#> [15,]  4  9.402e-02 0.040890
#> [16,]  4  9.729e-02 0.037260
#> [17,]  4  1.002e-01 0.033950
#> [18,]  4  1.027e-01 0.030930
#> [19,]  4  1.050e-01 0.028180
#> [20,]  4  1.070e-01 0.025680
#> [21,]  4  1.087e-01 0.023400
#> [22,]  4  1.102e-01 0.021320
#> [23,]  4  1.116e-01 0.019430
#> [24,]  4  1.128e-01 0.017700
#> [25,]  4  1.138e-01 0.016130
#> [26,]  4  1.148e-01 0.014700
#> [27,]  4  1.156e-01 0.013390
#> [28,]  4  1.162e-01 0.012200
#> [29,]  4  1.168e-01 0.011120
#> [30,]  4  1.174e-01 0.010130
#> [31,]  4  1.178e-01 0.009229
#> [32,]  4  1.182e-01 0.008409
#> [33,]  4  1.186e-01 0.007662
#> [34,]  4  1.188e-01 0.006981
#> [35,]  4  1.191e-01 0.006361
#> [36,]  4  1.193e-01 0.005796
#> [37,]  4  1.195e-01 0.005281
#> [38,]  4  1.196e-01 0.004812
#> [39,]  4  1.198e-01 0.004385
#> [40,]  4  1.199e-01 0.003995
#> [41,]  4  1.200e-01 0.003640
#> [42,]  4  1.200e-01 0.003317
#> [43,]  4  1.201e-01 0.003022
#> [44,]  4  1.202e-01 0.002754
#> [45,]  4  1.202e-01 0.002509
#> [46,]  4  1.203e-01 0.002286
#> [47,]  4  1.203e-01 0.002083
#> [48,]  4  1.203e-01 0.001898
#> [49,]  4  1.203e-01 0.001729
#> [50,]  4  1.204e-01 0.001576
#> [51,]  4  1.204e-01 0.001436
#> [52,]  4  1.204e-01 0.001308
#> [53,]  4  1.204e-01 0.001192
#> [54,]  4  1.204e-01 0.001086

#> lambda min is at location 21

#> the leave-out cells in the source subpop is 112

#> use 112 target subpops cells for leave-out test set

#> standardizing the leave-out target and source subpops...

#> start ElasticNet prediction for estimating accuracy...

#> start LDA prediction for estimating accuracy for 215 cells and 5 genes...

#> evaluation accuracy ElasticNet 0.623255813953488

#> evaluation accuracy LDA 0.627906976744186

#> done training for bootstrap 1, moving to prediction...

#> standardizing target subpops before prediction...

#> predicting from source to target subpop 1...

#> number of cells in the target subpop 1 is 187

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 1 is 70.0534759358289

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 187 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 1 is 51.8716577540107

#> predicting from source to target subpop 2...

#> number of cells in the target subpop 2 is 140

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 2 is 83.5714285714286

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 140 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 2 is 89.2857142857143

#> predicting from source to target subpop 3...

#> number of cells in the target subpop 3 is 133

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 3 is 47.3684210526316

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 133 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 3 is 62.406015037594

#> predicting from source to target subpop 4...

#> number of cells in the target subpop 4 is 40

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running elasticNet classification...

#> class probability prediction ElasticNet for target subpop 4 is 57.5

#> Number of genes in the target data, but not in model genes is 4995

#> Number of genes in the model present in the target data is 5

#> There are 0 genes that are in the model, but not in target subpopulations

#> the prediction (target) subop has 5 genes and 40 cells. The trained model has 5 genes

#> first 10 genes in model 

#> GATA4VIMSNAI2GJA1TMEM88

#> first 10 genes in target 

#> GATA4VIMSNAI2GJA1TMEM88

#> running LDA classification...

#> class probability prediction LDA for target subpop 4 is 62.5

summary_prediction_lda(LSOLDA_dat=LSOLDA_dat, nPredSubpop=4)

#>                 V1                                names
#> 1 51.8716577540107 LDA for subpop 1 in target mixedpop2
#> 2 89.2857142857143 LDA for subpop 2 in target mixedpop2
#> 3  62.406015037594 LDA for subpop 3 in target mixedpop2
#> 4             62.5 LDA for subpop 4 in target mixedpop2