get percent deviance explained for LDA model, from n bootstraps

the training results from training_scGPS were written to the object LSOLDA_dat, the summary_prediction summarises prediction explained for n bootstrap runs and also returns the best deviance matrix for plotting, as well as the best matrix with Lasso genes and coefficients

summary_prediction_lda(LSOLDA_dat = NULL, nPredSubpop = NULL)

Arguments

LSOLDA_dat
is a list containing the training results from training_scGPS

nPredSubpop
is the number of subpopulations in the target mixed population

Value

a dataframe containg information for the LDA prediction results, each column contains prediction results for all subpopulations from each bootstrap run

Examples

c_selectID<-1
day2 <- sample1
mixedpop1 <-NewscGPS(ExpressionMatrix = day2$dat2_counts,
GeneMetadata = day2$dat2geneInfo, CellMetadata = day2$dat2_clusters)
day5 <- sample2
mixedpop2 <-NewscGPS(ExpressionMatrix = day5$dat5_counts,
    GeneMetadata = day5$dat5geneInfo, CellMetadata = day5$dat5_clusters)
genes <-GeneList
genes <-genes$Merged_unique
LSOLDA_dat <- bootstrap_scGPS(nboots = 2,mixedpop1 = mixedpop1,
mixedpop2 = mixedpop2, genes=genes, c_selectID, listData =list(),
    cluster_mixedpop1 = colData(mixedpop1)[,1],
    cluster_mixedpop2=colData(mixedpop2)[,1])

#> [1] "Total 224 cells as source subpop"
#> [1] "Total 366 cells in remaining subpops"
#> [1] "subsampling 112 cells for training source subpop"
#> [1] "subsampling 112 cells in remaining subpops for training"
#> [1] "use 107 genes for training model"
#> [1] "use 107 genes 224 cells for testing model"
#> [1] "rename remaining subpops to 2_3"
#> [1] "there are 112 cells in class 2_3 and 112 cells in class 1"
#> [1] "removing 11 genes with no variance"
#> [1] "standardizing prediction/target dataset"
#> [1] "performning elasticnet model training..."
#> [1] "performning LDA model training..."

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> Warning: variables are collinear

#> [1] "extracting deviance and best gene features..."
#> 
#> Call:  glmnet(x = as.matrix(dataset[, -which(colnames(dataset) == "Cluster_class")]),      y = y_cat, family = "binomial") 
#> 
#>        Df       %Dev    Lambda
#>   [1,]  0 -2.563e-15 3.413e-01
#>   [2,]  1  5.719e-02 3.110e-01
#>   [3,]  1  1.053e-01 2.833e-01
#>   [4,]  1  1.462e-01 2.582e-01
#>   [5,]  1  1.812e-01 2.352e-01
#>   [6,]  2  2.153e-01 2.143e-01
#>   [7,]  2  2.458e-01 1.953e-01
#>   [8,]  2  2.722e-01 1.779e-01
#>   [9,]  2  2.951e-01 1.621e-01
#>  [10,]  2  3.151e-01 1.477e-01
#>  [11,]  3  3.342e-01 1.346e-01
#>  [12,]  3  3.531e-01 1.226e-01
#>  [13,]  3  3.697e-01 1.118e-01
#>  [14,]  6  3.885e-01 1.018e-01
#>  [15,]  6  4.066e-01 9.278e-02
#>  [16,]  6  4.226e-01 8.454e-02
#>  [17,]  7  4.370e-01 7.703e-02
#>  [18,]  7  4.503e-01 7.018e-02
#>  [19,]  9  4.648e-01 6.395e-02
#>  [20,] 10  4.789e-01 5.827e-02
#>  [21,] 10  4.919e-01 5.309e-02
#>  [22,] 11  5.041e-01 4.837e-02
#>  [23,] 14  5.177e-01 4.408e-02
#>  [24,] 14  5.304e-01 4.016e-02
#>  [25,] 15  5.438e-01 3.659e-02
#>  [26,] 17  5.567e-01 3.334e-02
#>  [27,] 19  5.689e-01 3.038e-02
#>  [28,] 20  5.809e-01 2.768e-02
#>  [29,] 21  5.920e-01 2.522e-02
#>  [30,] 21  6.022e-01 2.298e-02
#>  [31,] 23  6.118e-01 2.094e-02
#>  [32,] 26  6.214e-01 1.908e-02
#>  [33,] 30  6.328e-01 1.739e-02
#>  [34,] 33  6.456e-01 1.584e-02
#>  [35,] 35  6.612e-01 1.443e-02
#>  [36,] 39  6.768e-01 1.315e-02
#>  [37,] 39  6.911e-01 1.198e-02
#>  [38,] 40  7.042e-01 1.092e-02
#>  [39,] 40  7.161e-01 9.948e-03
#>  [40,] 41  7.270e-01 9.065e-03
#>  [41,] 42  7.372e-01 8.259e-03
#>  [42,] 46  7.471e-01 7.526e-03
#>  [43,] 47  7.574e-01 6.857e-03
#>  [44,] 48  7.672e-01 6.248e-03
#>  [45,] 53  7.766e-01 5.693e-03
#>  [46,] 54  7.856e-01 5.187e-03
#>  [47,] 55  7.948e-01 4.726e-03
#>  [48,] 58  8.040e-01 4.306e-03
#>  [49,] 59  8.134e-01 3.924e-03
#>  [50,] 61  8.240e-01 3.575e-03
#>  [51,] 60  8.343e-01 3.258e-03
#>  [52,] 62  8.442e-01 2.968e-03
#>  [53,] 65  8.541e-01 2.705e-03
#>  [54,] 67  8.641e-01 2.464e-03
#>  [55,] 67  8.737e-01 2.245e-03
#>  [56,] 67  8.825e-01 2.046e-03
#>  [57,] 68  8.912e-01 1.864e-03
#>  [58,] 67  8.997e-01 1.699e-03
#>  [59,] 70  9.082e-01 1.548e-03
#>  [60,] 69  9.162e-01 1.410e-03
#>  [61,] 69  9.238e-01 1.285e-03
#>  [62,] 68  9.307e-01 1.171e-03
#>  [63,] 68  9.368e-01 1.067e-03
#>  [64,] 68  9.426e-01 9.720e-04
#>  [65,] 68  9.476e-01 8.856e-04
#>  [66,] 68  9.525e-01 8.069e-04
#>  [67,] 68  9.567e-01 7.353e-04
#>  [68,] 68  9.607e-01 6.699e-04
#>  [69,] 68  9.642e-01 6.104e-04
#>  [70,] 68  9.674e-01 5.562e-04
#>  [71,] 68  9.704e-01 5.068e-04
#>  [72,] 69  9.730e-01 4.618e-04
#>  [73,] 69  9.755e-01 4.207e-04
#>  [74,] 69  9.777e-01 3.834e-04
#>  [75,] 69  9.797e-01 3.493e-04
#>  [76,] 69  9.815e-01 3.183e-04
#>  [77,] 69  9.832e-01 2.900e-04
#>  [78,] 69  9.847e-01 2.642e-04
#>  [79,] 69  9.861e-01 2.408e-04
#>  [80,] 69  9.873e-01 2.194e-04
#>  [81,] 70  9.885e-01 1.999e-04
#>  [82,] 70  9.895e-01 1.821e-04
#>  [83,] 70  9.904e-01 1.659e-04
#>  [84,] 70  9.913e-01 1.512e-04
#>  [85,] 70  9.920e-01 1.378e-04
#>  [86,] 70  9.927e-01 1.255e-04
#>  [87,] 70  9.934e-01 1.144e-04
#>  [88,] 70  9.940e-01 1.042e-04
#>  [89,] 70  9.945e-01 9.496e-05
#>  [90,] 70  9.950e-01 8.653e-05
#>  [91,] 70  9.954e-01 7.884e-05
#>  [92,] 70  9.958e-01 7.184e-05
#>  [93,] 70  9.962e-01 6.545e-05
#>  [94,] 70  9.965e-01 5.964e-05
#>  [95,] 70  9.968e-01 5.434e-05
#>  [96,] 70  9.971e-01 4.951e-05
#>  [97,] 70  9.974e-01 4.511e-05
#>  [98,] 71  9.976e-01 4.111e-05
#>  [99,] 71  9.978e-01 3.745e-05
#> [100,] 70  9.980e-01 3.413e-05
#> [1] "lambda min is at location 31"
#> [1] "the leave-out cells in the source subpop is 112"
#> [1] "use 112 target subpops cells for leave-out test set"
#> [1] "standardizing the leave-out target and source subpops..."
#> [1] "start ElasticNet prediction for estimating accuracy..."
#> [1] "start LDA prediction for estimating accuracy for 224 cells and 96 genes..."
#> [1] "evaluation accuracy ElasticNet 0.861607142857143"
#> [1] "evaluation accuracy LDA 0.754464285714286"
#> [1] "done training for bootstrap 1, moving to prediction..."
#> [1] "standardizing target subpops before prediction..."
#> [1] "predicting from source to target subpop 1..."
#> [1] "number of cells in the target subpop 1 is 187"
#> [1] "Number of genes in the target data, but not in model genes is 17306"
#> [1] "Number of genes in the model present in the target data is 96"
#> [1] "There are 0 genes that are in the model, but not in target subpopulations"
#> [1] "the prediction (target) subop has 96 genes and 187 cells. The trained model has 96 genes"
#> [1] "first 10 genes in model "
#>  [1] "NPPB"   "FABP3"  "LCK"    "DMBX1"  "TPM3"   "RXRG"   "NR5A2"  "LEFTY1"
#>  [9] "LEFTY2" "ACTN2" 
#> [1] "first 10 genes in target "
#>  [1] "NPPB"   "FABP3"  "LCK"    "DMBX1"  "TPM3"   "RXRG"   "NR5A2"  "LEFTY1"
#>  [9] "LEFTY2" "ACTN2" 
#> [1] "running elasticNet classification..."

#> Error in which(ElasticNet_result[, 1] == c_selectID): object 'c_selectID' not found

summary_prediction_lda(LSOLDA_dat=LSOLDA_dat, nPredSubpop=4)

#> Error in summary_prediction_lda(LSOLDA_dat = LSOLDA_dat, nPredSubpop = 4): object 'LSOLDA_dat' not found